Set action levels: failure thresholds and functions to invoke

The action_levels() function works with the actions argument that is present in the create_agent() function and in every validation step function. With it, we can provide threshold fail levels for any combination of warn, stop, or notify states. We can react to any entrance of a state by supplying corresponding functions to the fns argument. They will undergo evaluation at the time when the matching state is entered. If provided to create_agent() then the policies will be applied to every validation step, acting as a default for the validation as a whole. Calls of action_levels() could also be applied directly to any validation step and this will act as an override if set also in create_agent(). Usage of action_levels() is required to have any useful side effects (i.e., warnings, throwing errors) in the case of validation step functions operating directly on data.

action_levels(warn_at = NULL, stop_at = NULL, notify_at = NULL, fns = NULL)

Arguments

warn_at, stop_at, notify_at
The threshold number or fraction of validation units that can provide a fail result before entering the warn, stop, or notify failure states. If this a decimal value between 0 and 1 then it's a proportional failure threshold (e.g., 0.15 indicates that if 15% percent of the validation units are found to fail, then the designated failure state is entered). Absolute values starting from 1 can be used instead, and this constitutes an absolute failure threshold (e.g., 10 means that if 10 of the validation units are found to fail, the failure state is entered).

fns
A named list of functions that is to be paired with the appropriate failure states. The syntax for this list involves using failure state names from the set of warn, stop, and notify. The functions corresponding to the failure states are provided as formulas (e.g., list(warn = ~ warning("Too many failures.")). A series of expressions for each named state can be used by enclosing the set of statements with { }.

Details

The output of the action_levels() call in actions will be interpreted slightly differently if using an agent or using validation step functions directly on a data table. For convenience when working directly on data, any values supplied to warn_at or stop_at will be automatically given a stock warning() or stop() function. If you were to supply those manually then the stock functions would be overridden. Furthermore, if actions is NULL in this workflow, pointblank will use a warn_at value of 1 (providing a warning if there are any fail units). We can absolutely suppress this automatic warning behavior by at each validation step by setting active = FALSE. In this interactive data case there is no stock function given for the notify_at. The notify failure state is less commonly used in this workflow as it is in the agent-based one.

When using an agent, we often opt to not use any functions in fns as the warn, stop, and notify failure states will be reported on when using create_agent_report() (and, usually that's sufficient).

Function ID

1-3

See also

Examples

# Create a simple data frame with
# a column of numerical values
tbl <- dplyr::tibble(a = c(5, 7, 8, 5))

# Create an `action_levels()` list
# with fractional values for the
# `warn`, `stop`, and `notify` states
al <-
  action_levels(
    warn_at = 0.2,
    stop_at = 0.8,
    notify_at = 0.345
  )

# Validate that values in column
# `a` are always greater than 7 and
# apply the list of action levels
agent <-
  create_agent(tbl = tbl) %>%
  col_vals_gt(vars(a), 7, actions = al) %>%
  interrogate()

#> 

#> ── Interrogation Started - there is a single validation step ───────────────────

#> ✖ Step 1: STOP and NOTIFY conditions met.

#> 

#> ── Interrogation Completed ─────────────────────────────────────────────────────

# The report from the agent will show
# that the `warn` state has been entered
# for the first and only validation step.

# agent

# In the context of using validation
# stop functions directly on data, their
# use is commonly to trigger warnings
# and raise errors. The following will
# provide a warning (but that's suppressed
# here).
suppressWarnings(
  tbl %>%
    col_vals_gt(vars(a), 5, actions = al)
)

#> # A tibble: 4 x 1
#>       a
#>   <dbl>
#> 1     5
#> 2     7
#> 3     8
#> 4     5