Activate one or more of an agent's validation steps

If certain validation steps need to be activated after the creation of the validation plan for an agent, use the activate_steps() function. This is equivalent to using the active = TRUE for the selected validation steps (active is an argument in all validation functions). This will replace any function that may have been defined for the active argument during creation of the targeted validation steps.

activate_steps(agent, i = NULL)

Arguments

agent

An agent object of class ptblank_agent.

i

The validation step number, which is assigned to each validation step in the order of definition.

Value

A ptblank_agent object.

Function ID

9-6

See also

Examples

# Create an agent that has the
# `small_table` object as the
# target table, add a few inactive
# validation steps, and then use
# `interrogate()`
agent_1 <- 
  create_agent(
    read_fn = ~ small_table,
    tbl_name = "small_table",
    label = "An example."
  ) %>%
  col_exists(
    vars(date),
    active = FALSE
  ) %>%
  col_vals_regex(
    vars(b), regex = "[0-9]-[a-z]{3}-[0-9]{3}",
    active = FALSE
  ) %>%
  interrogate()
#> ℹ Step 1 is not set as active. Skipping.
#> ℹ Step 2 is not set as active. Skipping.

# In the above, the data is
# not actually interrogated
# because the `active` setting
# was `FALSE` in all steps; we
# can selectively change this
# with `activate_steps()`
agent_2 <-
  agent_1 %>%
  activate_steps(i = 1) %>%
  interrogate()
#> ℹ Step 2 is not set as active. Skipping.