Ponga la fecha actual en un nombre de archivo

Esta función ayuda a fijar la fecha actual a un nombre de archivo. Esto es útil al escribir objetos agent y informante en el disco como parte de un proceso continuo. La fecha puede ser en términos de hora UTC o del sistema local. tiempo. La fecha se puede colocar al final del nombre del archivo (antes del extensión de archivo) o al principio con un delimitador personalizable.

Las funciones x_write_disk(), yaml_write() permiten la escritura de objetos pointblank al disco. Además, la función log4r_step() tiene el argumento append_to que acepta nombres de archivo, y es razonable que un La serie de archivos de registro se puede diferenciar por un componente de fecha en el nombre esquema. La modificación de la cadena del nombre del archivo tiene efecto inmediatamente, pero no en el momento de escribir un archivo en el disco. En la mayoría de los casos, especialmente cuando usando affix_date() con las funciones de escritura de archivos antes mencionadas, el archivo las marcas de tiempo deben aproximarse a los componentes de tiempo fijados a los nombres de archivo.

Usage

affix_date(
  filename,
  position = c("end", "start"),
  format = "%Y-%m-%d",
  delimiter = "_",
  utc_time = TRUE
)

Arguments

filename

El nombre de archivo que se va a modificar.

position

Dónde colocar la fecha formateada. Esto podría estar al "end" del nombre del archivo (el predeterminado) o al "start".

format

Una cadena de formato base::strptime() para formatear la fecha. Por predeterminado, esto es "%Y-%m-%d" que expresa la fecha de acuerdo con la ISO 8601 estándar (como 'AAAA-MM-DD'). Consulte la documentación en base::strptime() para las especificaciones de conversión si planea usar un cadena de formato diferente.

delimiter

Los caracteres delimitadores que se utilizarán para separar la fecha. cadena del nombre del archivo original.

utc_time

Una opción para utilizar la hora UTC actual para establecer la fecha (la predeterminada, con TRUE), o, utilizar el local del sistema tiempo (FALSE).

Value

Un vector de caracteres.

Function ID

13-3

See also

Examples

# Tomando el nombre genérico de `pb_file`
# para un archivo, le agregamos la fecha
# actual como sufijo
affix_date(filename = "pb_file")
#> [1] "pb_file_2021-12-15"

# Las extensiones de archivo no se
# interpondrán en el camino:
affix_date(filename = "pb_file.rds")
#> [1] "pb_file_2021-12-15.rds"

# La fecha se puede utilizar como prefijo
affix_date(
  filename = "pb_file",
  position = "start"
)
#> [1] "2021-12-15_pb_file"

# El patrón de fecha se puede cambiar
# y así puede el delimitador
affix_date(
  filename = "pb_file.yml",
  format = "%Y%m%d",
  delimiter = "-"
)
#> [1] "pb_file-20211215.yml"

if (interactive()) {

# Podemos usar una convención de
# nomenclatura de archivos que
# involucran fechas al escribir la
# salida archivos inmediatamente
# después de interrogar; útil al
# interrogar directamente de YAML
# en un proceso programado
yaml_agent_interrogate(
  filename = system.file(
    "yaml", "agent-small_table.yml",
    package = "pointblank"
  )
) %>% 
  x_write_disk(
    filename = affix_date(
      filename = "small_table_agent.rds",
      delimiter = "-"
    ),
    keep_tbl = TRUE,
    keep_extracts = TRUE
  )

}