Ponga la fecha y hora actual en un nombre de archivo

Esta función ayuda a fijar la fecha y hora actual a un nombre de archivo. Este es útil al escribir objetos agent y informant en el disco como parte de un proceso continuo. La cadena de fecha y hora puede basarse en la hora UTC actual. o la hora del sistema local. La fecha y hora se puede colocar al final de el nombre del archivo (antes de la extensión del archivo) o al principio con un delimitador personalizable. Opcionalmente, la información de la zona horaria puede ser incluido. Si la fecha y hora se basa en la hora del sistema local, el sistema del usuario La zona horaria se muestra con el formato <time>(+/-)hhmm. Si usa la hora UTC, entonces se adopta el formato <time>Z.

Las funciones x_write_disk(), yaml_write() permiten la escritura de objetos pointblank al disco. La modificación de la cadena de nombre de archivo toma efecto inmediatamente, pero no en el momento de escribir un archivo en el disco. En la mayoría casos, especialmente cuando se usa affix_datetime() con el mencionado funciones de escritura de archivos, las marcas de tiempo del archivo deben aproximarse al tiempo componentes adheridos a los nombres de archivo.

Usage

affix_datetime(
  filename,
  position = c("end", "start"),
  format = "%Y-%m-%d_%H-%M-%S",
  delimiter = "_",
  utc_time = TRUE,
  add_tz = FALSE
)

Arguments

filename

El nombre de archivo que se va a modificar.

position

Dónde colocar la fecha y hora formateada. Esto podría ser al "end" del nombre del archivo (predeterminado) o al "start".

format

Una cadena de formato base::strptime() para formatear el fecha y hora. De forma predeterminada, es "%Y-%m-%dT%H:%M:%S" que expresa la fecha según la norma ISO 8601. Por ejemplo, si la corriente fecha-hora es 2020-12-04 13:11:23, la cadena formateada se convertiría en "2020-12-04T13:11:23". Consulte la documentación en base::strptime() para las especificaciones de conversión si planea utilizar una cadena de formato diferente.

delimiter

Los caracteres delimitadores que se utilizarán para separar la cadena de fecha y hora del nombre del archivo original.

utc_time

Una opción para usar la hora UTC actual para establecer la fecha y hora (la predeterminada, con TRUE), o usar la hora local del sistema (FALSE).

add_tz

¿Debería proporcionarse la zona horaria (como un desfase de UTC)? Si es TRUE, la compensación UTC se proporcionará como <time>Z (si utc_time = TRUE) o <time>(+/-)hhmm. De forma predeterminada, es FALSE.

Value

Un vector de caracteres.

Function ID

13-4

See also

Examples

# Tomando el nombre genérico de `pb_file`
# para un archivo, le agregamos la fecha y
# hora actual como sufijo
affix_datetime(filename = "pb_file")
#> [1] "pb_file_2021-12-15_04-07-22"

# Las extensiones de archivo no se
# interpondrán en el camino:
affix_datetime(filename = "pb_file.rds")
#> [1] "pb_file_2021-12-15_04-07-22.rds"

# La fecha y hora se puede utilizar
# como prefijo
affix_datetime(
  filename = "pb_file",
  position = "start"
)
#> [1] "2021-12-15_04-07-22_pb_file"

# El patrón de fecha y hora se puede
# cambiar y así puede el delimitador
affix_datetime(
  filename = "pb_file.yml",
  format = "%Y%m%d_%H%M%S",
  delimiter = "-"
)
#> [1] "pb_file-20211215_040722.yml"

# Se puede incluir información sobre
# la zona horaria
affix_datetime(
  filename = "pb_file.yml",
  add_tz = TRUE
)
#> [1] "pb_file_2021-12-15_04-07-22Z.yml"

if (interactive()) {

# Podemos usar una convención de
# nomenclatura de archivos que
# involucran fechas y horas al
# escribir la salida archivos
# inmediatamente después de interrogar;
# útil al interrogar directamente
# de YAML en un proceso programado
yaml_agent_interrogate(
  filename = system.file(
    "yaml", "agent-small_table.yml",
    package = "pointblank"
  )
) %>% 
  x_write_disk(
    filename = affix_datetime(
      filename = "small_table_agent.rds",
      delimiter = "-"
    ),
    keep_tbl = TRUE,
    keep_extracts = TRUE
  )

}