Did all of the validations fully pass?

Given an agent's validation plan that had undergone interrogation via interrogate(), did every single validation step result in zero fail levels? Using the all_passed() function will let us know whether that's TRUE or not.

all_passed(agent)

Arguments

agent
An agent object of class ptblank_agent.

Value

A logical value.

Function ID

8-4

See also

Examples

# Create a simple table with
# a column of numerical values
tbl <- 
  dplyr::tibble(a = c(5, 7, 8, 5))

# Validate that values in column
# `a` are always greater than 4
agent <-
  create_agent(tbl = tbl) %>%
  col_vals_gt(vars(a), value = 4) %>%
  interrogate()

# Determine if these column
# validations have all passed
# by using `all_passed()`
all_passed(agent = agent)
#> [1] TRUE