¿Existe realmente una o más columnas?

La función de validación col_exists(), la expectativa expect_col_exists() función, y la función de prueba test_col_exists() comprueban si uno o existen más columnas en la tabla de destino. El único requisito es la especificación de los nombres de las columnas. La función de validación se puede utilizar directamente en un dato. tabla o con un objeto * agent * (técnicamente, un objeto ptblank_agent) mientras que las funciones de expectativa y prueba solo se pueden usar con datos mesa. Los tipos de tablas de datos que se pueden utilizar incluyen marcos de datos, tibbles, tablas de base de datos (tbl_dbi) y Spark DataFrames (tbl_spark). Cada paso de validación o expectativa operará en una sola unidad de prueba, que es si la columna existe o no.

Uso

col_exists(
  x,
  columns,
  actions = NULL,
  step_id = NULL,
  label = NULL,
  brief = NULL,
  active = TRUE
)

expect_col_exists(object, columns, threshold = 1)

test_col_exists(object, columns, threshold = 1)

Argumentos

x

Un data.frame, tibble (tbl_df o tbl_dbi), Spark DataFrame (tbl_spark), o un agent objeto de clase ptblank_agent que se crea con create_agent().

columns

Una o más columnas de la tabla en foco. Esto se puede proporcionar como un vector de nombres de columnas usando c() o nombres de columnas desnudos encerrados entre vars().

actions

Una lista que contiene los niveles de umbral para que el paso de validación pueda reaccionar en consecuencia al superar los niveles establecidos. Esto se creará con la función auxiliar action_levels().

step_id

Uno o más identificadores opcionales para los pasos de validación únicos o múltiples generados al llamar a una función de validación. El uso de ID de pasos sirve para distinguir los pasos de validación entre sí y brinda la oportunidad de proporcionar una etiqueta más significativa en comparación con el índice de pasos. De forma predeterminada, es NULL, y pointblank generará automáticamente el valor de ID de paso (basado en el índice de paso) en este caso. Se pueden proporcionar uno o más valores, y el número exacto de valores de ID debe (1) coincidir con el número de pasos de validación que producirá la llamada a la función de validación (influenciado por el número de columns proporcionadas), (2) ser un ID cadena no utilizada en ningún paso de validación anterior, y (3) ser un vector con valores únicos.

label

Una etiqueta opcional para el paso de validación. Esta etiqueta aparece en el informe del agent y, para una mejor apariencia, debe ser breve.

brief

Una descripción opcional basada en texto para el paso de validación. Si no se proporciona nada aquí, el objeto agent genera un autobrief, utilizando el lenguaje proporcionado en el argumento lang de create_agent() (que por defecto es "en" o inglés). El autobrief incorpora detalles del paso de validación, por lo que a menudo es la opción preferida en la mayoría de los casos (donde un label podría ser más adecuada para describir sucintamente la validación).

active

Un valor lógico que indica si el paso de validación debe estar activo. Si la función de validación está trabajando con un objeto agent, FALSE hará que el paso de validación esté inactivo (aún informando su presencia y manteniendo los índices de los pasos sin cambios). Si la función de validación operará directamente en los datos (sin participación de agent), entonces cualquier paso con active = FALSE simplemente pasará los datos sin validación alguna. Aparte de un vector lógico, una fórmula R unilateral que usa un ~ inicial se puede usar con . (que sirve como la tabla de datos de entrada) para evaluar a un solo valor lógico. Con este enfoque, la función pointblank has_columns() se puede utilizar para determinar si se debe activar un paso de validación sobre la base de una o más columnas existentes en la tabla (por ejemplo, ~ . %>% has_columns(vars(d, e))). El valor predeterminado de active es TRUE.

object

Un data.frame, tibble (tbl_df o tbl_dbi) o Spark DataFrame (tbl_spark) que sirve como tabla de destino para la función de expectativa o la función de prueba.

threshold

Un valor de umbral de falla simple para usar con las variantes de función expectativa (expect_) y prueba (test_). De forma predeterminada, se establece en 1, lo que significa que cualquier unidad de falla en la validación de datos da como resultado una falla general de la prueba. Los números enteros más allá de 1 indican que cualquier unidad defectuosa hasta ese valor de umbral absoluto dará como resultado una thatthat prueba o evalúe como TRUE. Asimismo, los valores fraccionarios (entre 0 y 1) actúan como un umbral de falla proporcional, donde 0.15 significa que el 15 por ciento de las unidades de prueba que fallan dan como resultado una falla general de la prueba.

Valor

Para la función de validación, el valor de retorno es un objeto ptblank_agent o un objeto de tabla (dependiendo de si se pasó un objeto agent o una tabla a x). La función de expectativa devuelve invisiblemente su entrada pero, en el contexto de los datos de prueba, la función se llama principalmente por sus posibles efectos secundarios (por ejemplo, falla de señalización). La función de prueba devuelve un valor lógico.

Nombres de columnas

Si proporciona varios nombres de columna, el resultado será una expansión de pasos de validación para ese número de nombres de columna (por ejemplo, vars(col_a, col_b) dará lugar a la entrada de dos pasos de validación). Aparte de los nombres de las columnas en comillas y en vars(), tidyselect funciones auxiliares están disponibles para especificando columnas. Ellos son: starts_with(), ends_with(), contains(), matches(), y everything().

Actions

A menudo, querremos especificar actions para la validación. Este argumento, presente en cada función de validación, toma un objeto de lista especialmente diseñado que se produce mejor con la función action_levels(). Lee esa función documentación para la verdad sobre cómo crear reacciones por encima del umbral niveles de falla en la validación. La esencia básica es que querrás al menos un nivel de umbral único (especificado como la fracción de unidades de prueba falló, o un valor absoluto), a menudo usando el argumento warn_at. Utilizando action_levels(warn_at = 1) o action_levels(stop_at = 1) son buenas opciones dependiendo de la situación (el primero produce una advertencia, el otro stop()).

Briefs

¿Quiere describir este paso de validación con algún detalle? Tenga en cuenta que esto sólo es útil si x es un objeto agent. Si ese es el caso, use un texto brief que se ajuste a la situación. No se preocupe si no quiere hacerlo. Un autobrief se activa cuando brief = NULL y el texto luego se generará automáticamente.

YAML

Se puede escribir un agente pointblank en YAML con yaml_write() y el YAML resultante se puede usar para regenerar un agente (con yaml_read_agent()) o interrogar la tabla de destino (a través de yaml_agent_interrogate()). Cuando col_exists() se representa en YAML (bajo la clave de nivel superior steps como un miembro de la lista), la sintaxis sigue de cerca la firma de la función de validación. A continuación se muestra un ejemplo de cómo una llamada compleja de col_exists() como paso de validación se expresa en código R y en la representación YAML correspondiente.

# Código R
agent %>% 
  col_exists(
    vars(a),
    actions = action_levels(warn_at = 0.1, stop_at = 0.2),
    label = "El paso `col_exists()`.",
    active = FALSE
  )

# Representación YAML
steps:
- col_exists:
    columns: vars(a)
    actions:
      warn_fraction: 0.1
      stop_fraction: 0.2
    label: El paso `col_exists()`.
    active: false

En la práctica, ambos serán a menudo más cortos, ya que solo el argumento de las columns requiere un valor. Los argumentos con valores predeterminados no se escribirán en YAML cuando se use yaml_write() (aunque es aceptable incluirlos con sus valores predeterminados al generar el YAML por otros medios). También es posible obtener una vista previa de la transformación de un agente a YAML sin escribir en el disco usando la función yaml_agent_string().

Function ID

2-29

Ver también

Ejemplos

# Para todos los ejemplos aquí,
# usaremos una tabla simple con
# dos columnas: `a` y` b`
tbl <-
  dplyr::tibble(
    a = c(5, 7, 6, 5, 8, 7),
    b = c(7, 1, 0, 0, 0, 3)
  )

# A: Usando un `agent` con funciones de
#    validación y luego `interrogate()`

# Valide que las columnas `a` y` b`
# existan en la tabla `tbl`; esto hace
# dos pasos de validación distintos ya
# que se proporcionaron dos columnas
# a `vars()`
agent <-
  create_agent(tbl) %>%
  col_exists(vars(a, b)) %>%
  interrogate()

# Determine si esta validación no tuvo
# unidades de prueba defectuosas (1)
all_passed(agent)
#> [1] TRUE

# Llamar a `agent` en la consola imprime
# el informe del agente; pero podemos
# obtener un objeto `gt_tbl` directamente
# con `get_agent_report(agent)`

# B: Usando la función de validación
#    directamente en los datos
#    (sin `agent`)

# Esta forma de utilizar las funciones
# de validación actúa como un filtro de
# datos: los datos se pasan a través,
# pero deben `stop()` si hay una sola
# unidad de prueba que falla; el
# comportamiento de los efectos
# secundarios se puede personalizar con
# la opción `actions`
tbl %>% col_exists(vars(a, b))
#> # A tibble: 6 × 2
#>       a     b
#>   <dbl> <dbl>
#> 1     5     7
#> 2     7     1
#> 3     6     0
#> 4     5     0
#> 5     8     0
#> 6     7     3

# C: Usando la función de expectativa

# Con el formulario `expect_*()`,
# necesitamos ser más exactos y
# proporcionar una columna a la vez;
# esto se usa principalmente en
# pruebas `testthat`
expect_col_exists(tbl, vars(a))
expect_col_exists(tbl, vars(b))

# D: Usando la función de prueba

# Con el formulario `test_*()`,
# deberíamos obtener un único valor
# lógico que se nos devuelva (incluso
# si hay varias columnas probadas,
# como es el caso a continuación)
tbl %>% test_col_exists(vars(a, b))
#> [1] TRUE