Do the columns contain R factor objects?

The col_is_factor() validation step function and the expect_col_is_factor() expectation function both check whether one or more columns in a table is of the factor type. Like many of the col_is_*()-type functions in pointblank, the only requirement is a specification of the column names. The validation step function can be used directly on a data table or with an agent object (technically, a ptblank_agent object) whereas the expectation function can only be used with a data table. The types of data tables that can be used include data frames, tibbles, and even database tables of the tbl_dbi class. Each validation step or expectation will operate over a single test unit, which is whether the column is a factor-type column or not.

col_is_factor(x, columns, actions = NULL, brief = NULL, active = TRUE)

expect_col_is_factor(object, columns, threshold = 1)

Arguments

x
A data frame, tibble (tbl_df or tbl_dbi), or, an agent object of class ptblank_agent that can be created with create_agent().

columns
The column (or a set of columns, provided as a character vector) to which this validation should be applied.

actions
A list containing threshold levels so that the validation step can react accordingly when exceeding the set levels. This is to be created with the action_levels() helper function.

brief
An optional, text-based description for the validation step.

active
A logical value indicating whether the validation step should be active. If the step function is working with an agent, FALSE will make the validation step inactive (still reporting its presence and keeping indexes for the steps unchanged). If the step function will be operating directly on data, then any step with active = FALSE will simply pass the data through with no validation whatsoever. The default for this is TRUE.

object
A data frame or tibble (tbl_df or tbl_dbi) that serves as the target table for the expectation function.

threshold
A simple failure threshold value for use with the expectation function. By default, this is set to 1 meaning that any single unit of failure in data validation results in an overall test failure. Whole numbers beyond 1 indicate that any failing units up to that absolute threshold value will result in a succeeding testthat test. Likewise, fractional values (between 0 and 1) act as a proportional failure threshold.

Value

For the validation step function, the return value is either a ptblank_agent object or a table object (depending on whether an agent object or a table was passed to x). The expectation function invisibly returns its input but, in the context of testing data, the function is called primarily for its potential side-effects (e.g., signaling failure).

Details

If providing multiple column names, the result will be an expansion of validation steps to that number of column names (e.g., vars(col_a, col_b) will result in the entry of two validation steps). Aside from column names in quotes and in vars(), tidyselect helper functions are available for specifying columns. They are: starts_with(), ends_with(), contains(), matches(), and everything().

Often, we will want to specify actions for the validation. This argument, present in every validation step function, takes a specially-crafted list object that is best produced by the action_levels() function. Read that function's documentation for the lowdown on how to create reactions to above-threshold failure levels in validation. The basic gist is that you'll want at least a single threshold level (specified as either the fraction test units failed, or, an absolute value), often using the warn_at argument. This is especially true when x is a table object because, otherwise, nothing happens. For the col_is_*()-type functions, using action_levels(warn_at = 1) or action_levels(stop_at = 1) are good choices depending on the situation (the first produces a warning, the other stop()s).

Want to describe this validation step in some detail? Keep in mind that this is only useful if x is an agent. If that's the case, brief the agent with some text that fits. Don't worry if you don't want to do it. The autobrief protocol is kicked in when brief = NULL and a simple brief will then be automatically generated.

Function ID

2-22

See also

Examples

# Create a simple table with
# a column of `factor` values
tbl <-
  dplyr::tibble(a = factor(c("one", "two")))

# Validate that column `a` in the
# table is classed as `factor`
agent <-
  create_agent(tbl = tbl) %>%
  col_is_factor(vars(a)) %>%
  interrogate()

# Determine if these column
# validations have all passed
# by using `all_passed()`
all_passed(agent)

#> [1] TRUE