¿Las columnas contienen fechas POSIXct?

La función de validación col_is_posix(), la función de expectativa expect_col_is_posix() y la función de prueba test_col_is_posix() comprueban si una o más columnas en una tabla es del tipo de fecha y hora R POSIXct. Como muchas de las funciones de tipo col_is_*() en pointblank, el único requisito es una especificación de los nombres de las columnas. La función de validación se puede usar directamente en una tabla de datos o con un objeto agent (técnicamente, un objeto ptblank_agent) mientras que las funciones de expectativa y prueba solo se pueden usar con una tabla de datos. Los tipos de tablas de datos que se pueden utilizar incluyen marcos de datos, tibbles, tablas de base de datos (tbl_dbi) y Spark DataFrames (tbl_spark). Cada paso de validación o expectativa operará sobre una sola unidad de prueba, que es si la columna es una columna de tipo POSIXct o no.

Usage

col_is_posix(
  x,
  columns,
  actions = NULL,
  step_id = NULL,
  label = NULL,
  brief = NULL,
  active = TRUE
)

expect_col_is_posix(object, columns, threshold = 1)

test_col_is_posix(object, columns, threshold = 1)

Arguments

x

Un data.frame, tibble (tbl_df o tbl_dbi), Spark DataFrame (tbl_spark), o un agent objeto de clase ptblank_agent que se crea con create_agent().

columns

La columna (o un conjunto de columnas, proporcionado como un vector de caracteres) a la que se debe aplicar esta validación.

actions

Una lista que contiene los niveles de umbral para que el paso de validación pueda reaccionar en consecuencia al superar los niveles establecidos. Esto se creará con la función auxiliar action_levels().

step_id

Uno o más identificadores opcionales para los pasos de validación únicos o múltiples generados al llamar a una función de validación. El uso de ID de pasos sirve para distinguir los pasos de validación entre sí y brinda la oportunidad de proporcionar una etiqueta más significativa en comparación con el índice de pasos. De forma predeterminada, es NULL, y pointblank generará automáticamente el valor de ID de paso (basado en el índice de paso) en este caso. Se pueden proporcionar uno o más valores, y el número exacto de valores de ID debe (1) coincidir con el número de pasos de validación que producirá la llamada a la función de validación (influenciado por el número de columns proporcionadas), (2) ser un ID cadena no utilizada en ningún paso de validación anterior, y (3) ser un vector con valores únicos.

label

Una etiqueta opcional para el paso de validación. Esta etiqueta aparece en el informe del agent y, para una mejor apariencia, debe ser breve.

brief

Una descripción opcional basada en texto para el paso de validación. Si no se proporciona nada aquí, el objeto agent genera un autobrief, utilizando el lenguaje proporcionado en el argumento lang de create_agent() (que por defecto es "en" o inglés). El autobrief incorpora detalles del paso de validación, por lo que a menudo es la opción preferida en la mayoría de los casos (donde un label podría ser más adecuada para describir sucintamente la validación).

active

Un valor lógico que indica si el paso de validación debe estar activo. Si la función de validación está trabajando con un objeto agent, FALSE hará que el paso de validación esté inactivo (aún informando su presencia y manteniendo los índices de los pasos sin cambios). Si la función de validación operará directamente en los datos (sin participación de agent), entonces cualquier paso con active = FALSE simplemente pasará los datos sin validación alguna. Aparte de un vector lógico, una fórmula R unilateral que usa un ~ inicial se puede usar con . (que sirve como la tabla de datos de entrada) para evaluar a un solo valor lógico. Con este enfoque, la función pointblank has_columns() se puede utilizar para determinar si se debe activar un paso de validación sobre la base de una o más columnas existentes en la tabla (por ejemplo, ~ . %>% has_columns(vars(d, e))). El valor predeterminado de active es TRUE.

object

Un data.frame, tibble (tbl_df o tbl_dbi) o Spark DataFrame (tbl_spark) que sirve como tabla de destino para la función de expectativa o la función de prueba.

threshold

Un valor de umbral de falla simple para usar con las variantes de función expectativa (expect_) y prueba (test_). De forma predeterminada, se establece en 1, lo que significa que cualquier unidad de falla en la validación de datos da como resultado una falla general de la prueba. Los números enteros más allá de 1 indican que cualquier unidad defectuosa hasta ese valor de umbral absoluto dará como resultado una thatthat prueba o evalúe como TRUE. Asimismo, los valores fraccionarios (entre 0 y 1) actúan como un umbral de falla proporcional, donde 0.15 significa que el 15 por ciento de las unidades de prueba que fallan dan como resultado una falla general de la prueba.

Value

Para la función de validación, el valor de retorno es un objeto ptblank_agent o un objeto de tabla (dependiendo de si se pasó un objeto agent o una tabla a x). La función de expectativa devuelve invisiblemente su entrada pero, en el contexto de los datos de prueba, la función se llama principalmente por sus posibles efectos secundarios (por ejemplo, falla de señalización). La función de prueba devuelve un valor lógico.

Nombres de columnas

Si proporciona varios nombres de columna, el resultado será una expansión de pasos de validación para ese número de nombres de columna (por ejemplo, vars(col_a, col_b) dará lugar a la entrada de dos pasos de validación). Aparte de los nombres de las columnas en comillas y en vars(), tidyselect funciones auxiliares están disponibles para especificando columnas. Ellos son: starts_with(), ends_with(), contains(), matches(), y everything().

Actions

A menudo, querremos especificar actions para la validación. Este argumento, presente en cada función de validación, toma un objeto de lista especialmente diseñado que es mejor producido por la función action_levels(). Lea la documentación de esa función para obtener información sobre cómo crear reacciones a niveles de falla por encima del umbral en la validación. La esencia básica es que querrá al menos un nivel de umbral único (especificado como la fracción de unidades de prueba fallidas o un valor absoluto), a menudo utilizando el argumento warn_at. Esto es especialmente cierto cuando x es un objeto de tabla porque, de lo contrario, no sucede nada. Para las funciones de tipo col_is_*(), usar action_levels(warn_at = 1) o action_levels(stop_at = 1) son buenas opciones dependiendo de la situación (la primera produce una advertencia, la otra stop()).

Briefs

¿Quiere describir este paso de validación con algún detalle? Tenga en cuenta que esto sólo es útil si x es un objeto agent. Si ese es el caso, use un texto brief que se ajuste a la situación. No se preocupe si no quiere hacerlo. Un autobrief se activa cuando brief = NULL y el texto luego se generará automáticamente.

YAML

Se puede escribir un agente pointblank en YAML con yaml_write() y el YAML resultante se puede usar para regenerar un agente (con yaml_read_agent()) o interrogar la tabla de destino (a través de yaml_agent_interrogate()). Cuando col_is_posix() se representa en YAML (bajo la clave de nivel superior steps como un miembro de la lista), la sintaxis sigue de cerca la firma de la función de validación. A continuación se muestra un ejemplo de cómo una llamada compleja de col_is_posix() como paso de validación se expresa en código R y en la representación YAML correspondiente.

# Código R
agent %>% 
  col_is_posix(
    vars(a),
    actions = action_levels(warn_at = 0.1, stop_at = 0.2),
    label = "El paso `col_is_posix()`.",
    active = FALSE
  )

# Representación YAML
steps:
- col_is_posix:
    columns: vars(a)
    actions:
      warn_fraction: 0.1
      stop_fraction: 0.2
    label: El paso `col_is_posix()`.
    active: false

En la práctica, ambos serán a menudo más cortos, ya que solo el argumento de las columns requiere un valor. Los argumentos con valores predeterminados no se escribirán en YAML cuando se use yaml_write() (aunque es aceptable incluirlos con sus valores predeterminados al generar el YAML por otros medios). También es posible obtener una vista previa de la transformación de un agente a YAML sin escribir en el disco usando la función yaml_agent_string().

Function ID

2-27

See also

Examples

# La tabla `small_table` del paquete
# tiene una columna `date_time`; los
# siguientes ejemplos validarán que esa
# columna es de las clases `POSIXct`
# y `POSIXt`

# A: Usando un `agent` con funciones de
#    validación y luego `interrogate()`

# Valide que la columna `date_time` sea
# de hecho una columna de fecha y hora
agent <-
  create_agent(small_table) %>%
  col_is_posix(vars(date_time)) %>%
  interrogate()
  
# Determine si esta validación no tuvo
# unidades de prueba defectuosas (1)
all_passed(agent)
#> [1] TRUE

# Llamar a `agent` en la consola imprime
# el informe del agente; pero podemos
# obtener un objeto `gt_tbl` directamente
# con `get_agent_report(agent)`

# B: Usando la función de validación
#    directamente en los datos
#    (sin `agent`)

# Esta forma de utilizar las funciones
# de validación actúa como un filtro de
# datos: los datos se pasan a través,
# pero deben `stop()` si hay una sola
# unidad de prueba que falla; el
# comportamiento de los efectos
# secundarios se puede personalizar con
# la opción `actions`
small_table %>%
  col_is_posix(vars(date_time)) %>%
  dplyr::slice(1:5)
#> # A tibble: 5 × 8
#>   date_time           date           a b             c      d e     f    
#>   <dttm>              <date>     <int> <chr>     <dbl>  <dbl> <lgl> <chr>
#> 1 2016-01-04 11:00:00 2016-01-04     2 1-bcd-345     3  3423. TRUE  high 
#> 2 2016-01-04 00:32:00 2016-01-04     3 5-egh-163     8 10000. TRUE  low  
#> 3 2016-01-05 13:32:00 2016-01-05     6 8-kdg-938     3  2343. TRUE  high 
#> 4 2016-01-06 17:23:00 2016-01-06     2 5-jdo-903    NA  3892. FALSE mid  
#> 5 2016-01-09 12:36:00 2016-01-09     8 3-ldm-038     7   284. TRUE  low  

# C: Usando la función de expectativa

# Con el formulario `expect_*()`,
# necesitamos ser más exactos y
# proporcionar una columna a la vez;
# esto se usa principalmente en
# pruebas `testthat`
expect_col_is_posix(
  small_table, vars(date_time)
)

# D: Usando la función de prueba

# Con la forma `test_*()`, deberíamos
# obtener un único valor lógico devuelto
small_table %>%
  test_col_is_posix(vars(date_time))
#> [1] TRUE