Are column data less than or equal to a specified value?

The col_vals_lte() validation step function and the expect_col_vals_lte() expectation function both check whether column values in a table are less than or equal to a specified value (the exact comparison used in this function is col_val <= value). The value can be specified as a single, literal value or as a column name given in vars(). The validation step function can be used directly on a data table or with an agent object (technically, a ptblank_agent object) whereas the expectation function can only be used with a data table. The types of data tables that can be used include data frames, tibbles, and even database tables of tbl_dbi class. Each validation step or expectation will operate over the number of test units that is equal to the number of rows in the table (after any preconditions have been applied).

col_vals_lte(
  x,
  columns,
  value,
  na_pass = FALSE,
  preconditions = NULL,
  actions = NULL,
  brief = NULL,
  active = TRUE
)

expect_col_vals_lte(
  object,
  columns,
  value,
  na_pass = FALSE,
  preconditions = NULL,
  threshold = 1
)

Arguments

x
A data frame, tibble (tbl_df or tbl_dbi), or, an agent object of class ptblank_agent that can be created with create_agent().

columns
The column (or a set of columns, provided as a character vector) to which this validation should be applied.

value
A numeric value used for this test. Any column values <= value are considered passing.

na_pass
Should any encountered NA values be allowed to pass a validation unit? This is by default FALSE. Set to TRUE to give NAs a pass.

preconditions
expressions used for mutating the input table before proceeding with the validation. This is ideally as a one-sided R formula using a leading ~. In the formula representation, the . serves as the input data table to be transformed (e.g., ~ . %>% dplyr::mutate(col = col + 10).

actions
A list containing threshold levels so that the validation step can react accordingly when exceeding the set levels. This is to be created with the action_levels() helper function.

brief
An optional, text-based description for the validation step.

active
A logical value indicating whether the validation step should be active. If the step function is working with an agent, FALSE will make the validation step inactive (still reporting its presence and keeping indexes for the steps unchanged). If the step function will be operating directly on data, then any step with active = FALSE will simply pass the data through with no validation whatsoever. The default for this is TRUE.

object
A data frame or tibble (tbl_df or tbl_dbi) that serves as the target table for the expectation function.

threshold
A simple failure threshold value for use with the expectation function. By default, this is set to 1 meaning that any single unit of failure in data validation results in an overall test failure. Whole numbers beyond 1 indicate that any failing units up to that absolute threshold value will result in a succeeding testthat test. Likewise, fractional values (between 0 and 1) act as a proportional failure threshold.

Value

For the validation step function, the return value is either a ptblank_agent object or a table object (depending on whether an agent object or a table was passed to x). The expectation function invisibly returns its input but, in the context of testing data, the function is called primarily for its potential side-effects (e.g., signaling failure).

Details

If providing multiple column names to columns, the result will be an expansion of validation steps to that number of column names (e.g., vars(col_a, col_b) will result in the entry of two validation steps). Aside from column names in quotes and in vars(), tidyselect helper functions are available for specifying columns. They are: starts_with(), ends_with(), contains(), matches(), and everything().

This validation step function supports special handling of NA values. The na_pass argument will determine whether an NA value appearing in a test unit should be counted as a pass or a fail. The default of na_pass = FALSE means that any NAs encountered will accumulate failing test units.

Having table preconditions means pointblank will mutate the table just before interrogation. Such a table mutation is isolated in scope to the validation step(s) produced by the validation step function call. Using dplyr code is suggested here since the statements can be translated to SQL if necessary. The code is most easily supplied as a one-sided R formula (using a leading ~). In the formula representation, the . serves as the input data table to be transformed (e.g., ~ . %>% dplyr::mutate(col_a = col_b + 10)). Alternatively, a function could instead be supplied (e.g., function(x) dplyr::mutate(x, col_a = col_b + 10)).

Often, we will want to specify actions for the validation. This argument, present in every validation step function, takes a specially-crafted list object that is best produced by the action_levels() function. Read that function's documentation for the lowdown on how to create reactions to above-threshold failure levels in validation. The basic gist is that you'll want at least a single threshold level (specified as either the fraction test units failed, or, an absolute value), often using the warn_at argument. This is especially true when x is a table object because, otherwise, nothing happens. For the col_vals_*()-type functions, using action_levels(warn_at = 0.25) or action_levels(stop_at = 0.25) are good choices depending on the situation (the first produces a warning when a quarter of the total test units fails, the other stop()s at the same threshold level).

Want to describe this validation step in some detail? Keep in mind that this is only useful if x is an agent. If that's the case, brief the agent with some text that fits. Don't worry if you don't want to do it. The autobrief protocol is kicked in when brief = NULL and a simple brief will then be automatically generated.

Function ID

2-2

See also

Examples

# Create a simple table with a
# column of numerical values
tbl <-
  dplyr::tibble(
    a = c(5, 4, 1, 2),
    b = c(3, 2, 5, 6)
  )

# Validate that the sum of
# values across columns `a`
# and `b` are always less
# than or equal to 10
agent <-
  create_agent(tbl = tbl) %>%
  col_vals_lte(vars(a_b), 10,
    preconditions = ~
      . %>% dplyr::mutate(a_b = a + b)
  ) %>%
  interrogate()

#> 

#> ── Interrogation Started - there is a single validation step ───────────────────

#> ✔ Step 1: OK.

#> 

#> ── Interrogation Completed ─────────────────────────────────────────────────────

# Determine if this column
# validation has passed by using
# `all_passed()`
all_passed(agent)

#> [1] TRUE