¿Es un conjunto de valores un subconjunto de una columna de valores?

La función de validación col_vals_make_subset(), la función de expectativa expect_col_vals_make_subset() y la función de prueba test_col_vals_make_subset() comprueban si todos los valores de set se ven al menos una vez en una columna de tabla. La función de validación se puede usar directamente en una tabla de datos o con un objeto agent (técnicamente, un objeto ptblank_agent) mientras que las funciones de expectativa y prueba solo se pueden usar con una tabla de datos. Los tipos de tablas de datos que se pueden utilizar incluyen marcos de datos, tibbles, tablas de base de datos (tbl_dbi) y Spark DataFrames (tbl_spark). Cada paso de validación o expectativa operará sobre el número de unidades de prueba que es igual al número de elementos en el set.

Usage

col_vals_make_subset(
  x,
  columns,
  set,
  preconditions = NULL,
  segments = NULL,
  actions = NULL,
  step_id = NULL,
  label = NULL,
  brief = NULL,
  active = TRUE
)

expect_col_vals_make_subset(
  object,
  columns,
  set,
  preconditions = NULL,
  threshold = 1
)

test_col_vals_make_subset(
  object,
  columns,
  set,
  preconditions = NULL,
  threshold = 1
)

Arguments

x

Un data.frame, tibble (tbl_df o tbl_dbi), Spark DataFrame (tbl_spark), o un agent objeto de clase ptblank_agent que se crea con create_agent().

columns

La columna (o un conjunto de columnas, proporcionado como un vector de caracteres) a la que se debe aplicar esta validación.

set

A vector of elements that is expected to be a subset of the unique values in the target column.

preconditions

Una expresión opcional para mutar la tabla de entrada antes de continuar con la validación. Esto se puede proporcionar como una fórmula R unilateral usando un ~ inicial (por ejemplo, ~ . %>% dplyr::mutate(col = col + 10) o como una función (por ejemplo, function (x) dplyr::mutate(x, col = col + 10). Consulte la sección Preconditions para obtener más información.

segments

Una expresión opcional o un conjunto de expresiones (contenidas en una lista) que sirven para segmentar la tabla de destino por valores de columna. Cada expresión se puede dar de una de dos maneras: (1) como nombres de columna, o (2) como una fórmula de dos lados donde el LHS contiene un nombre de columna y el RHS contiene los valores de columna para segmentar. Consulte la sección Segments para obtener más detalles sobre esto.

actions

Una lista que contiene los niveles de umbral para que el paso de validación pueda reaccionar en consecuencia al superar los niveles establecidos. Esto se creará con la función auxiliar action_levels().

step_id

Uno o más identificadores opcionales para los pasos de validación únicos o múltiples generados al llamar a una función de validación. El uso de ID de pasos sirve para distinguir los pasos de validación entre sí y brinda la oportunidad de proporcionar una etiqueta más significativa en comparación con el índice de pasos. De forma predeterminada, es NULL, y pointblank generará automáticamente el valor de ID de paso (basado en el índice de paso) en este caso. Se pueden proporcionar uno o más valores, y el número exacto de valores de ID debe (1) coincidir con el número de pasos de validación que producirá la llamada a la función de validación (influenciado por el número de columns proporcionadas), (2) ser un ID cadena no utilizada en ningún paso de validación anterior, y (3) ser un vector con valores únicos.

label

Una etiqueta opcional para el paso de validación. Esta etiqueta aparece en el informe del agent y, para una mejor apariencia, debe ser breve.

brief

Una descripción opcional basada en texto para el paso de validación. Si no se proporciona nada aquí, el objeto agent genera un autobrief, utilizando el lenguaje proporcionado en el argumento lang de create_agent() (que por defecto es "en" o inglés). El autobrief incorpora detalles del paso de validación, por lo que a menudo es la opción preferida en la mayoría de los casos (donde un label podría ser más adecuada para describir sucintamente la validación).

active

Un valor lógico que indica si el paso de validación debe estar activo. Si la función de validación está trabajando con un objeto agent, FALSE hará que el paso de validación esté inactivo (aún informando su presencia y manteniendo los índices de los pasos sin cambios). Si la función de validación operará directamente en los datos (sin participación de agent), entonces cualquier paso con active = FALSE simplemente pasará los datos sin validación alguna. Aparte de un vector lógico, una fórmula R unilateral que usa un ~ inicial se puede usar con . (que sirve como la tabla de datos de entrada) para evaluar a un solo valor lógico. Con este enfoque, la función pointblank has_columns() se puede utilizar para determinar si se debe activar un paso de validación sobre la base de una o más columnas existentes en la tabla (por ejemplo, ~ . %>% has_columns(vars(d, e))). El valor predeterminado de active es TRUE.

object

Un data.frame, tibble (tbl_df o tbl_dbi) o Spark DataFrame (tbl_spark) que sirve como tabla de destino para la función de expectativa o la función de prueba.

threshold

Un valor de umbral de falla simple para usar con las variantes de función expectativa (expect_) y prueba (test_). De forma predeterminada, se establece en 1, lo que significa que cualquier unidad de falla en la validación de datos da como resultado una falla general de la prueba. Los números enteros más allá de 1 indican que cualquier unidad defectuosa hasta ese valor de umbral absoluto dará como resultado una thatthat prueba o evalúe como TRUE. Asimismo, los valores fraccionarios (entre 0 y 1) actúan como un umbral de falla proporcional, donde 0.15 significa que el 15 por ciento de las unidades de prueba que fallan dan como resultado una falla general de la prueba.

Value

Para la función de validación, el valor de retorno es un objeto ptblank_agent o un objeto de tabla (dependiendo de si se pasó un objeto agent o una tabla a x). La función de expectativa devuelve invisiblemente su entrada pero, en el contexto de los datos de prueba, la función se llama principalmente por sus posibles efectos secundarios (por ejemplo, falla de señalización). La función de prueba devuelve un valor lógico.

Nombres de columnas

Si proporciona varios nombres de columna, el resultado será una expansión de pasos de validación para ese número de nombres de columna (por ejemplo, vars(col_a, col_b) dará lugar a la entrada de dos pasos de validación). Aparte de los nombres de las columnas en comillas y en vars(), tidyselect funciones auxiliares están disponibles para especificando columnas. Ellos son: starts_with(), ends_with(), contains(), matches(), y everything().

Preconditions

Proporcionar expresiones como preconditions significa que pointblank preprocesará la tabla de destino durante la interrogación como paso preparatorio. Puede suceder que una validación en particular requiera una columna calculada, algún filtrado de filas o la adición de columnas a través de una combinación, etc. Especialmente para un informe basado en agent, esto puede ser ventajoso ya que podemos desarrollar un gran plan de validación con una sola tabla de destino y realice ajustes menores en ella, según sea necesario, a lo largo del camino.

La mutación de la tabla está totalmente aislada en el alcance de los pasos de validación en los que se utilizan las preconditions. Aquí se sugiere usar el código dplyr ya que las declaraciones se pueden traducir a SQL si es necesario (es decir, si la tabla de destino reside en una base de datos). El código se proporciona más fácilmente como una fórmula R unilateral (utilizando un ~ inicial). En la representación de la fórmula, el . sirve como la tabla de datos de entrada que se va a transformar (por ejemplo, ~ . %>% dplyr::mutate(col_b = col_a + 10)). Alternativamente, se podría proporcionar una función (por ejemplo, function(x) dplyr::mutate(x, col_b = col_a + 10)).

Segmentos

Al usar el argumento segments, es posible definir una validación particular con segmentos (o porciones de fila) de la tabla de destino. Una expresión opcional o un conjunto de expresiones que sirven para segmentar la tabla de destino por valores de columna. Cada expresión se puede dar de una de dos maneras: (1) como nombres de columna, o (2) como una fórmula de dos lados donde el LHS contiene un nombre de columna y el RHS contiene los valores de columna para segmentar.

Como ejemplo del primer tipo de expresión que se puede utilizar, vars(a_column) segmentará la tabla de destino en la forma en que estén presentes muchos valores únicos en la columna llamada a_column. Esto es excelente si cada valor único en una columna en particular (como diferentes ubicaciones o diferentes fechas) requiere su propia validación repetida.

Con una fórmula, podemos ser más selectivos con los valores de columna que se deben usar para la segmentación. El uso de a_column ~ c("group_1", "group_2") intentará obtener dos segmentos donde uno es una porción de datos donde el valor "group_1" existe en la columna llamada "a_column", y el otro es un segmento donde existe "group_2" en la misma columna. Cada grupo de filas resuelto a partir de la fórmula dará como resultado un paso de validación independiente.

Si hay varias columns especificadas, el número potencial de pasos de validación será m columnas multiplicadas por n segmentos resueltos.

La segmentación siempre ocurrirá después de que se apliquen las preconditions (es decir, declaraciones que mutan la tabla de destino), si las hay. Con este tipo de combinación, es posible generar etiquetas para la segmentación usando una expresión para preconditions y hacer referencia a esas etiquetas en segments sin tener que generar una versión separada de la tabla de destino.

Actions

A menudo, querremos especificar actions para la validación. Este argumento, presente en cada función de validación, toma un objeto de lista especialmente diseñado que es mejor producido por la función action_levels(). Lea la documentación de esa función para obtener información sobre cómo crear reacciones a niveles de falla por encima del umbral en la validación. La esencia básica es que querrá al menos un nivel de umbral único (especificado como la fracción de unidades de prueba fallidas o un valor absoluto), a menudo utilizando el argumento warn_at. Esto es especialmente cierto cuando x es un objeto de tabla porque, de lo contrario, no sucede nada. Para las funciones de tipo col_vals_*(), usar action_levels(warn_at = 0.25) o action_levels(stop_at = 0.25) son buenas opciones dependiendo de la situación (la primera produce una advertencia cuando una cuarta parte de la prueba total las unidades fallan, las otras stop() en el mismo nivel de umbral).

Briefs

¿Quiere describir este paso de validación con algún detalle? Tenga en cuenta que esto sólo es útil si x es un objeto agent. Si ese es el caso, use un texto brief que se ajuste a la situación. No se preocupe si no quiere hacerlo. Un autobrief se activa cuando brief = NULL y el texto luego se generará automáticamente.

YAML

Se puede escribir un agente pointblank en YAML con yaml_write() y el YAML resultante se puede usar para regenerar un agente (con yaml_read_agent()) o interrogar la tabla de destino (a través de yaml_agent_interrogate()). Cuando col_vals_make_subset() se representa en YAML (bajo la clave de nivel superior steps como un miembro de la lista), la sintaxis sigue de cerca la firma de la función de validación. A continuación se muestra un ejemplo de cómo una llamada compleja de col_vals_make_subset() como paso de validación se expresa en código R y en la representación YAML correspondiente.

# Código R
agent %>% 
  col_vals_make_subset(
    columns = vars(a),
    set = c(1, 2, 3, 4),
    preconditions = ~ . %>% dplyr::filter(a < 10),
    segments = b ~ c("group_1", "group_2"),
    actions = action_levels(warn_at = 0.1, stop_at = 0.2),
    label = "El paso `col_vals_make_subset()`.",
    active = FALSE
  )

# Representación YAML
steps:
- col_vals_make_subset:
    columns: vars(a)
   set:
   - 1.0
   - 2.0
   - 3.0
   - 4.0
   preconditions: ~. %>% dplyr::filter(a < 10)
   segments: b ~ c("group_1", "group_2")
   actions:
     warn_fraction: 0.1
     stop_fraction: 0.2
   label: El paso `col_vals_make_subset()`.
   active: false

In practice, both of these will often be shorter as only the columns, and set arguments require values. Arguments with default values won't be written to YAML when using yaml_write() (though it is acceptable to include them with their default when generating the YAML by other means). It is also possible to preview the transformation of an agent to YAML without any writing to disk by using the yaml_agent_string() function.

Function ID

2-12

See also

Examples

# The `small_table` dataset in the
# package will be used to validate that
# column values are part of a given set

# A: Usando un `agent` con funciones de
#    validación y luego `interrogate()`

# Validate that the distinct set of values
# in column `f` contains at least the
# subset defined as `low` and `high` (the
# column actually has both of those and
# some `mid` values)
agent <-
  create_agent(small_table) %>%
  col_vals_make_subset(
    vars(f), c("low", "high")
  ) %>%
  interrogate()
  
# Determine if this validation
# had no failing test units (there
# are 2 test units, one per element
# in the `set`)
all_passed(agent)
#> [1] TRUE

# Llamar a `agent` en la consola imprime
# el informe del agente; pero podemos
# obtener un objeto `gt_tbl` directamente
# con `get_agent_report(agent)`

# B: Usando la función de validación
#    directamente en los datos
#    (sin `agent`)

# Esta forma de utilizar las funciones
# de validación actúa como un filtro de
# datos: los datos se pasan a través,
# pero deben `stop()` si hay una sola
# unidad de prueba que falla; el
# comportamiento de los efectos
# secundarios se puede personalizar con
# la opción `actions`
small_table %>%
  col_vals_make_subset(
    vars(f), c("low", "high")
  ) %>%
  dplyr::pull(f) %>%
  unique()
#> [1] "high" "low"  "mid" 

# C: Usando la función de expectativa

# Con el formulario `expect_*()`,
# necesitamos ser más exactos y
# proporcionar una columna a la vez;
# esto se usa principalmente en
# pruebas `testthat`
expect_col_vals_make_subset(
  small_table,
  vars(f), c("low", "high")
)

# D: Usando la función de prueba

# Con la forma `test_*()`, deberíamos
# obtener un único valor lógico devuelto
small_table %>%
  test_col_vals_make_subset(
    vars(f), c("low", "high")
  )
#> [1] TRUE