Realice múltiples validaciones por filas para la validez conjunta

La función de validación conjointly(), la función de expectativa expect_conjointly() y la función de prueba test_conjointly() comprueban si las unidades de prueba en cada índice (típicamente cada fila) pasan todas las validaciones múltiples. Podemos usar funciones de validación que validan las unidades de fila (la serie col_vals_*()), verificar la existencia de la columna (col_exists()), o validar el tipo de columna (la serie col_is_*()). Debido a la restricción impuesta sobre las funciones de validación permitidas, el conjunto de unidades de prueba está compuesto por filas de la tabla (después de que se hayan aplicado las preconditions comunes) o son unidades de prueba únicas (para aquellas funciones que validan columnas).

Cada una de las funciones usadas en un paso de validación conjointly() (compuesto usando múltiples llamadas a funciones de validación) finalmente realiza una prueba por filas de si todas las subvalidaciones reportaron un aprobado para las mismas unidades de prueba. En la práctica, un ejemplo de validación conjunta es probar si los valores de la columna a son mayores que un valor específico, mientras que los valores adyacentes en la columna b se encuentran dentro de un rango específico. Las funciones de validación que serán parte de la validación conjunta deben ser suministradas como fórmulas R unilaterales (usando un ~ inicial y con un . como el objeto de datos). La función de validación se puede usar directamente en una tabla de datos o con un objeto agent (técnicamente, un objeto ptblank_agent) mientras que las funciones de expectativa y prueba solo se pueden usar con una tabla de datos.

Uso

conjointly(
  x,
  ...,
  .list = list2(...),
  preconditions = NULL,
  segments = NULL,
  actions = NULL,
  step_id = NULL,
  label = NULL,
  brief = NULL,
  active = TRUE
)

expect_conjointly(
  object,
  ...,
  .list = list2(...),
  preconditions = NULL,
  threshold = 1
)

test_conjointly(
  object,
  ...,
  .list = list2(...),
  preconditions = NULL,
  threshold = 1
)

Argumentos

x

Un data.frame, tibble (tbl_df o tbl_dbi), Spark DataFrame (tbl_spark), o un agent objeto de clase ptblank_agent que se crea con create_agent().

...

Una colección de fórmulas unilaterales que constan de funciones de validación que validan unidades de fila (la serie col_vals_*()), existencia de columna (col_exists()) o tipo de columna (la serie col_is_*()). Un ejemplo de esto es ~ col_vals_gte(., vars(a), 5.5), ~ col_vals_not_null(., vars(b)).

.list

Permite el uso de una lista como alternativa de entrada a ....

preconditions

Una expresión opcional para mutar la tabla de entrada antes de continuar con la validación. Esto se puede proporcionar como una fórmula R unilateral usando un ~ inicial (por ejemplo, ~ . %>% dplyr::mutate(col = col + 10) o como una función (por ejemplo, function (x) dplyr::mutate(x, col = col + 10). Consulte la sección Preconditions para obtener más información.

segments

Una expresión opcional o un conjunto de expresiones (contenidas en una lista) que sirven para segmentar la tabla de destino por valores de columna. Cada expresión se puede dar de una de dos maneras: (1) como nombres de columna, o (2) como una fórmula de dos lados donde el LHS contiene un nombre de columna y el RHS contiene los valores de columna para segmentar. Consulte la sección Segments para obtener más detalles sobre esto.

actions

Una lista que contiene los niveles de umbral para que el paso de validación pueda reaccionar en consecuencia al superar los niveles establecidos. Esto se creará con la función auxiliar action_levels().

step_id

Uno o más identificadores opcionales para los pasos de validación únicos o múltiples generados al llamar a una función de validación. El uso de ID de pasos sirve para distinguir los pasos de validación entre sí y brinda la oportunidad de proporcionar una etiqueta más significativa en comparación con el índice de pasos. De forma predeterminada, es NULL, y pointblank generará automáticamente el valor de ID de paso (basado en el índice de paso) en este caso. Se pueden proporcionar uno o más valores, y el número exacto de valores de ID debe (1) coincidir con el número de pasos de validación que producirá la llamada a la función de validación (influenciado por el número de columns proporcionadas), (2) ser un ID cadena no utilizada en ningún paso de validación anterior, y (3) ser un vector con valores únicos.

label

Una etiqueta opcional para el paso de validación. Esta etiqueta aparece en el informe del agent y, para una mejor apariencia, debe ser breve.

brief

Una descripción opcional basada en texto para el paso de validación. Si no se proporciona nada aquí, el objeto agent genera un autobrief, utilizando el lenguaje proporcionado en el argumento lang de create_agent() (que por defecto es "en" o inglés). El autobrief incorpora detalles del paso de validación, por lo que a menudo es la opción preferida en la mayoría de los casos (donde un label podría ser más adecuada para describir sucintamente la validación).

active

Un valor lógico que indica si el paso de validación debe estar activo. Si la función de validación está trabajando con un objeto agent, FALSE hará que el paso de validación esté inactivo (aún informando su presencia y manteniendo los índices de los pasos sin cambios). Si la función de validación operará directamente en los datos (sin participación de agent), entonces cualquier paso con active = FALSE simplemente pasará los datos sin validación alguna. Aparte de un vector lógico, una fórmula R unilateral que usa un ~ inicial se puede usar con . (que sirve como la tabla de datos de entrada) para evaluar a un solo valor lógico. Con este enfoque, la función pointblank has_columns() se puede utilizar para determinar si se debe activar un paso de validación sobre la base de una o más columnas existentes en la tabla (por ejemplo, ~ . %>% has_columns(vars(d, e))). El valor predeterminado de active es TRUE.

object

Un data.frame, tibble (tbl_df o tbl_dbi) o Spark DataFrame (tbl_spark) que sirve como tabla de destino para la función de expectativa o la función de prueba.

threshold

Un valor de umbral de falla simple para usar con las variantes de función expectativa (expect_) y prueba (test_). De forma predeterminada, se establece en 1, lo que significa que cualquier unidad de falla en la validación de datos da como resultado una falla general de la prueba. Los números enteros más allá de 1 indican que cualquier unidad defectuosa hasta ese valor de umbral absoluto dará como resultado una thatthat prueba o evalúe como TRUE. Asimismo, los valores fraccionarios (entre 0 y 1) actúan como un umbral de falla proporcional, donde 0.15 significa que el 15 por ciento de las unidades de prueba que fallan dan como resultado una falla general de la prueba.

Valor

Para la función de validación, el valor de retorno es un objeto ptblank_agent o un objeto de tabla (dependiendo de si se pasó un objeto agent o una tabla a x). La función de expectativa devuelve invisiblemente su entrada pero, en el contexto de los datos de prueba, la función se llama principalmente por sus posibles efectos secundarios (por ejemplo, falla de señalización). La función de prueba devuelve un valor lógico.

Nombres de columnas

Si proporciona varios nombres de columna en cualquiera de los pasos de validación proporcionados, el resultado será una expansión de los pasos de subvalidación a ese número de nombres de columna. Aparte de los nombres de las columnas entre comillas y en vars(), las funciones auxiliares tidyselect están disponibles para especificar columnas. Son: starts_with(), ends_with(), contains(), matches() y everything().

Preconditions

Proporcionar expresiones como preconditions significa que pointblank preprocesará la tabla de destino durante la interrogación como paso preparatorio. Puede suceder que una validación en particular requiera una columna calculada, algún filtrado de filas o la adición de columnas a través de una combinación, etc. Especialmente para un informe basado en agent, esto puede ser ventajoso ya que podemos desarrollar un gran plan de validación con una sola tabla de destino y realice ajustes menores en ella, según sea necesario, a lo largo del camino.

La mutación de la tabla está totalmente aislada en el alcance de los pasos de validación en los que se utilizan las preconditions. Aquí se sugiere usar el código dplyr ya que las declaraciones se pueden traducir a SQL si es necesario (es decir, si la tabla de destino reside en una base de datos). El código se proporciona más fácilmente como una fórmula R unilateral (utilizando un ~ inicial). En la representación de la fórmula, el . sirve como la tabla de datos de entrada que se va a transformar (por ejemplo, ~ . %>% dplyr::mutate(col_b = col_a + 10)). Alternativamente, se podría proporcionar una función (por ejemplo, function(x) dplyr::mutate(x, col_b = col_a + 10)).

Segmentos

Al usar el argumento segments, es posible definir una validación particular con segmentos (o porciones de fila) de la tabla de destino. Una expresión opcional o un conjunto de expresiones que sirven para segmentar la tabla de destino por valores de columna. Cada expresión se puede dar de una de dos maneras: (1) como nombres de columna, o (2) como una fórmula de dos lados donde el LHS contiene un nombre de columna y el RHS contiene los valores de columna para segmentar.

Como ejemplo del primer tipo de expresión que se puede utilizar, vars(a_column) segmentará la tabla de destino en la forma en que estén presentes muchos valores únicos en la columna llamada a_column. Esto es excelente si cada valor único en una columna en particular (como diferentes ubicaciones o diferentes fechas) requiere su propia validación repetida.

Con una fórmula, podemos ser más selectivos con los valores de columna que se deben usar para la segmentación. El uso de a_column ~ c("group_1", "group_2") intentará obtener dos segmentos donde uno es una porción de datos donde el valor "group_1" existe en la columna llamada "a_column", y el otro es un segmento donde existe "group_2" en la misma columna. Cada grupo de filas resuelto a partir de la fórmula dará como resultado un paso de validación independiente.

Si hay varias columns especificadas, el número potencial de pasos de validación será m columnas multiplicadas por n segmentos resueltos.

La segmentación siempre ocurrirá después de que se apliquen las preconditions (es decir, declaraciones que mutan la tabla de destino), si las hay. Con este tipo de combinación, es posible generar etiquetas para la segmentación usando una expresión para preconditions y hacer referencia a esas etiquetas en segments sin tener que generar una versión separada de la tabla de destino.

Actions

A menudo, querremos especificar actions para la validación. Este argumento, presente en cada función de validación, toma un objeto de lista especialmente diseñado que es mejor producido por la función action_levels(). Lea la documentación de esa función para obtener información sobre cómo crear reacciones a niveles de falla por encima del umbral en la validación. La esencia básica es que querrá al menos un nivel de umbral único (especificado como la fracción de unidades de prueba fallidas o un valor absoluto), a menudo utilizando el argumento warn_at. Esto es especialmente cierto cuando x es un objeto de tabla porque, de lo contrario, no sucede nada. Para las funciones de tipo col_vals_*(), usar action_levels(warn_at = 0.25) o action_levels(stop_at = 0.25) son buenas opciones dependiendo de la situación (la primera produce una advertencia cuando una cuarta parte de la prueba total las unidades fallan, las otras stop()s en el mismo nivel de umbral).

Briefs

¿Quiere describir este paso de validación con algún detalle? Tenga en cuenta que esto sólo es útil si x es un objeto agent. Si ese es el caso, use un texto brief que se ajuste a la situación. No se preocupe si no quiere hacerlo. Un autobrief se activa cuando brief = NULL y el texto luego se generará automáticamente.

YAML

Se puede escribir un agente pointblank en YAML con yaml_write() y el YAML resultante se puede usar para regenerar un agente (con yaml_read_agent()) o interrogar la tabla de destino (a través de yaml_agent_interrogate()). Cuando conjointly() se representa en YAML (bajo la clave de nivel superior steps como un miembro de la lista), la sintaxis sigue de cerca la firma de la función de validación. A continuación se muestra un ejemplo de cómo una llamada compleja de conjointly() como paso de validación se expresa en código R y en la representación YAML correspondiente.

# Código R
agent %>% 
  conjointly(
    ~ col_vals_lt(., vars(a), 8),
    ~ col_vals_gt(., vars(c), vars(a)),
    ~ col_vals_not_null(., vars(b)),
    preconditions = ~ . %>% dplyr::filter(a < 10),
    segments = b ~ c("group_1", "group_2"),
    actions = action_levels(warn_at = 0.1, stop_at = 0.2), 
    label = "El paso `conjointly()`.",
    active = FALSE
  )

# Representación YAML
steps:
- conjointly:
    fns:
    - ~col_vals_lt(., vars(a), 8)
    - ~col_vals_gt(., vars(c), vars(a))
    - ~col_vals_not_null(., vars(b))
    preconditions: ~. %>% dplyr::filter(a < 10)
    segments: b ~ c("group_1", "group_2")
    actions:
      warn_fraction: 0.1
      stop_fraction: 0.2
    label: El paso `conjointly()`.
    active: false

En la práctica, ambos serán a menudo más cortos, ya que solo son necesarias las expresiones para los pasos de validación. Los argumentos con valores predeterminados no se escribirán en YAML cuando se use yaml_write() (aunque es aceptable incluirlos con sus valores predeterminados al generar el YAML por otros medios). También es posible obtener una vista previa de la transformación de un agente a YAML sin escribir en el disco usando la función yaml_agent_string().

Function ID

2-31

Ver también

Ejemplos

# Para todos los ejemplos aquí, usaremos
# una tabla simple con tres columnas
# numéricas (`a`, `b` y `c`); esta es
# una tabla muy básica, pero será más
# útil cuando se expliquen las cosas
# más adelante
tbl <-
  dplyr::tibble(
    a = c(5, 2, 6),
    b = c(3, 4, 6),
    c = c(9, 8, 7)
  )
  
tbl
#> # A tibble: 3 × 3
#>       a     b     c
#>   <dbl> <dbl> <dbl>
#> 1     5     3     9
#> 2     2     4     8
#> 3     6     6     7
  
# A: Usando un `agent` con funciones de
#    validación y luego `interrogate()`

# Valida una serie de cosas fila por fila
# utilizando funciones de validación del
# tipo `col_vals*` (todas tienen el mismo
# número de unidades de prueba): (1) los
# valores de `a` son menores que `8`,
# (2) los valores de `c` son mayores que
# los valores adyacentes de `a`, y (3) no
# hay valores `NA` en `b`.
agent <-
  create_agent(tbl = tbl) %>%
  conjointly(
    ~ col_vals_lt(., vars(a), value = 8),
    ~ col_vals_gt(., vars(c), value = vars(a)),
    ~ col_vals_not_null(., vars(b))
    ) %>%
  interrogate()
  
# Determine si esta validación no tuvo
# unidades de prueba fallidas (hay 3
# unidades de prueba, una para cada fila)
all_passed(agent)
#> [1] TRUE

# Llamar a `agent` en la consola imprime
# el informe del agente; pero podemos
# obtener un objeto `gt_tbl` directamente
# con `get_agent_report(agent)`

# ¿Que esta pasando? Piense en que hay
# tres validaciones paralelas, cada una
# de las cuales produce una columna de
# valores `TRUE` o `FALSE` ("pasa" o
# "falla") y alinéelos uno al lado del
# otro, cualquier fila con cualquier
# valor `FALSE` da como resultado una
# unidad de prueba conjunta "fallida"

# B: Usando la función de validación
#    directamente en los datos
#    (sin `agent`)

# Esta forma de utilizar las funciones
# de validación actúa como un filtro de
# datos: los datos se pasan a través,
# pero deben `stop()` si hay una sola
# unidad de prueba que falla; el
# comportamiento de los efectos
# secundarios se puede personalizar con
# la opción `actions`
tbl %>%
  conjointly(
    ~ col_vals_lt(., vars(a), value = 8),
    ~ col_vals_gt(., vars(c), value = vars(a)),
    ~ col_vals_not_null(., vars(b))
  )
#> # A tibble: 3 × 3
#>       a     b     c
#>   <dbl> <dbl> <dbl>
#> 1     5     3     9
#> 2     2     4     8
#> 3     6     6     7

# C: Usando la función de expectativa

# Con el formulario `expect_*()`,
# necesitamos ser más exactos y
# proporcionar una columna a la vez;
# esto se usa principalmente en
# pruebas `testthat`
expect_conjointly(
  tbl,
  ~ col_vals_lt(., vars(a), value = 8),
  ~ col_vals_gt(., vars(c), value = vars(a)),
  ~ col_vals_not_null(., vars(b))
)

# D: Usando la función de prueba

# Con la forma `test_*()`, deberíamos
# obtener un único valor lógico devuelto
tbl %>%
  test_conjointly(
    ~ col_vals_lt(., vars(a), value = 8),
    ~ col_vals_gt(., vars(c), value = vars(a)),
    ~ col_vals_not_null(., vars(b))
  )
#> [1] TRUE