Create a pointblank agent object

Creates an agent object, which is used in a data quality reporting workflow. The overall aim of this workflow is to generate some reporting on the state of data quality for a target table. We can supply as many validation step functions as the user wishes to write to get validation coverage on that table. We supply a single table to create_agent(), which becomes the sole focus of the data quality analysis. After application of one or more validation step functions, we need to use the interrogate() function to complete the process; the validation step functions, when called on an agent, are merely instructions up to that point. The agent then has information on how the interrogation went. Reporting of the interrogation can be performed with the get_agent_report() function. If we just need to know whether all validations completely passed (i.e., all steps had no failing test units), the all_passed() function should be used.

create_agent(
  tbl,
  name = NULL,
  actions = NULL,
  end_fns = NULL,
  embed_report = FALSE,
  reporting_lang = NULL
)

Arguments

tbl
The input table. This can be a data frame, a tibble, or a tbl_dbi object.

name
An optional name for the validation plan that the agent will eventually carry out during the interrogation process. If no value is provided, a name will be generated based on the current system time.

actions
A list containing threshold levels so that all validation steps can react accordingly when exceeding the set levels. This is to be created with the action_levels() helper function. Should an action levels list be used for specific validation step, any default set here will be overridden.

end_fns
A list of functions that should be performed at the end of an interrogation.

embed_report
An option to embed a gt-based validation report into the ptblank_agent object. If FALSE (the default) then the table object will be not generated and available with the agent upon returning from the interrogation.

reporting_lang
The language to use for automatic creation of briefs (short descriptions for each validation step). By default, NULL will create English ("en") text. Other options include French ("fr"), German ("de"), Italian ("it"), and Spanish ("es").

Value

A ptblank_agent object.

Function ID

1-2

See also

Examples

# Create a simple table with a
# column of numerical values
tbl <-
  dplyr::tibble(a = c(5, 7, 8, 7))

# Create a pointblank `agent` object
agent <- create_agent(tbl = tbl)

# Then, as with any `ptblank_agent`
# object, we can add validation steps
# to the validation plan and then
# eventually use `interrogate()`
# to perform the validations; here,
# with a single validation step, we
# expect that values in column `a`
# are always greater than 4
agent <-
  agent %>%
  col_vals_gt(vars(a), 4) %>%
  interrogate()