Create a pointblank metadata object

The create_metadata() function creates a metadata object, which is used in a metadata management workflow. The overall aim of this workflow is to record, collect, and generate useful information on data tables. We can supply as many metadata parameters that are useful for describing a particular data table. The metadata object created by the create_metadata() function takes metadeta-focused functions (the meta_*() series of functions).

create_metadata(
  tbl = NULL,
  read_fn = NULL,
  agent = NULL,
  tbl_name = NULL,
  label = NULL,
  lang = NULL,
  locale = NULL
)

Arguments

tbl
The input table. This can be a data frame, a tibble, a tbl_dbi object, or a tbl_spark object. Alternatively, a function can be used to read in the input data table with the read_fn argument (in which case, tbl can be NULL).

read_fn
A function that's used for reading in the data. Even if a tbl is provided, this function will be invoked to obtain the data (i.e., the read_fn takes priority). There are two ways to specify a read_fn: (1) using a function (e.g., function() { <table reading code> }) or, (2) with an R formula expression.

agent
A pointblank agent object. This object can be used instead of supplying a table in tbl or a table-reading function in read_fn.

tbl_name
A optional name to assign to the input table object. If no value is provided, a name will be generated based on whatever information is available.

label
An optional label for the metadata report. If no value is provided, a label will be generated based on the current system time. Markdown can be used here to make the label more visually appealing (it will appear in the header area of the metadata report).

lang
The language to use for the metadata report (a summary table that provides all of the available metadata for the table. By default, NULL will create English ("en") text. Other options include French ("fr"), German ("de"), Italian ("it"), Spanish ("es"), Portuguese, ("pt"), and Chinese ("zh").

locale
An optional locale ID to use for formatting values in the metadata report summary table according the locale's rules. Examples include "en_US" for English (United States) and "fr_FR" for French (France); more simply, this can be a language identifier without a country designation, like "es" for Spanish (Spain, same as "es_ES").

Value

A ptblank_metadata object.

Examples

# Let's walk through how we can
# generate some useful metadata for an
# extremely small table; it's actually
# called `small_table` and we can find
# it as a dataset in this package
small_table

#> # A tibble: 13 x 8
#>    date_time           date           a b             c      d e     f    
#>    <dttm>              <date>     <int> <chr>     <dbl>  <dbl> <lgl> <chr>
#>  1 2016-01-04 11:00:00 2016-01-04     2 1-bcd-345     3  3423. TRUE  high 
#>  2 2016-01-04 00:32:00 2016-01-04     3 5-egh-163     8 10000. TRUE  low  
#>  3 2016-01-05 13:32:00 2016-01-05     6 8-kdg-938     3  2343. TRUE  high 
#>  4 2016-01-06 17:23:00 2016-01-06     2 5-jdo-903    NA  3892. FALSE mid  
#>  5 2016-01-09 12:36:00 2016-01-09     8 3-ldm-038     7   284. TRUE  low  
#>  6 2016-01-11 06:15:00 2016-01-11     4 2-dhe-923     4  3291. TRUE  mid  
#>  7 2016-01-15 18:46:00 2016-01-15     7 1-knw-093     3   843. TRUE  high 
#>  8 2016-01-17 11:27:00 2016-01-17     4 5-boe-639     2  1036. FALSE low  
#>  9 2016-01-20 04:30:00 2016-01-20     3 5-bce-642     9   838. FALSE high 
#> 10 2016-01-20 04:30:00 2016-01-20     3 5-bce-642     9   838. FALSE high 
#> 11 2016-01-26 20:07:00 2016-01-26     4 2-dmx-010     7   834. TRUE  low  
#> 12 2016-01-28 02:51:00 2016-01-28     2 7-dmx-010     8   108. FALSE low  
#> 13 2016-01-30 11:23:00 2016-01-30     1 3-dka-303    NA  2230. TRUE  high 

# We create a pointblank `metadata`
# object with `create_metadata()`;
# let's use that function with the
# `small_table` dataset
metadata <- create_metadata(small_table)

# This function creates some metadata
# without any extra help by profiling
# the supplied table object; it adds
# the sections: (1) 'table' and
# (2) 'columns' and we can print the
# object to see the metadata report
# metadata

# Alternatively, we can visualize this
# metadata into the same report by using
# `get_metadata_report()`
report <- get_metadata_report(metadata)
class(report)

#> [1] "gt_tbl" "list"