Exportar un agent, informant, multiagent o escaneo de tabla a HTML

El agent, el informant, el multiagent y el objeto de exploración de la tabla pueden escribirse fácilmente como HTML con export_report(). Además, cualquier objeto de informe del agente, informante y multiagente (generados mediante get_agent_report(), get_informant_report() y get_multiagent_report()) puede ser proporcionado aquí para su exportación en HTML. Cada documento HTML escrito en el disco es autocontenido y fácilmente visualizable en un en un navegador web.

Uso

export_report(x, filename, path = NULL, quiet = FALSE)

Argumentos

x

Un objeto agent de la clase ptblank_agent, un informant de la clase ptblank_informant, un multiagente de la clase ptblank_multiagent, un escaneo de tabla de la clase ptblank_tbl_scan, o, objetos de informe personalizados (ptblank_agent_report, ptblank_informant_report, ptblank_multiagent_report.wide, ptblank_multiagent_report.long).

filename

El nombre del archivo que se creará en el disco para la exportación HTML del objeto proporcionado. Se recomienda incluir la extensión ".html".

path

Una ruta opcional en la que se debe guardar el archivo (esto es se combina automáticamente con filename).

quiet

¿La función no debe informar cuando se escribe el archivo? En por defecto es FALSE.

Valor

Devuelve invisiblemente TRUE si el archivo ha sido escrito.

Function ID

9-3

Ver también

Ejemplos

if (interactive()) {

# A: Escribir un informe de agente como HTML

# Vamos a recorrer el proceso de
# (1) desarrollar un agente con un plan de
# validación (que se utilizará para el
# análisis de la calidad de los datos de
# la tabla `small_table`), (2) interrogar
# al agente con la función `interrogate()`,
# y (3) escribir el agente y toda su
# información en un archivo

# La creación de un objeto `action_levels`
# es un paso común del flujo de trabajo al
# crear un agente pointblank; designamos
# los umbrales de fallo a los estados
# `warn`, `stop` y `notify` utilizando
# `action_levels()`
al <- 
  action_levels(
    warn_at = 0.10,
    stop_at = 0.25,
    notify_at = 0.35
  )

# Ahora cree un objeto pointblank `agent`
# y déle el objeto `al` (que sirve como
# valor por defecto para todos los pasos
# de validación que pueden ser anulados);
# el datos serán referenciados en un
# `read_fn`.
agent <- 
  create_agent(
    read_fn = ~ small_table,
    tbl_name = "small_table",
    label = "`export_report()`",
    actions = al
  )

# Luego, como con cualquier objeto
# agente, podemos añadir pasos al plan
# de validación utilizando tantas
# funciones de validación como que
# queramos; entonces, `interrogate()`
agent <-
  agent %>% 
  col_exists(vars(date, date_time)) %>%
  col_vals_regex(
    vars(b), regex = "[0-9]-[a-z]{3}-[0-9]{3}"
  ) %>%
  rows_distinct() %>%
  col_vals_gt(vars(d), value = 100) %>%
  col_vals_lte(vars(c), value = 5) %>%
  interrogate()

# El informe del agente se puede
# escribir en un archivo HTML con
# `export_report()`
export_report(
  agent,
  filename = "agent-small_table.html"
)

# Si estás escribiendo constantemente
# informes de agente cuando compruebas
# periódicamente los datos, podríamos
# hacer uso de `affix_date()` o
# `affix_datetime()` dependiendo de la
# granularidad que necesites; aquí
# tienes un ejemplo que escribe el
# fichero con el formato
# 'agent-small_table-YYYY-mm-dd_HH-MM-SS'
export_report(
  agent,
  filename = affix_datetime(
    "agent-small_table.html"
  )
)

# B: Escribir un informe de informante
#    como HTML

# Vamos a repasar el proceso de (1)
# crear un objeto informante que
# describa mínimamente la tabla
# `small_table`, (2) asegurarnos de que
# se capturan los datos de la tabla de
# destino utilizando la función
# `incorporate()`, y (3) escribir el
# informe informante en HTML

# Crear un objeto `informant` en
# blanco con `create_informant()` y
# el conjunto de datos `small_table`;
# `incorporate()` para que los
# fragmentos de información se integren
# en el texto
informant <- 
  create_informant(
    read_fn = ~ small_table,
    tbl_name = "small_table",
    label = "`export_report()`"
  ) %>%
  info_snippet(
    snippet_name = "high_a",
    fn = snip_highest(column = "a")
  ) %>%
  info_snippet(
    snippet_name = "low_a",
    fn = snip_lowest(column = "a")
  ) %>%
  info_columns(
    columns = vars(a),
    info = "From {low_a} to {high_a}."
  ) %>%
  info_columns(
    columns = starts_with("date"),
    info = "Time-based values."
  ) %>%
  info_columns(
    columns = "date",
    info = "The date part of `date_time`."
  ) %>%
  incorporate()

# El informe del informante puede
# escribirse en un archivo HTML con
# `export_report()`; hagámoslo con
# `affix_date()` para que el nombre del
# archivo tenga una marca de fecha
export_report(
  informant,
  filename = affix_date(
    "informant-small_table.html"
  )
)

# C: Escribir una exploración de l
#    tabla como HTML

# Podemos obtener un informe qu
# describa todos los datos de la
# tabla `storms`
tbl_scan <- scan_data(tbl = dplyr::storms)

# El objeto de escaneo de la tabla
# puede ser escrito en un archivo
# HTML con `export_report()`
export_report(
  tbl_scan,
  filename = "tbl_scan-storms.html"
)

}