Obtener una tabla de un archivo local o remoto

Si su tabla de destino se encuentra en un archivo, almacenado local o remotamente, la función file_tbl() puede hacer posible el acceso a la misma en una sola llamada a la función. Los tipos de archivo compatibles con esta función son: CSV (.csv), TSV (.tsv), RDA (.rda) y RDS (.rds). Esta función genera un objeto tbl_dbl en memoria, que puede utilizarse como tabla de destino para create_agent() y create_informant(). La opción ideal para el acceso a los datos con file_tbl() es utilizar esta función como parámetro read_fn en cualquiera de las funciones create_*() antes mencionadas. Esto puede hacerse utilizando un ~ inicial (por ejemplo,. read_fn = ~file_tbl(...)).

En el caso de uso de datos remotos, podemos especificar una URL que empiece por http://, https://, etc., y que termine con el archivo que contiene la tabla de datos. Si los archivos de datos están disponibles en un repositorio de GitHub, podemos utilizar la función from_github() para especificar el nombre y la ubicación de los datos de la tabla en un repositorio.

Uso

file_tbl(file, type = NULL, ..., keep = FALSE, verify = TRUE)

Argumentos

file

La ruta completa del archivo que lleva a una tabla de datos compatible, ya sea en el sistema del usuario o en una URL http://, https://, ftp:// o ftps://. En el caso de un archivo alojado en un repositorio de GitHub, una llamada a la función from_github() se puede utilizar aquí.

type

El tipo de archivo. Normalmente se infiere por la extensión del archivo y por defecto es NULL para indicar que la extensión dictará el tipo de lectura del archivo que se realiza internamente. Sin embargo, si no hay extensión (y las extensiones válidas son .csv, .tsv, .rda y .rds), podemos podemos proporcionar el tipo como csv, tsv, rda, o rds.

...

Opciones pasadas a la función read_csv() o read_tsv() de readr. Ambas funciones tienen los mismos argumentos y se utilizará una u otra internamente en función de la extensión del archivo o de un valor explícito dado a type.

keep

En el caso de un archivo descargado, ¿debe almacenarse en el directorio de trabajo (keep = TRUE) o debe descargarse en un directorio temporal? Por defecto, esto es FALSE.

verify

Si es TRUE (el valor por defecto) entonces se realizará una verificación del objeto de datos con la clase data.frame.

Valor

Un objeto tbl_df.

Function ID

1-7

Ver también

Ejemplos

# Un archivo CSV local puede obtenerse
# como un objeto tbl suministrando una
# ruta al archivo y algunas opciones
# de lectura CSV (las utilizadas por
# `readr::read_csv()`) a la función
# `file_tbl()`; para este ejemplo
# podríamos obtener una ruta a un
# archivo CSV en el paquete pointblank
# con `system.file()`:
csv_path <- 
  system.file(
    "data_files", "small_table.csv",
    package = "pointblank"
  )

# A continuación, utilice esa ruta
# en `file_tbl()` con la opción de
# especificar los tipos de columna
# en ese CSV  
tbl <- 
  file_tbl(
    file = csv_path,
    col_types = "TDdcddlc"
  )
  
# Ahora que tenemos un objeto `tbl`
# que es un tibble, se puede
# introducir en `create_agent()`
# para su validación
agent <- create_agent(tbl = tbl)

# Una estrategia diferente es
# proporcionar la llamada a la
# función de lectura de datos
# directamente a `create_agent()`:
agent <- 
  create_agent(
    read_fn = ~ file_tbl(
      file = system.file(
        "data_files", "small_table.csv",
        package = "pointblank"
      ),
      col_types = "TDdcddlc"
    )
  ) %>%
  col_vals_gt(vars(a), value = 0)

# Todas las instrucciones de lectura
# de archivos se encapsulan en el
# `read_fn` para que el agente obtenga
# siempre la versión más reciente del
# conjunto de datos (y la lógica puede
# traducirse a YAML, para uso posterior)

if (interactive()) {

# Se puede obtener un CSV de un
# repositorio público de GitHub
# utilizando la función de ayuda
# `from_github()`; creemos un agente
# que suministre una fórmula de
# preparación de tablas que obtenga el
# mismo archivo CSV del repositorio de
# GitHub para el paquete pointblank 
agent <- 
  create_agent(
    read_fn = ~ file_tbl(
      file = from_github(
        file = "inst/data_files/small_table.csv",
        repo = "rich-iannone/pointblank"
      ),
      col_types = "TDdcddlc"
    )
  ) %>%
  col_vals_gt(vars(a), value = 0) %>%
  interrogate()

# Esto interroga los datos que se
# obtuvieron del archivo fuente remoto,
# y, no hay nada que limpiar (por defecto,
# el archivo descargado va a un
# directorio temporal del sistema)

# El almacenamiento de fórmulas de
# preparación de tablas en un almacén
# de tablas facilita el trabajo con
# datos tabulares procedentes de
# archivos; a continuación se explica
# cómo generar un almacén de tablas con
# dos entradas con nombre para la tabla
# preparaciones
tbls <-
  tbl_store(
    small_table_file ~ file_tbl(
      file = system.file(
        "data_files", "small_table.csv",
        package = "pointblank"
      ),
      col_types = "TDdcddlc"
    ),
    small_high_file ~ file_tbl(
      file = system.file(
        "data_files", "small_table.csv",
        package = "pointblank"
      ),
      col_types = "TDdcddlc"
    ) %>%
      dplyr::filter(f == "high")
  )

# Ahora es fácil acceder a cualquiera
# de estas tablas (la segunda es una
# versión mutada) mediante la
# función `tbl_get()`
tbl_get("small_table_file", store = tbls)
tbl_get("small_high_file", store = tbls)

# Las fórmulas de preparación de
# tablas en `tbls` también pueden
# utilizarse en funciones con el
# argumento `read_fn`; esto es gracias
# a la función `tbl_source()`
agent <- 
  create_agent(
    read_fn = ~ tbl_source(
      "small_table_file",
      store = tbls
    )
  )

informant <- 
  create_informant(
    read_fn = ~ tbl_source(
      "small_high_file",
      store = tbls
    )
  )

}