Obtenga un informe resumido de una agent

Podemos obtener una tabla resumen informativa de un agente utilizando la función get_agent_report(). La tabla puede ser proporcionada en dos formas sustancialmente diferentes: como una tabla de visualización basada en gt (el valor por defecto), o, como un tibble. La cantidad de campos con intel es diferente dependiendo de si el agente realizó o no una interrogación (con la función interrogate()). Básicamente, antes de llamar a interrogate(), el agente contendrá sólo el plan de validación (el número de filas que tenga depende de cuántas funciones de validación se hayan suministrado como parte de ese plan). Después de la interrogación, se proporciona información sobre las unidades de prueba que pasan y que fallan, junto con indicadores sobre si se han introducido ciertos estados de fallo (siempre que se hayan establecido a través de actions). La variante de la tabla de visualización del informe del agente, la forma por defecto, tendrá las siguientes columnas:

i (sin etiquetas): el número de paso de validación

STEP: el nombre de la función de validación utilizada para el paso de validación

COLUMNS: los nombres de las columnas de destino utilizadas en el paso de validación (si procede)

VALUES: los valores utilizados en el paso de validación, cuando sea aplicable; esto podría ser como valores literales, como nombres de columnas, una expresión, un conjunto de subvalidaciones (para un paso de validación conjointly()), etc.

TBL: indica si había alguna condición previa que aplicar antes de la interrogación; si no es así, aparece una aparece una flecha hacia la derecha

EVAL: un valor de carácter que denota el resultado de cada validación evaluación de las funciones de paso durante la interrogación

N: el número total de unidades de prueba para la etapa de validación

PASS: el número de unidades de prueba que han recibido un paso.

FAIL: la fracción de unidades de prueba que recibieron un paso.

W, S, N: Indicadores que muestran si se han introducido los estados de warn, stop o notify; los estados no establecidos aparecen como guiones, los estados establecidos con umbrales aparecen como círculos sin rellenar cuando no se han introducido y rellenos cuando se superan los umbrales (los colores para W, S y N son ámbar, rojo y azul).

EXT: una columna que proporciona botones con extractos de datos para cada paso de validación en el que están disponibles las filas fallidas (como archivos CSV)

La versión pequeña de la tabla de visualización (obtenida con size = "small") omite las columnas COLUMNS, TBLyEXT`. El ancho de la tabla pequeña es de 575px; la tabla estándar tiene 875px de ancho.

Si se elige obtener un tibble (con display_table = FALSE), tendrá las siguientes columnas:

i: el número de paso de validación

type: el nombre de la función de validación utilizada para el paso de validación

columns: los nombres de las columnas de destino utilizadas en el paso de validación (si procede)

values: los valores utilizados en el paso de validación, cuando sea aplicable; esto podría ser como valores literales, como nombres de columnas, una expresión, un conjunto de subvalidaciones (para un paso de validación conjointly()), etc.

precon: indica si hay que aplicar alguna condición previa antes del interrogatorio y, en caso afirmativo, el número de declaraciones utilizadas

active: un valor lógico que indica si un paso de validación se establece como "active" durante una interrogación

eval: un valor de carácter que denota el resultado de cada validación evaluación de las funciones de paso durante la interrogación

units: el número total de unidades de prueba para la etapa de validación

n_pass: el número de unidades de prueba que han recibido un paso

f_pass: la fracción de unidades de prueba que recibieron un paso

W, S, N: valor lógico que indica si los estados warn, stop o notify fueron introducidos

extract: un valor lógico que indica si un extracto de datos está disponible para el paso de validación

Uso

get_agent_report(
  agent,
  arrange_by = c("i", "severity"),
  keep = c("all", "fail_states"),
  display_table = TRUE,
  size = "standard",
  title = ":default:",
  lang = NULL,
  locale = NULL
)

Argumentos

agent

Un objeto de agente de clase ptblank_agent.

arrange_by

Una opción para organizar las filas de la tabla de informes por el número de paso de validación ("i", el valor predeterminado), o, para organizar en orden descendente por severidad del estado de fallo (con "severity").

keep

Una opción para mantener "all" las filas de la tabla del informe (el valor predeterminado), o, mantener sólo aquellas filas que reflejan uno o más "fail_states".

display_table

¿Debe generarse una tabla de visualización? Si es TRUE (por defecto), y si el paquete gt está instalado, se mostrará una tabla de visualización del informe en el Visor. Si es FALSE, o si gt no está está disponible, entonces se devolverá un tibble.

size

El tamaño de la tabla de visualización, que puede ser "standard" (por defecto) o "small". Esto sólo se aplica a una tabla de visualización (donde display_table = TRUE).

title

Opciones para personalizar el título del informe. La opción por defecto es la palabra clave ":default:", que produce un texto de título genérico que hace referencia al paquete pointblank en el idioma regido por la opción lang. Otra opción de palabra clave es ":tbl_name:", que presenta el nombre de la tabla como título del informe. Si no se desea ningún título, se puede utilizar la opción de palabra clave ":none:". Aparte de las opciones de palabras clave, se puede proporcionar texto para el título y se pueden utilizar llamadas a glue::glue() para construir la cadena de texto. Si se proporciona texto, se interpretará como texto Markdown y se transformará internamente en HTML. Para evitar dicha transformación, utilice text en I() para indicar explícitamente que el texto suministrado no debe ser transformado.

lang

El lenguaje que se utilizará para la creación automática de resúmenes (descripciones breves para cada paso de validación) y para el agent report (una tabla resumen que proporciona el plan de validación y los resultados de la interrogación. De forma predeterminada, NULL creará texto en inglés ("en"). Otras opciones incluyen francés ("fr"), alemán ("de"), italiano ("it"), español ("es"), portugués ("pt"), turco ("tr"), chino ("zh"), ruso ("ru"), polaco ("pl"), danés ("da"), sueco ("sv" ) y holandés ("nl"). Esta opción lang anulará cualquier previamente establecido (por ejemplo, por la llamada create_agent()).

locale

Un identificador opcional de la configuración regional que se utilizará para formatear los valores en la tabla de resumen del informe de agente de acuerdo con las reglas de la configuración regional. Los ejemplos incluyen "en_US" para el inglés (Estados Unidos) y "fr_FR" para el francés (Francia); más simplemente, puede ser un identificador de idioma sin designación de país, como "es" para el español (España, igual que "es_ES"). Esta opción locale anulará cualquier valor de locale previamente establecido (por ejemplo, por la opción create_agent()).

Valor

Un objeto de tabla gt si display_table = TRUE o un tibble si display_table = FALSE.

Function ID

6-2

Ver también

Ejemplos

# Crear una tabla simple con una
# columna de valores numéricos
tbl <- 
  dplyr::tibble(a = c(5, 7, 8, 5))

# Validar que los valores de la
# columna `a` sean siempre mayores
# que 4
agent <-
  create_agent(tbl = tbl) %>%
  col_vals_gt(vars(a), value = 4) %>%
  interrogate()

# Obtenga un informe basado en
# tibble del agente utilizando
# `get_agent_report()` y
# `display_table = FALSE`
agent %>%
  get_agent_report(display_table = FALSE)
#> # A tibble: 1 × 14
#>       i type  columns values precon active eval  units n_pass f_pass W     S    
#>   <int> <chr> <chr>   <chr>  <chr>  <lgl>  <chr> <dbl>  <dbl>  <dbl> <lgl> <lgl>
#> 1     1 col_… a       4      NA     TRUE   OK        4      4      1 NA    NA   
#> # … with 2 more variables: N <lgl>, extract <int>
  
# Ver un informe imprimiendo el
# objeto `agente` en cualquier momento,
# pero, devolver un objeto de tabla gt
# usando esto con `display_table = TRUE`
# (por defecto)
report <- get_agent_report(agent)
class(report)
#> [1] "ptblank_agent_report" "gt_tbl"               "list"                

# ¿Qué puedes hacer con el informe?
# Imprimirlo desde un código R Markdown,
# utilizarlo en un correo electrónico,
# ponerlo en una página web, o
# modificarlo aún más con el paquete
# **gt**

# El informe del agente como tabla
# de visualización **gt** viene en
# dos tamaños: "standard" (el
# predeterminado) y "small"
small_report <- 
  get_agent_report(
    agent = agent,
    size = "small"
  )

class(small_report)
#> [1] "ptblank_agent_report" "gt_tbl"               "list"                

# El informe estándar tiene 875px
# de ancho y el pequeño 575px