Get a table information report from an informant object

We can get a table information report from an informant object that's generated by the create_informant() function. The report is provided as a gt based display table. The amount of information shown depends on the extent of that added via the use of the info_*() functions or through direct editing of a pointblank YAML file.

get_informant_report(informant)

Arguments

informant
An informant object of class ptblank_informant.

Value

A gt table object.

Function ID

6-2

See also

Examples

# Generate an informant object using
# the `small_table` dataset
informant <- create_informant(small_table)

# This function creates some information
# without any extra help by profiling
# the supplied table object; it adds
# the sections 'table' and columns' and
# we can print the object to see the
# table information report

# Alternatively, we can get the same report
# by using `get_informant_report()`
report <- get_informant_report(informant)
class(report)

#> [1] "gt_tbl" "list"