Get a summary report using multiple agents

We can get an informative summary table from a collective of agents by using the get_multiagent_report() function. The table can be provided in either of two very different forms: as a gt based display table (the default), or, as a tibble with packed columns. The display table variant of the multiagent report, the default form, will have the following columns:

STEP: the SHA1 hash for the validation step, possibly shared among several interrogations.

subsequent columns: each column beyond STEP represents a separate interrogation from an agent object. The time stamp for the completion of each interrogation is shown as the column label.

Each step is represented with an icon standing in for the name of the validation function and the associated SHA1 hash. This is a highly trustworthy way for ascertaining which validation steps are effectively identical across interrogations. This way of organizing the report is beneficial because different agents may have used different steps and we want to track the validation results where the validation step doesn't change but the target table does.

get_multiagent_report(multiagent, display_table = TRUE, title = ":default:")

Arguments

multiagent
A multiagent object of class ptblank_multiagent.

display_table
Should a display table be generated? If TRUE (the default) a display table for the report will be shown in the Viewer. If FALSE then a tibble will be returned.

title
Options for customizing the title of the report. The default is the keyword ":default:" which produces generic title text. If no title is wanted, then the ":none:" keyword option can be used. Aside from keyword options, text can be provided for the title and glue::glue() calls can be used to construct the text string. If providing text, it will be interpreted as Markdown text and transformed internally to HTML. To circumvent such a transformation, use text in I() to explicitly state that the supplied text should not be transformed.

Value

A gt table object if display_table = TRUE or a tibble if display_table = FALSE.

Function ID

10-3

See also