Sunder the data, splitting it into 'pass' and 'fail' pieces

Validation of the data is one thing but, sometimes, you want to use the best part of the input dataset for something else. The get_sundered_data() function works with an agent object that has intel (i.e., post interrogate()) and gets either the 'pass' data piece (rows with no failing units across all row-based validation step functions), or, the 'fail' data piece (rows with at least one failing unit across the same series of validations). There are some caveats, only those validation steps with no preconditions are considered. And, the validation steps used for this must, again, be of the row-based variety (e.g., the col_vals_*() functions, and conjointly()).

get_sundered_data(agent, type = "pass", id_cols = NULL)

Arguments

agent
An agent object of class ptblank_agent. It should have had interrogate() called on it, such that the validation steps were actually carried out.

type
The desired piece of data resulting from the splitting. Options are "pass" (the default) and "fail".

id_cols
An optional specification of one or more identifying columns. When taken together, we can count on this single column or grouping of columns to distinguish rows.

Value

A list of table objects if type is NULL, or, a table object piece if a type is given.

Function ID

5-4

See also

Examples

# Create a series of three validation
# steps focus on test row values for
# the `small_table` tibble object;
# `interrogate()` immediately
agent <-
  create_agent(tbl = small_table) %>%
  col_vals_gt(vars(d), 100) %>%
  col_vals_equal(
    vars(d), vars(d),
    na_pass = TRUE
  ) %>%
  col_vals_between(
    vars(c), left = vars(a), right = vars(d),
    na_pass = TRUE
  ) %>%
  interrogate()

#> 

#> ── Interrogation Started - there are 3 steps ───────────────────────────────────

#> ✔ Step 1: OK.

#> ✔ Step 2: OK.

#> ✔ Step 3: OK.

#> 

#> ── Interrogation Completed ─────────────────────────────────────────────────────

# Get the sundered data piece that
# contains only rows that passed all
# validation steps (the default piece)
agent %>% get_sundered_data()

#> # A tibble: 9 x 8
#>   date_time           date           a b             c      d e     f    
#>   <dttm>              <date>     <int> <chr>     <dbl>  <dbl> <lgl> <chr>
#> 1 2016-01-04 11:00:00 2016-01-04     2 1-bcd-345     3  3423. TRUE  high 
#> 2 2016-01-04 00:32:00 2016-01-04     3 5-egh-163     8 10000. TRUE  low  
#> 3 2016-01-06 17:23:00 2016-01-06     2 5-jdo-903    NA  3892. FALSE mid  
#> 4 2016-01-11 06:15:00 2016-01-11     4 2-dhe-923     4  3291. TRUE  mid  
#> 5 2016-01-20 04:30:00 2016-01-20     3 5-bce-642     9   838. FALSE high 
#> 6 2016-01-20 04:30:00 2016-01-20     3 5-bce-642     9   838. FALSE high 
#> 7 2016-01-26 20:07:00 2016-01-26     4 2-dmx-010     7   834. TRUE  low  
#> 8 2016-01-28 02:51:00 2016-01-28     2 7-dmx-010     8   108. FALSE low  
#> 9 2016-01-30 11:23:00 2016-01-30     1 3-dka-303    NA  2230. TRUE  high