Given a metadata object, update and incorporate table snippets

When the metadata object has a number of snippets available (by using meta_snippet()) and the strings to use them (by using the meta_*() functions and {<snippet_name>} in the text elements), the process of incorporating aspects of the table into the metadata text can occur by using the incorporate() function. After that, the metadata will fully updated (getting the current state of table dimensions, rerendering the metadata text, etc.) and we can print the metadata object or use the get_metadata_report() function to see the metadata report.

incorporate(metadata)

Arguments

metadata
A metadata object of class ptblank_metadata.

Value

A ptblank_metadata object.

Examples

# Take the `small_table` and
# assign it to `test_table`; we'll
# modify it later
test_table <- small_table

# Generate a metadata object, add
# two snippets with `meta_snippet()`,
# add metadata with some other
# `meta_*()` functions and then
# `incorporate()` the snippets into
# the metadata text
metadata <- 
  create_metadata(
    read_fn = ~test_table,
    tbl_name = "test_table"
  ) %>%
  meta_snippet(
    snippet_name = "row_count",
    fn = ~ . %>% nrow()
  ) %>%
  meta_snippet(
    snippet_name = "col_count",
    fn = ~ . %>% ncol()
  ) %>%
  meta_columns(
    columns = vars(a),
    info = "In the range of 1 to 10. (SIMPLE)"
  ) %>%
  meta_columns(
    columns = starts_with("date"),
    info = "Time-based values (e.g., `Sys.time()`)."
  ) %>%
  meta_columns(
    columns = "date",
    info = "The date part of `date_time`. (CALC)"
  ) %>%
  meta_section(
    section_name = "rows",
    row_count = "There are {row_count} rows available."
  ) %>%
  incorporate()

#> Error in rlang::eval_tidy(.): object 'test_table' not found

# We can print the `metadata`
# object to see the metadata report
# metadata

# Let's modify `test_table` to give
# it more rows and an extra column
test_table <- 
  dplyr::bind_rows(test_table, test_table) %>%
  dplyr::mutate(h = a + c)

# Using `incorporate()` will cause
# the snippets to be reprocessed, and,
# the strings to be updated
# metadata %>% incorporate()