Does the row count match that of a different table?

The row_count_match() validation function, the expect_row_count_match() expectation function, and the test_row_count_match() test function all check whether the row count in the target table matches that of a comparison table. The validation function can be used directly on a data table or with an agent object (technically, a ptblank_agent object) whereas the expectation and test functions can only be used with a data table. The types of data tables that can be used include data frames, tibbles, database tables (tbl_dbi), and Spark DataFrames (tbl_spark). As a validation step or as an expectation, there is a single test unit that hinges on whether the row counts for the two tables are the same (after any preconditions have been applied).

Usage

row_count_match(
  x,
  tbl_compare,
  preconditions = NULL,
  segments = NULL,
  actions = NULL,
  step_id = NULL,
  label = NULL,
  brief = NULL,
  active = TRUE
)

expect_row_count_match(
  object,
  tbl_compare,
  preconditions = NULL,
  threshold = 1
)

test_row_count_match(object, tbl_compare, preconditions = NULL, threshold = 1)

Arguments

x

Un data.frame, tibble (tbl_df o tbl_dbi), Spark DataFrame (tbl_spark), o un agent objeto de clase ptblank_agent que se crea con create_agent().

tbl_compare

A table to compare against the target table in terms of row count values. This can either be a table object, a table-prep formula.This can be a table object such as a data frame, a tibble, a tbl_dbi object, or a tbl_spark object. Alternatively, a table-prep formula (~ <table reading code>) or a function (function() <table reading code>) can be used to lazily read in the table at interrogation time.

preconditions

Una expresión opcional para mutar la tabla de entrada antes de continuar con la validación. Esto se puede proporcionar como una fórmula R unilateral usando un ~ inicial (por ejemplo, ~ . %>% dplyr::mutate(col = col + 10) o como una función (por ejemplo, function (x) dplyr::mutate(x, col = col + 10). Consulte la sección Preconditions para obtener más información.

segments

Una expresión opcional o un conjunto de expresiones (contenidas en una lista) que sirven para segmentar la tabla de destino por valores de columna. Cada expresión se puede dar de una de dos maneras: (1) como nombres de columna, o (2) como una fórmula de dos lados donde el LHS contiene un nombre de columna y el RHS contiene los valores de columna para segmentar. Consulte la sección Segments para obtener más detalles sobre esto.

actions

Una lista que contiene los niveles de umbral para que el paso de validación pueda reaccionar en consecuencia al superar los niveles establecidos. Esto se creará con la función auxiliar action_levels().

step_id

Uno o más identificadores opcionales para los pasos de validación únicos o múltiples generados al llamar a una función de validación. El uso de ID de pasos sirve para distinguir los pasos de validación entre sí y brinda la oportunidad de proporcionar una etiqueta más significativa en comparación con el índice de pasos. De forma predeterminada, es NULL, y pointblank generará automáticamente el valor de ID de paso (basado en el índice de paso) en este caso. Se pueden proporcionar uno o más valores, y el número exacto de valores de ID debe (1) coincidir con el número de pasos de validación que producirá la llamada a la función de validación (influenciado por el número de columns proporcionadas), (2) ser un ID cadena no utilizada en ningún paso de validación anterior, y (3) ser un vector con valores únicos.

label

Una etiqueta opcional para el paso de validación. Esta etiqueta aparece en el informe del agent y, para una mejor apariencia, debe ser breve.

brief

Una descripción opcional basada en texto para el paso de validación. Si no se proporciona nada aquí, el objeto agent genera un autobrief, utilizando el lenguaje proporcionado en el argumento lang de create_agent() (que por defecto es "en" o inglés). El autobrief incorpora detalles del paso de validación, por lo que a menudo es la opción preferida en la mayoría de los casos (donde un label podría ser más adecuada para describir sucintamente la validación).

active

Un valor lógico que indica si el paso de validación debe estar activo. Si la función de validación está trabajando con un objeto agent, FALSE hará que el paso de validación esté inactivo (aún informando su presencia y manteniendo los índices de los pasos sin cambios). Si la función de validación operará directamente en los datos (sin participación de agent), entonces cualquier paso con active = FALSE simplemente pasará los datos sin validación alguna. Aparte de un vector lógico, una fórmula R unilateral que usa un ~ inicial se puede usar con . (que sirve como la tabla de datos de entrada) para evaluar a un solo valor lógico. Con este enfoque, la función pointblank has_columns() se puede utilizar para determinar si se debe activar un paso de validación sobre la base de una o más columnas existentes en la tabla (por ejemplo, ~ . %>% has_columns(vars(d, e))). El valor predeterminado de active es TRUE.

object

Un data.frame, tibble (tbl_df o tbl_dbi) o Spark DataFrame (tbl_spark) que sirve como tabla de destino para la función de expectativa o la función de prueba.

threshold

Un valor de umbral de falla simple para usar con las variantes de función expectativa (expect_) y prueba (test_). De forma predeterminada, se establece en 1, lo que significa que cualquier unidad de falla en la validación de datos da como resultado una falla general de la prueba. Los números enteros más allá de 1 indican que cualquier unidad defectuosa hasta ese valor de umbral absoluto dará como resultado una thatthat prueba o evalúe como TRUE. Asimismo, los valores fraccionarios (entre 0 y 1) actúan como un umbral de falla proporcional, donde 0.15 significa que el 15 por ciento de las unidades de prueba que fallan dan como resultado una falla general de la prueba.

Value

For the validation function, the return value is either a ptblank_agent object or a table object (depending on whether an agent object or a table was passed to x). The expectation function invisibly returns its input but, in the context of testing data, the function is called primarily for its potential side-effects (e.g., signaling failure). The test function returns a logical value.

Preconditions

Providing expressions as preconditions means pointblank will preprocess the target table during interrogation as a preparatory step. It might happen that this particular validation requires some operation on the target table before the row count comparison takes place. Using preconditions can be useful at times since since we can develop a large validation plan with a single target table and make minor adjustments to it, as needed, along the way.

The table mutation is totally isolated in scope to the validation step(s) where preconditions is used. Using dplyr code is suggested here since the statements can be translated to SQL if necessary (i.e., if the target table resides in a database). The code is most easily supplied as a one-sided R formula (using a leading ~). In the formula representation, the . serves as the input data table to be transformed. Alternatively, a function could instead be supplied.

Segments

By using the segments argument, it's possible to define a particular validation with segments (or row slices) of the target table. An optional expression or set of expressions that serve to segment the target table by column values. Each expression can be given in one of two ways: (1) as column names, or (2) as a two-sided formula where the LHS holds a column name and the RHS contains the column values to segment on.

As an example of the first type of expression that can be used, vars(a_column) will segment the target table in however many unique values are present in the column called a_column. This is great if every unique value in a particular column (like different locations, or different dates) requires it's own repeating validation.

With a formula, we can be more selective with which column values should be used for segmentation. Using a_column ~ c("group_1", "group_2") will attempt to obtain two segments where one is a slice of data where the value "group_1" exists in the column named "a_column", and, the other is a slice where "group_2" exists in the same column. Each group of rows resolved from the formula will result in a separate validation step.

Segmentation will always occur after preconditions (i.e., statements that mutate the target table), if any, are applied. With this type of one-two combo, it's possible to generate labels for segmentation using an expression for preconditions and refer to those labels in segments without having to generate a separate version of the target table.

Actions

Often, we will want to specify actions for the validation. This argument, present in every validation function, takes a specially-crafted list object that is best produced by the action_levels() function. Read that function's documentation for the lowdown on how to create reactions to above-threshold failure levels in validation. The basic gist is that you'll want at least a single threshold level (specified as either the fraction of test units failed, or, an absolute value), often using the warn_at argument. Using action_levels(warn_at = 1) or action_levels(stop_at = 1) are good choices depending on the situation (the first produces a warning, the other stop()s).

Briefs

Want to describe this validation step in some detail? Keep in mind that this is only useful if x is an agent. If that's the case, brief the agent with some text that fits. Don't worry if you don't want to do it. The autobrief protocol is kicked in when brief = NULL and a simple brief will then be automatically generated.

YAML

A pointblank agent can be written to YAML with yaml_write() and the resulting YAML can be used to regenerate an agent (with yaml_read_agent()) or interrogate the target table (via yaml_agent_interrogate()). When row_count_match() is represented in YAML (under the top-level steps key as a list member), the syntax closely follows the signature of the validation function. Here is an example of how a complex call of row_count_match() as a validation step is expressed in R code and in the corresponding YAML representation.

# R statement
agent %>% 
  row_count_match(
    tbl_compare = ~ file_tbl(
      file = from_github(
        file = "all_revenue_large.rds",
        repo = "rich-iannone/intendo",
        subdir = "data-large"
        )
      ),
    preconditions = ~ . %>% dplyr::filter(a < 10),
    segments = b ~ c("group_1", "group_2"),
    actions = action_levels(warn_at = 0.1, stop_at = 0.2),
    label = "The `row_count_match()` step.",
    active = FALSE
  )

# YAML representation
steps:
- row_count_match:
    tbl_compare: ~ file_tbl(
      file = from_github(
        file = "all_revenue_large.rds",
        repo = "rich-iannone/intendo",
        subdir = "data-large"
        )
      )
    preconditions: ~. %>% dplyr::filter(a < 10)
    segments: b ~ c("group_1", "group_2")
    actions:
      warn_fraction: 0.1
      stop_fraction: 0.2
    label: The `row_count_match()` step.
    active: false

In practice, both of these will often be shorter. Arguments with default values won't be written to YAML when using yaml_write() (though it is acceptable to include them with their default when generating the YAML by other means). It is also possible to preview the transformation of an agent to YAML without any writing to disk by using the yaml_agent_string() function.

Function ID

2-31

See also

Examples

# Create a simple table with three
# columns and four rows of values
tbl <-
  dplyr::tibble(
    a = c(5, 7, 6, 5),
    b = c(7, 1, 0, 0),
    c = c(1, 1, 1, 3)
  )

# Create a second table which is
# quite different but has the
# same number of rows as `tbl`
tbl_2 <-
  dplyr::tibble(
    e = c("a", NA, "a", "c"),
    f = c(2.6, 1.2, 0, NA)
  )

# Validate that the count of rows
# in the target table (`tbl`) matches
# that of the comparison table
# (`tbl_2`)
agent <-
  create_agent(tbl = tbl) %>%
  row_count_match(tbl_compare = tbl_2) %>%
  interrogate()

# Determine if this validation passed
# by using `all_passed()`
all_passed(agent)
#> [1] TRUE