Verify that row data are not duplicated (deprecated)

Verify that row data are not duplicated (deprecated)

rows_not_duplicated(
  x,
  columns = NULL,
  preconditions = NULL,
  brief = NULL,
  actions = NULL,
  active = TRUE
)

Arguments

x
An agent object of class ptblank_agent.

columns
The column (or a set of columns, provided as a character vector) to which this validation should be applied.

preconditions
expressions used for mutating the input table before proceeding with the validation. This is ideally as a one-sided R formula using a leading ~. In the formula representation, the . serves as the input data table to be transformed (e.g., ~ . %>% dplyr::mutate(col = col + 10).

brief
An optional, text-based description for the validation step.

actions
A list containing threshold levels so that the validation step can react accordingly when exceeding the set levels. This is to be created with the action_levels() helper function.

active
A logical value indicating whether the validation step should be active. If the step function is working with an agent, FALSE will make the validation step inactive (still reporting its presence and keeping indexes for the steps unchanged). If the step function will be operating directly on data, then any step with active = FALSE will simply pass the data through with no validation whatsoever. The default for this is TRUE.

Value

A ptblank_agent object.