Ejecutar varias pruebas y una validación final en serie

La función de validación serially() permite ejecutar una serie de pruebas en secuencia antes de culminar con un paso de validación final o simplemente salir de la serie. Esta construcción permite realizar pruebas previas que pueden tener sentido antes de un paso de validación. Por ejemplo, puede haber situaciones en las que es vital comprobar el tipo de columna antes de realizar una validación en la misma columna (ya que tener un tipo incorrecto puede dar lugar a un error de evaluación para la validación posterior). Otro flujo de trabajo en serie podría implicar tener un conjunto de comprobaciones en un orden prescrito y, si todas pasan, entonces el objetivo de estas pruebas se ha alcanzado (por ejemplo, comprobar si una tabla coincide con otra a través de una serie de pruebas cada vez más específicas).

Una serie como se especifica dentro de serially() se compone con un listado de llamadas, y nos basaríamos en las funciones de prueba (T) para describir las pruebas y opcionalmente proporcionar una llamada final con una función de validación (V). Se aplican las siguientes restricciones:

debe haber al menos una función de prueba en la serie (T -> V es buena, V es no)

sólo puede haber una llamada a la función de validación, V; es opcional pero, si se incluye, debe colocarse al final (T -> T -> V es buena, estas secuencias son malas: (1) T -> V -> T, (2) T -> T -> V -> V)

una llamada a la función de validación (V), si se incluye, no debe producir por sí misma múltiples pasos de validación (esto puede ocurrir cuando se proporcionan múltiples columns o cualquier segments)

Este es un ejemplo de cómo organizar las expresiones:

~ test_col_exists(., columns = vars(count)),
~ test_col_is_numeric(., columns = vars(count)),
~ col_vals_gt(., columns = vars(count), value = 2)

Esta serie se concentra en la columna llamada count y primero comprueba si la columna existe, luego comprueba si esa columna es numérica, y luego finalmente valida si todos los valores de la columna son mayores que 2.

Tenga en cuenta que en la lista de llamadas anterior, el . representa la tabla de destino y siempre es necesario aquí. También es importante que todas las funciones test_*() tienen un argumento threshold que se establece en 1 por defecto. Si necesita aumentar el valor del umbral, puede cambiarlo a un valor entero diferente (como umbral absoluto de unidades de prueba que fallan) o a un valor decimal entre 0 y 1 (que sirve como umbral fraccionario de unidades de prueba que fallan). unidades de prueba que fallan).

Usage

serially(
  x,
  ...,
  .list = list2(...),
  preconditions = NULL,
  actions = NULL,
  step_id = NULL,
  label = NULL,
  brief = NULL,
  active = TRUE
)

expect_serially(
  object,
  ...,
  .list = list2(...),
  preconditions = NULL,
  threshold = 1
)

test_serially(
  object,
  ...,
  .list = list2(...),
  preconditions = NULL,
  threshold = 1
)

Arguments

x

Un data.frame, tibble (tbl_df o tbl_dbi), Spark DataFrame (tbl_spark), o un agent objeto de clase ptblank_agent que se crea con create_agent().

...

Una colección de fórmulas unilaterales que consisten en llamadas a funciones test_*() (por ejemplo, test_col_vals_between(), etc.) dispuestas en la secuencia del orden de interrogación previsto. Típicamente, las validaciones hasta la final tendrían algún valor de threshold establecido (por defecto es 1) para el cortocircuito dentro de la serie. Una llamada a la función de validación final (por ejemplo, col_vals_increasing(), etc.) puede insertarse opcionalmente al final de la serie, sirviendo como un paso de validación que sólo se somete a interrogación si las pruebas anteriores pasan adecuadamente. Un ejemplo de esto es ~ test_column_exists(., vars(a)), ~ col_vals_not_null(., vars(a))).

.list

Permite el uso de una lista como alternativa de entrada a ....

preconditions

Una expresión opcional para mutar la tabla de entrada antes de continuar con la validación. Esto se puede proporcionar como una fórmula R unilateral usando un ~ inicial (por ejemplo, ~ . %>% dplyr::mutate(col = col + 10) o como una función (por ejemplo, function (x) dplyr::mutate(x, col = col + 10). Consulte la sección Preconditions para obtener más información.

actions

Una lista que contiene los niveles de umbral para que el paso de validación pueda reaccionar en consecuencia al superar los niveles establecidos. Esto se creará con la función auxiliar action_levels().

step_id

Uno o más identificadores opcionales para los pasos de validación únicos o múltiples generados al llamar a una función de validación. El uso de ID de pasos sirve para distinguir los pasos de validación entre sí y brinda la oportunidad de proporcionar una etiqueta más significativa en comparación con el índice de pasos. De forma predeterminada, es NULL, y pointblank generará automáticamente el valor de ID de paso (basado en el índice de paso) en este caso. Se pueden proporcionar uno o más valores, y el número exacto de valores de ID debe (1) coincidir con el número de pasos de validación que producirá la llamada a la función de validación (influenciado por el número de columns proporcionadas), (2) ser un ID cadena no utilizada en ningún paso de validación anterior, y (3) ser un vector con valores únicos.

label

Una etiqueta opcional para el paso de validación. Esta etiqueta aparece en el informe del agent y, para una mejor apariencia, debe ser breve.

brief

Una descripción opcional basada en texto para el paso de validación. Si no se proporciona nada aquí, el objeto agent genera un autobrief, utilizando el lenguaje proporcionado en el argumento lang de create_agent() (que por defecto es "en" o inglés). El autobrief incorpora detalles del paso de validación, por lo que a menudo es la opción preferida en la mayoría de los casos (donde un label podría ser más adecuada para describir sucintamente la validación).

active

Un valor lógico que indica si el paso de validación debe estar activo. Si la función de validación está trabajando con un objeto agent, FALSE hará que el paso de validación esté inactivo (aún informando su presencia y manteniendo los índices de los pasos sin cambios). Si la función de validación operará directamente en los datos (sin participación de agent), entonces cualquier paso con active = FALSE simplemente pasará los datos sin validación alguna. Aparte de un vector lógico, una fórmula R unilateral que usa un ~ inicial se puede usar con . (que sirve como la tabla de datos de entrada) para evaluar a un solo valor lógico. Con este enfoque, la función pointblank has_columns() se puede utilizar para determinar si se debe activar un paso de validación sobre la base de una o más columnas existentes en la tabla (por ejemplo, ~ . %>% has_columns(vars(d, e))). El valor predeterminado de active es TRUE.

object

Un data.frame, tibble (tbl_df o tbl_dbi) o Spark DataFrame (tbl_spark) que sirve como tabla de destino para la función de expectativa o la función de prueba.

threshold

Un valor de umbral de falla simple para usar con las variantes de función expectativa (expect_) y prueba (test_). De forma predeterminada, se establece en 1, lo que significa que cualquier unidad de falla en la validación de datos da como resultado una falla general de la prueba. Los números enteros más allá de 1 indican que cualquier unidad defectuosa hasta ese valor de umbral absoluto dará como resultado una thatthat prueba o evalúe como TRUE. Asimismo, los valores fraccionarios (entre 0 y 1) actúan como un umbral de falla proporcional, donde 0.15 significa que el 15 por ciento de las unidades de prueba que fallan dan como resultado una falla general de la prueba.

Value

Para la función de validación, el valor de retorno es un objeto ptblank_agent o un objeto de tabla (dependiendo de si se pasó un objeto agent o una tabla a x). La función de expectativa devuelve invisiblemente su entrada pero, en el contexto de los datos de prueba, la función se llama principalmente por sus posibles efectos secundarios (por ejemplo, falla de señalización). La función de prueba devuelve un valor lógico.

Nombres de columnas

Si se proporcionan varios nombres de columna en cualquiera de los pasos de validación suministrados, el resultado será una expansión de los pasos de subvalidación a ese número de nombres de columna. Aparte de los nombres de columnas entre comillas y en vars(), las funciones de ayuda tidyselect están disponibles para especificar columnas. Son: starts_with(), ends_with(), contains(), matches() y everything().

Preconditions

Proporcionar expresiones como preconditions significa que pointblank preprocesará la tabla de destino durante la interrogación como paso preparatorio. Puede suceder que una validación en particular requiera una columna calculada, algún filtrado de filas o la adición de columnas a través de una combinación, etc. Especialmente para un informe basado en agent, esto puede ser ventajoso ya que podemos desarrollar un gran plan de validación con una sola tabla de destino y realice ajustes menores en ella, según sea necesario, a lo largo del camino.

La mutación de la tabla está totalmente aislada en el alcance de los pasos de validación en los que se utilizan las preconditions. Aquí se sugiere usar el código dplyr ya que las declaraciones se pueden traducir a SQL si es necesario (es decir, si la tabla de destino reside en una base de datos). El código se proporciona más fácilmente como una fórmula R unilateral (utilizando un ~ inicial). En la representación de la fórmula, el . sirve como la tabla de datos de entrada que se va a transformar (por ejemplo, ~ . %>% dplyr::mutate(col_b = col_a + 10)). Alternativamente, se podría proporcionar una función (por ejemplo, function(x) dplyr::mutate(x, col_b = col_a + 10)).

Actions

A menudo, querremos especificar actions para la validación. Este argumento, presente en cada función de validación, toma un objeto de lista especialmente diseñado que es mejor producido por la función action_levels(). Lea la documentación de esa función para obtener información sobre cómo crear reacciones a niveles de falla por encima del umbral en la validación. La esencia básica es que querrá al menos un nivel de umbral único (especificado como la fracción de unidades de prueba fallidas o un valor absoluto), a menudo utilizando el argumento warn_at. Esto es especialmente cierto cuando x es un objeto de tabla porque, de lo contrario, no sucede nada. Para las funciones de tipo col_vals_*(), usar action_levels(warn_at = 0.25) o action_levels(stop_at = 0.25) son buenas opciones dependiendo de la situación (la primera produce una advertencia cuando una cuarta parte de la prueba total las unidades fallan, las otras stop()s en el mismo nivel de umbral).

Briefs

¿Quiere describir este paso de validación con algún detalle? Tenga en cuenta que esto sólo es útil si x es un objeto agent. Si ese es el caso, use un texto brief que se ajuste a la situación. No se preocupe si no quiere hacerlo. Un autobrief se activa cuando brief = NULL y el texto luego se generará automáticamente.

YAML

Se puede escribir un agente pointblank en YAML con yaml_write() y el YAML resultante se puede usar para regenerar un agente (con yaml_read_agent()) o interrogar la tabla de destino (a través de yaml_agent_interrogate()). Cuando serially() se representa en YAML (bajo la clave de nivel superior steps como un miembro de la lista), la sintaxis sigue de cerca la firma de la función de validación. A continuación se muestra un ejemplo de cómo una llamada compleja de serially() como paso de validación se expresa en código R y en la representación YAML correspondiente.

# Código R
agent %>% 
  serially(
    ~ col_vals_lt(., vars(a), 8),
    ~ col_vals_gt(., vars(c), vars(a)),
    ~ col_vals_not_null(., vars(b)),
    preconditions = ~ . %>% dplyr::filter(a < 10),
    actions = action_levels(warn_at = 0.1, stop_at = 0.2), 
    label = "El paso `serially()`.",
    active = FALSE
  )

# Representación YAML
steps:
- serially:
    fns:
    - ~col_vals_lt(., vars(a), 8)
    - ~col_vals_gt(., vars(c), vars(a))
    - ~col_vals_not_null(., vars(b))
    preconditions: ~. %>% dplyr::filter(a < 10)
    actions:
      warn_fraction: 0.1
      stop_fraction: 0.2
    label: El paso `serially()`.
    active: false

En la práctica, ambos serán a menudo más cortos, ya que solo son necesarias las expresiones para los pasos de validación. Los argumentos con valores predeterminados no se escribirán en YAML cuando se use yaml_write() (aunque es aceptable incluirlos con sus valores predeterminados al generar el YAML por otros medios). También es posible obtener una vista previa de la transformación de un agente a YAML sin escribir en el disco usando la función yaml_agent_string().

Function ID

2-32

See also

Examples

# Para todos los ejemplos aquí,
# usaremos una tabla simple con tres
# columnas numéricas (`a`, `b` y `c`);
# esta es una tabla muy básica pero
# será más útil a la hora de explicar
# las cosas más adelante
tbl <-
  dplyr::tibble(
    a = c(5, 2, 6),
    b = c(6, 4, 9),
    c = c(1, 2, 3)
  )
  
tbl
#> # A tibble: 3 × 3
#>       a     b     c
#>   <dbl> <dbl> <dbl>
#> 1     5     6     1
#> 2     2     4     2
#> 3     6     9     3
  
# A: Usando un `agent` con funciones de
#    validación y luego `interrogate()`

# La función `serially()` puede
# configurarse para realizar una serie
# de pruebas y luego realizar una
# validación (sólo si se superan todas
# las pruebas); en este caso, vamos a
# (1) comprobar si las columnas `a` y `b`
# son numéricas, (2) comprobar que ambas
# no tienen ningún valor `NA`, y
# (3) realizar una validación final que
# compruebe si los valores de `b` son
# mayores que los valores de `a`.
agent_1 <-
  create_agent(tbl = tbl) %>%
  serially(
    ~ test_col_is_numeric(., vars(a, b)),
    ~ test_col_vals_not_null(., vars(a, b)),
    ~ col_vals_gt(., vars(b), vars(a))
    ) %>%
  interrogate()
  
# Determine si esta validación no tuvo
# unidades de prueba fallidas (hay 4
# pruebas y una validación final)
all_passed(agent_1)
#> [1] TRUE

# Llamar a `agent` en la consola imprime
# el informe del agente; pero podemos
# obtener un objeto `gt_tbl` directamente
# con `get_agent_report(agent_1)`

# ¿Qué ocurre? Las cuatro pruebas han
# sido aprobadas y, por tanto, se ha
# producido la validación final;
# ¡tampoco ha habido unidades de prueba
# que hayan fallado!

# La validación final es opcional;
# aquí hay un agente diferente en el
# que sólo se realizan las pruebas en
# serie se realizan
agent_2 <-
  create_agent(tbl = tbl) %>%
  serially(
    ~ test_col_is_numeric(., vars(a, b)),
    ~ test_col_vals_not_null(., vars(a, b))
  ) %>%
  interrogate()
  
# Todo es bueno aquí también:
all_passed(agent_2)
#> [1] TRUE

# B: Usando la función de validación
#    directamente en los datos
#    (sin `agent`)

# Esta forma de utilizar las funciones
# de validación actúa como un filtro de
# datos: los datos se pasan a través,
# pero deben `stop()` si hay una sola
# unidad de prueba que falla; el
# comportamiento de los efectos
# secundarios se puede personalizar con
# la opción `actions`
tbl %>%
  serially(
    ~ test_col_is_numeric(., vars(a, b)),
    ~ test_col_vals_not_null(., vars(a, b)),
    ~ col_vals_gt(., vars(b), vars(a))
  )
#> # A tibble: 3 × 3
#>       a     b     c
#>   <dbl> <dbl> <dbl>
#> 1     5     6     1
#> 2     2     4     2
#> 3     6     9     3

# C: Usando la función de expectativa

# Con el formulario `expect_*()`,
# necesitamos ser más exactos y
# proporcionar una columna a la vez;
# esto se usa principalmente en
# pruebas `testthat`
expect_serially(
  tbl,
  ~ test_col_is_numeric(., vars(a, b)),
  ~ test_col_vals_not_null(., vars(a, b)),
  ~ col_vals_gt(., vars(b), vars(a))
)

# D: Usando la función de prueba

# Con la forma `test_*()`, deberíamos
# obtener un único valor lógico devuelto
tbl %>%
  test_serially(
    ~ test_col_is_numeric(., vars(a, b)),
    ~ test_col_vals_not_null(., vars(a, b)),
    ~ col_vals_gt(., vars(b), vars(a))
  )
#> [1] TRUE