Set a data table to an agent or informant

Setting a data table to an agent or informant with set_tbl() replaces any associated table (a data frame, a tibble, objects of class tbl_dbi or tbl_spark). If a data table is associated with an agent or informant along with a table-reading function (settable in create_agent() and create_informant()'s read_fn argument or with set_read_fn()), the table-reading function will take precedence. If this is undesirable, it be removed with the remove_read_fn() function. The association to a table can be removed with with remove_tbl().

set_tbl(x, tbl)

Arguments

x
An agent object of class ptblank_agent, or, an informant of class ptblank_informant.

tbl
The input table for the agent. This can be a data frame, a tibble, a tbl_dbi object, or a tbl_spark object. Any table already associated with the agent or informant will be overwritten.

Function ID

8-3

See also