Un fn para info_snippet(): obtener el valor más bajo de una columna

La función snip_lowest() puede utilizarse como una función info_snippet() (es decir, proporcionada a fn) para obtener el valor numérico, de tiempo o alfabético más bajo de una columna de la tabla de destino.

Uso

snip_lowest(column)

Argumentos

column

El nombre de la columna que contiene los valores de destino.

Valor

Una fórmula necesaria para el argumento fn de info_snippet().

Function ID

3-8

Ver también

Ejemplos

# Generar un objeto informante,
# añadir un snippet con
# `info_snippet()` y `snip_lowest()`
# (dándonos un método para obtener
# el valor más alto de la columna `a`);
# definir una ubicación para el
# resultado del snippet en `{ }` y
# luego `incorporate()` el snippet en
# el info text
informant <- 
  create_informant(
    read_fn = ~ small_table,
    tbl_name = "small_table",
    label = "Un ejemplo."
  ) %>% 
  info_columns(
    columns = "a",
    `Lowest Value` = "Lowest value is {lowest_a}."
  ) %>%
  info_snippet(
    snippet_name = "lowest_a",
    fn = snip_lowest(column = "a")
  ) %>%
  incorporate()
#> 
#> ── Incorporation Started - there is a single snippet to process ────────────────
#> ✔ Information gathered.
#> ✔ Snippets processed.
#> ✔ Information built.
#> 
#> ── Incorporation Completed ─────────────────────────────────────────────────────

# Podemos imprimir el objeto
# `informant` para ver el informe
# de información