Table Transformer: obtain a summary stats table for numeric columns

With any table object, you can produce a summary table that is scoped to the numeric column values. The output summary table will have a leading column called ".param." with labels for each of the nine rows, each corresponding to the following summary statistics:

Minimum ("min")

5th Percentile ("p05")

1st Quartile ("q_1")

Median ("med")

3rd Quartile ("q_3")

95th Percentile ("p95")

Maximum ("max")

Interquartile Range ("iqr")

Range ("range")

Only numerical data from the input table will generate columns in the output table. Column names from the input will be used in the output, preserving order as well.

tt_summary_stats(tbl)

Arguments

tbl

A table object to be used as input for the transformation. This can be a data frame, a tibble, a tbl_dbi object, or a tbl_spark object.

Value

A tibble object.

Function ID

12-1

See also

Examples

# Get summary statistics for the
# `game_revenue` dataset that is
# included in the package
tt_summary_stats(game_revenue)
#> # A tibble: 9 × 3
#>   .param. item_revenue session_duration
#>   <chr>          <dbl>            <dbl>
#> 1 min             0                 3.2
#> 2 p05             0.02              8.2
#> 3 q_1             0.09             18.5
#> 4 med             0.38             26.5
#> 5 q_3             1.25             33.8
#> 6 p95            22.0              39.5
#> 7 max           143.               41  
#> 8 iqr             1.16             15.3
#> 9 range         143.               37.8

# Ensure that the maximum revenue
# for individual purchases in the
# `game_revenue` table is less than
# $150
tt_summary_stats(game_revenue) %>%
  col_vals_lt(
    columns = vars(item_revenue),
    value = 150,
    segments = .param. ~ "max"
  )
#> # A tibble: 9 × 3
#>   .param. item_revenue session_duration
#>   <chr>          <dbl>            <dbl>
#> 1 min             0                 3.2
#> 2 p05             0.02              8.2
#> 3 q_1             0.09             18.5
#> 4 med             0.38             26.5
#> 5 q_3             1.25             33.8
#> 6 p95            22.0              39.5
#> 7 max           143.               41  
#> 8 iqr             1.16             15.3
#> 9 range         143.               37.8

# For in-app purchases in the
# `game_revenue` table, check that
# median revenue is somewhere
# between $8 and $12
game_revenue %>% 
  dplyr::filter(item_type == "iap") %>%
  tt_summary_stats() %>%
  col_vals_between(
    columns = vars(item_revenue),
    left = 8, right = 12,
    segments = .param. ~ "med"
  )
#> # A tibble: 9 × 3
#>   .param. item_revenue session_duration
#>   <chr>          <dbl>            <dbl>
#> 1 min             0.4              3.2 
#> 2 p05             1.39             5.99
#> 3 q_1             4.49            14.0 
#> 4 med            10.5             22.6 
#> 5 q_3            20.3             30.6 
#> 6 p95            66.0             38.8 
#> 7 max           143.              41   
#> 8 iqr            15.8             16.7 
#> 9 range         143.              37.8 

# While performing validations of the
# `game_revenue` table with an agent
# we can include the same revenue
# check by using `tt_summary_stats()`
# in the `preconditions` argument (this
# will transform the target table for
# the validation step); we also need
# to get just a segment of that table
# (the row with the median values)
agent <- 
  create_agent(
    read_fn = ~ game_revenue,
    tbl_name = "game_revenue",
    label = "An example.",
    actions = action_levels(
      warn_at = 0.10,
      stop_at = 0.25,
      notify_at = 0.35
    )
  ) %>%
  rows_complete() %>%
  rows_distinct() %>%
  col_vals_between(
    columns = vars(item_revenue),
    left = 8, right = 12,
    preconditions = ~ . %>%
      dplyr::filter(item_type == "iap") %>%
      tt_summary_stats(),
    segments = .param. ~ "med"
  ) %>%
  interrogate()

# This should all pass but let's check:
all_passed(agent)
#> [1] TRUE