Table Transformer: get the dimensions of a table

With any table object, you can produce a summary table that contains nothing more than the table's dimensions: the number of rows and the number of columns. The output summary table will have two columns and two rows. The first is the ".param." column with the labels "rows" and "columns"; the second column, "value", contains the row and column counts.

tt_tbl_dims(tbl)

Arguments

tbl
A table object to be used as input for the transformation. This can be a data frame, a tibble, a tbl_dbi object, or a tbl_spark object.

Value

A tibble object.

Function ID

12-3

See also

Examples

# Get the dimensions of the
# `game_revenue` dataset that's
# included in the package
tt_tbl_dims(game_revenue)

#> # A tibble: 2 x 2
#>   .param. value
#>   <chr>   <int>
#> 1 rows     2000
#> 2 columns    11

# This output table is useful when
# you want to validate the
# dimensions of the table; here,
# we check that `game_revenue` has
# at least 1500 rows
tt_tbl_dims(game_revenue) %>%
  dplyr::filter(.param. == "rows") %>%
  test_col_vals_gt(
    columns = vars(value),
    value = 1500
  )

#> [1] TRUE

# We can check `small_table` for
# an exact number of columns (`8`)
tt_tbl_dims(small_table) %>%
  dplyr::filter(.param. == "columns") %>%
  test_col_vals_equal(
    columns = vars(value),
    value = 8
  )

#> [1] TRUE