Table Transformer: slice a table with a slice point on a time column

With any table object containing date or date-time columns, any one of those columns can be used to effectively slice the data in two with a slice_point: and you get to choose which of those slices you want to keep.

tt_time_slice(
  tbl,
  time_column = NULL,
  slice_point = 0,
  keep = c("left", "right"),
  auto_arrange = FALSE,
  omit_missing = FALSE
)

Arguments

tbl
A table object to be used as input for the transformation. This can be a data frame, a tibble, a tbl_dbi object, or a tbl_spark object.

time_column
The time-based column that will be used as a basis for the slicing. If no time column is provided then the first one found will be used.

slice_point
The location on the time_column where the slicing will occur. This can either be a real number from 0 to 1, an ISO 8601 formatted time string (as a date or a date-time), a POSIXct time, or a Date object.

keep
Which slice should be kept? The "left" side (the default) contains data rows that are earlier than the slice_point and the "right" side will have rows that are later.

auto_arrange
Should the table be arranged by the time_column just before slicing? This may be useful if the input tbl isn't ordered by the time_column.

omit_missing
Should rows with missing values in the time_column be removed just before slicing? This might be useful in conjunction with auto_arrange = TRUE and there is an expectation of missing values in the time_column (i.e., sorting by a column with missing values will result in the accumulation of rows with those missing values).

Value

A tibble object.

Function ID

12-5

See also