Transforma un agente pointblank en un testthat archivo de prueba

Con un objeto pointblank agent, podemos escribir un archivo de prueba testthat y optar por colocarlo en testthat/tests si está disponible en la ruta del proyecto (también podemos especificar una ruta alternativa). Esto funciona transformando los pasos de validación en una serie de llamadas a expect_*() dentro de declaraciones individuales testthat::test_that().

Un requisito indispensable para utilizar write_testthat_file() en un agente es la presencia de una read_fn, que es una función que se invoca para obtener la tabla de destino. La sentencia read_fn se colocará en la parte superior del archivo de prueba testthat para que la tabla de destino esté disponible para cada una de las sentencias testthat::test_that() que le siguen. Si un objeto agent no tiene una read_fn puede añadirse a través de set_read_fn().

Los umbrales se obtendrán a partir de los aplicados para el estado stop. Esto puede configurarse para un objeto agente* en blanco pasando un objeto action_levels al argumento actions de create_agent() o el mismo argumento de cualquier función de validación incluida. Si los umbrales stop no están disponibles, se utilizará un valor de umbral de 1 para cada sentencia expect_*() generada en el archivo de prueba testthat resultante.

No es necesario que el objeto agent se someta primero a una interrogación con interrogate(). Sin embargo, puede ser útil como prueba de funcionamiento realizar interactivamente una interrogación en los datos de destino antes de generar el archivo de prueba testthat.

Uso

write_testthat_file(
  agent,
  name = NULL,
  path = NULL,
  overwrite = FALSE,
  skips = NULL,
  quiet = FALSE
)

Argumentos

agent

Un objeto de agente de clase ptblank_agent.

name

Un nombre opcional para el archivo de prueba testhat. Debe ser un nombre sin extensión y sin el texto inicial "test-". Si no se suministra nada, el nombre se derivará del tbl_name en el objeto agente. Si no está presente, se utilizará un nombre genérico.

path

Aquí se puede especificar una ruta si no se debe intentar colocar el archivo en testthat/tests.

overwrite

¿Debe sobrescribirse un archivo testthat del mismo nombre? Por defecto, esto es FALSE.

skips

Se trata de un vector opcional de palabras clave de omisión de pruebas modelado a partir de las funciones testthat skip_on_*(). Las siguientes palabras clave pueden utilizarse para incluir declaraciones skip_on_*(): "cran" (testthat::skip_on_cran()), "travis" (testthat::skip_on_travis()), "appveyor" (testthat::skip_on_appveyor()), "ci" (testthat::skip_on_ci()), "covr" (testthat::skip_on_covr()), "bioc" (testthat::skip_on_bioc()). Existen palabras clave para omitir pruebas en determinados sistemas operativos y todas ellas insertarán una llamada específica testthat::skip_on_os(). Estas son "windows" (skip_on_os("windows")), "mac" (skip_on_os("mac")), "linux" (skip_on_os("linux")) y "solaris" (skip_on_os("solaris")). Estas declaraciones se colocarán en la parte superior del archivo de prueba testthat generado.

quiet

¿La función no debe informar cuando se escribe el archivo? En por defecto es FALSE.

Valor

Devuelve invisiblemente TRUE si el archivo testthat ha sido escrito.

Detalles

Las pruebas de los pasos de validación inactivos se omitirán con un mensaje claro que indique que el motivo de la omisión se debe a que la prueba no está activa. Cualquier paso de validación inactivo puede ser forzado a un estado activo utilizando la función activate_steps() en un objeto agent (lo contrario es posible con la función deactivate_steps()).

El paquete testthat viene con una serie de funciones skip_on_*() que convenientemente hacen que el archivo de prueba se salte por completo si se cumplen ciertas condiciones. Podemos establecer rápidamente cualquier número de ellas en la parte superior del archivo de prueba testthat suministrando palabras clave como un vector a la opción skips de write_testthat_file(). Por ejemplo, si establecemos skips = c("cran", "windows) añadiremos el código skip_on_cran() y skip_on_os("windows"), lo que significa que el archivo de prueba generado no se ejecutará en un sistema CRAN o si el sistema operativo es Windows.

Este es un ejemplo de la salida del archivo de prueba testthat:

test-small_table.R

# Generated by pointblank

tbl <- small_table

test_that("column `date_time` exists", {
  
  expect_col_exists(
    tbl,
    columns = vars(date_time),
    threshold = 1
  ) 
})

test_that("values in `c` should be <= `5`", {
  
  expect_col_vals_lte(
    tbl,
    columns = vars(c),
    value = 5,
    threshold = 0.25
  ) 
})

Esto fue generado por el siguiente conjunto de declaraciones:

library(pointblank)

agent <- 
  create_agent(
    read_fn = ~ small_table,
    actions = action_levels(stop_at = 0.25)
  ) %>%
  col_exists(vars(date_time)) %>%
  col_vals_lte(vars(c), value = 5)
  
write_testthat_file(
  agent = agent,
  name = "small_table",
  path = "."
)

Function ID

8-5

Ver también

Ejemplos

if (interactive()) {

# La creación de un objeto `action_levels`
# es un paso común del flujo de trabajo al
# crear un agente pointblank; designamos
# los umbrales de fallo a los estados
# `warn`, `stop` y `notify` utilizando
# `action_levels()`
al <- 
  action_levels(
    warn_at = 0.10,
    stop_at = 0.25,
    notify_at = 0.35
  )

# A pointblank `agent` object is now
# created and the `al` object is provided
# to the agent; the static thresholds
# provided by `al` make reports a bit
# more useful after interrogation
agent <- 
  create_agent(
    read_fn = ~ small_table,
    label = "Un ejemplo.",
    actions = al
  ) %>%
  col_exists(vars(date, date_time)) %>%
  col_vals_regex(
    vars(b),
    regex = "[0-9]-[a-z]{3}-[0-9]{3}"
  ) %>%
  col_vals_gt(vars(d), value = 100) %>%
  col_vals_lte(vars(c), value = 5) %>%
  interrogate()

# This agent and all of the checks can
# be transformed into a testthat file
# with `write_testthat_file()`; the `stop`
# thresholds will be ported over
write_testthat_file(
  agent = agent,
  name = "small_table",
  path = "."
)

# The above code will generate a file with
# the name `test-small_table.R`; the path
# was specified with `"."` but, by default,
# the function will place the file in the
# `tests/testthat` folder if it's available

# An agent on disk as a YAML file can be
# made into a testthat file; the
# 'agent-small_table.yml' file is
# available in the package through
# `system.file()`
yml_file <- 
  system.file(
    "yaml", "agent-small_table.yml",
    package = "pointblank"
  )

# Writing the testthat file into the
# working directory is much the same
# as before but we're reading the
# agent from disk this time
write_testthat_file(
  agent = yaml_read_agent(yml_file),
  name = "from_agent_yaml",
  path = "."
)

}