Read a pointblank agent or informant from disk

An agent or informant that has been written to disk (with x_write_disk()) can be read back into memory with the x_read_disk() function. Once the agent or informant has been generated in this way, it may not have a data table associated with it (depending on whether the keep_tbl option was TRUE or FALSE when writing to disk) but it should still be able to produce reporting (by printing the agent or informant to the console or using get_agent_report()/get_informant_report()). An agent will return an x-list with get_agent_x_list() and yield any available data extracts with get_data_extracts(). Furthermore, all of an agent's validation steps will still be present (along with results from the last interrogation).

x_read_disk(path)

Arguments

path
The path to a file that was previously written by x_write_disk().

Details

Should the agent or informant possess a table-reading function (can be set any time with set_read_fn()) or a specific table (settable with set_tbl()) we could use the interrogate() or incorporate() function again. For a data quality reporting workflow, it is useful to interrogate() target tables that evolve over time. While the same validation steps will be used, more can be added before calling interrogate(). For an information management workflow with an informant object, using incorporate() will update aspects of the reporting such as table dimensions, and info snippets/text will be regenerated.

Function ID

8-2

See also