Establecer niveles de acción: umbrales de falla y funciones para invocar

La función action_levels() trabaja con el argumento actions que está presente en la función create_agent() y en cada función de paso de validación (que también tiene un argumento actions). Con él, podemos proporcionar niveles de umbral falla para cualquier combinación de estados de advertencia, detención o notificación.

Podemos reaccionar ante cualquier entrada de un estado proporcionando las funciones correspondientes al argumento fns. Se someterán a evaluación en el momento en que se ingrese al estado de coincidencia. Si se proporciona a create_agent(), las políticas se aplicarán a cada paso de validación, actuando de forma predeterminada para la validación en su conjunto.

Las llamadas de action_levels() también podrían aplicarse directamente a cualquier paso de validación y esto actuará como una anulación si se establece también en create_agent(). Se requiere el uso de action_levels() para tener efectos secundarios útiles (es decir, advertencias, errores de lanzamiento) en el caso de funciones de validación que operan directamente sobre los datos (p. Ej., mtcars %>% col_vals_lt("mpg", 35)). Hay dos funciones auxiliares que son convenientes cuando se usan funciones de validación directamente en datos (el flujo de trabajo sin agente): warn_on_fail() y stop_on_fail(). Estos ayudantes advierten o se detienen (el umbral de falla predeterminado para cada uno se establece en "1") y lo hacen con advertencias informativas o mensajes de error. El ayudante stop_on_fail() se aplica de forma predeterminada cuando se utilizan funciones de validación directamente en los datos (se proporciona más información sobre esto en Detalles).

action_levels(warn_at = NULL, stop_at = NULL, notify_at = NULL, fns = NULL)

warn_on_fail(warn_at = 1)

stop_on_fail(stop_at = 1)

Arguments

warn_at, stop_at, notify_at
El número de umbral o fracción de unidades de prueba que pueden proporcionar una fallar resultado antes de entrar en las warn, stop o notify estados de fallo. Si este es un valor decimal entre 0 y 1 entonces es un umbral de falla proporcional (por ejemplo, 0.15 indica que si se encuentra que el 15% por ciento de las unidades de prueba fallan, entonces se ingresa el estado de falla designado). En 1 su lugar, se pueden usar valores absolutos a partir de, y esto constituye un umbral de falla absoluto (por ejemplo, 10 significa que si se encuentra que 10 de las unidades de prueba fallan , se ingresa el estado de falla).

fns
Una lista nombrada de funciones que se emparejará con los estados de falla apropiados. La sintaxis para esta lista implica el uso de nombres de estado fracaso desde el conjunto de warn, stop y notify. Las funciones correspondientes a los estados de falla se proporcionan como fórmulas (por ejemplo list(warn = ~ warning("Too many failures."))). Se puede usar una serie de expresiones para cada estado nombrado encerrando el conjunto de declaraciones con { }.

Details

El resultado de la action_levels() llamada actionsse interpretará de forma ligeramente diferente si se usa un agente o se usan funciones de validación directamente en una tabla de datos. Por conveniencia, cuando se trabaja directamente con datos, cualquier valor suministrado warn_at o stop_at se le asignará automáticamente un stock warning() o stop() función. Por ejemplo, el uso small_table %>% col_is_integer("date") proporcionará un mensaje de detención detallado de forma predeterminada, que indica el motivo del error. Si tuviera que suministrar el fns for stop o warn manualmente, las funciones de stock se anularían. Además, si actionses NULL en este flujo de trabajo (el predeterminado), pointblank usará un stop_at valor de 1 (que proporciona un mensaje de error detallado y específico del contexto si hay unidades que fallan). Podemos suprimir absolutamente este comportamiento de parada automática en cada paso de validación mediante la configuración active = FALSE. En este caso de datos interactivos, no se proporciona una función de stock para notify_at. El notifyestado de falla se usa con menos frecuencia en este flujo de trabajo que en el basado en agente .

Cuando se utiliza un agente , a menudo opta por no utilizar ninguna función en la fns que el warn, stop y notify serán reportados en los estados de insuficiencia cuando se usa create_agent_report() (y, por lo general eso es suficiente). En cambio, usar el end_fns argumento es una mejor opción, ya que ese esquema proporciona datos útiles sobre todo el interrogatorio, lo que permite un control más preciso de los efectos secundarios y reduce la posibilidad de duplicar los efectos secundarios.

Function ID

1-5

See also

Examples

# Para estos ejemplos, usaremos el conjunto
# de datos `small_table` incluido
small_table

#> # A tibble: 13 x 8
#>    date_time           date           a b             c      d e     f    
#>    <dttm>              <date>     <int> <chr>     <dbl>  <dbl> <lgl> <chr>
#>  1 2016-01-04 11:00:00 2016-01-04     2 1-bcd-345     3  3423. TRUE  high 
#>  2 2016-01-04 00:32:00 2016-01-04     3 5-egh-163     8 10000. TRUE  low  
#>  3 2016-01-05 13:32:00 2016-01-05     6 8-kdg-938     3  2343. TRUE  high 
#>  4 2016-01-06 17:23:00 2016-01-06     2 5-jdo-903    NA  3892. FALSE mid  
#>  5 2016-01-09 12:36:00 2016-01-09     8 3-ldm-038     7   284. TRUE  low  
#>  6 2016-01-11 06:15:00 2016-01-11     4 2-dhe-923     4  3291. TRUE  mid  
#>  7 2016-01-15 18:46:00 2016-01-15     7 1-knw-093     3   843. TRUE  high 
#>  8 2016-01-17 11:27:00 2016-01-17     4 5-boe-639     2  1036. FALSE low  
#>  9 2016-01-20 04:30:00 2016-01-20     3 5-bce-642     9   838. FALSE high 
#> 10 2016-01-20 04:30:00 2016-01-20     3 5-bce-642     9   838. FALSE high 
#> 11 2016-01-26 20:07:00 2016-01-26     4 2-dmx-010     7   834. TRUE  low  
#> 12 2016-01-28 02:51:00 2016-01-28     2 7-dmx-010     8   108. FALSE low  
#> 13 2016-01-30 11:23:00 2016-01-30     1 3-dka-303    NA  2230. TRUE  high 

# Crea un objeto `action_levels` con valores
# fraccionarios para el `warn_at`,
# `stop_at` y `notify_at` estados
al <- 
  action_levels(
    warn_at = 0.2,
    stop_at = 0.8,
    notify_at = 0.5
  )
  
# Un resumen de la configuración de `al`
# El objeto se muestra imprimiéndolo
al

#> ── The `action_levels` settings ────────────────────────────────────────────────

#> WARN failure threshold of 0.2 of all test units.

#> STOP failure threshold of 0.8 of all test units.

#> NOTIFY failure threshold of 0.5 of all test units.

#> ────────────────────────────────────────────────────────────────────────────────

# Cree un agente a quemarropa y
# aplicar el objeto `al` a las `actions`;
# agregue dos pasos de validación y
# interrogar a la `small_table`
agent_1 <-
  create_agent(
    tbl = small_table,
    actions = al
  ) %>%
  col_vals_gt(
    vars(a), value = 2
  ) %>%
  col_vals_lt(
    vars(d), value = 20000
  ) %>%
  interrogate()

# El informe del agente mostrará
# que se ha entrado en el estado de advertencia
# para el primer paso de validación pero no
# el segundo; podemos confirmar esto
# en la consola inspeccionando el
# Componente `warn` en la x-list del agente
x_list <- get_agent_x_list(agent_1)
x_list$warn

#> [1]  TRUE FALSE

# Aplicar el objeto `action_levels`
# para el agente significa que toda
# validación los pasos heredarán esta
# configuración, pero podemos anular
# esto aplicando otro objeto similar a
# la validación paso en su lugar (esta
# vez usando el `warn_on_fail()`
# abreviatura)
agent_2 <-
  create_agent(
    tbl = small_table,
    actions = al
  ) %>%
  col_vals_gt(
    vars(a), value = 2,
    actions = warn_on_fail(warn_at = 0.5)
  ) %>%
  col_vals_lt(
    vars(d), value = 20000
  ) %>%
  interrogate()

# En este caso, la primera validación
# el paso tiene una falla menos estricta
# umbral para el estado de advertencia y es
# lo suficientemente alto como para que la
# condición no sea ingresó; esto se puede
# confirmar en el consola a través de la
# inspección de la componente x-list `warn`
x_list <- get_agent_x_list(agent_2)
x_list$warn

#> [1] FALSE FALSE

if (interactive()) {

# En el contexto del uso de la validación
# funciona directamente en los datos (es
# decir, no participación de un agente)
# queremos desencadenar advertencias y
# generar errores; los siguiente dará una
# advertencia si se ejecuta (devolviendo el
# datos de `small_table`)
small_table %>%
  col_vals_gt(
    vars(a), value = 2,
    actions = warn_on_fail(warn_at = 2)
  )

# Con el mismo oleoducto, no suministrando
# cualquier cosa para `actions` (es` NULL`
# por predeterminado) tendrá el mismo
# efecto que usando `stop_on_fail (stop_at = 1)`
small_table %>%
  col_vals_gt(vars(a), value = 2)

small_table %>%
  col_vals_gt(
    vars(a), value = 2,
    actions = stop_on_fail(stop_at = 1)
  )

# Esto se debe a que el `stop_on_fail()`
# la llamada se inyecta automáticamente
# de forma predeterminada caso (cuando
# se opera con datos) para su conveniencia;
# detrás de escena un agente secreto' utiliza
# 'acciones encubiertas': todo para que
# puedas escribir menos

}