Obtener una tabla de una base de datos

Si su tabla de destino está en una base de datos, la función db_tbl() es una forma práctica de acceder a ella. Esta función simplifica el proceso de obtener un objeto tbl_dbi, que generalmente implica una combinación de construir una conexión a una base de datos y usar la función dplyr::tbl() con la conexión y el nombre de la tabla (o una referencia a una tabla en un esquema). Puede usar db_tbl() como base para obtener una tabla de base de datos para el parámetro read_fn en create_agent() o create_informant(). Esto se puede hacer usando un ~ antes de la llamada db_tbl() (por ejemplo, read_fn = ~ db_tbl (...)). Otra gran opción es proporcionar una fórmula de preparación de tablas que involucre db_tbl() a tbl_store() para que tenga acceso a las tablas de la base de datos a través de nombres únicos a través de un almacén de tablas.

El nombre de usuario y la contraseña se proporcionan a través de variables de entorno. Si lo desea, estos se pueden proporcionar directamente encerrando esos valores en I().

Uso

db_tbl(
  table,
  dbname,
  dbtype,
  host = NULL,
  port = NULL,
  user = NULL,
  password = NULL
)

Argumentos

table

El nombre de la tabla, o una referencia a una tabla en un esquema (vector de dos elementos con los nombres de esquema y tabla). Alternativamente, esto se puede proporcionar como una tabla de datos para copiar en una conexión de base de datos en memoria. Esto solo funciona si: (1) el db es "sqlite" o "duckdb" , (2) el dbname fue elegido como ":memory:", y (3) el data_tbl es un marco de datos o un objeto tibble.

dbname

El nombre de la base de datos.

dbtype

Ya sea una función de controlador apropiada (por ejemplo, RPostgres::Postgres()) o un nombre corto para el tipo de base de datos. Los nombres válidos son: "postgresql", "postgres", o "pgsql" (PostgreSQL, usando la función del controlador RPostgres::Postgres()); "mysql" (MySQL, usando RMySQL::MySQL()); "duckdb" (DuckDB, usando duckdb::duckdb()); y "sqlite" (SQLite, usando RSQLite::SQLite()).

host, port

El host de la base de datos y el número de puerto opcional.

user, password

Las variables de entorno utilizadas para acceder al nombre de usuario y la contraseña de la base de datos.

Valor

Un objeto tbl_dbi.

Function ID

1-6

Ver también

Ejemplos

# Puede usar una tabla de base de
# datos en memoria y proporcionarle
# una tabla en memoria también:
small_table_duckdb <- 
  db_tbl(
    table = small_table,
    dbname = ":memory:",
    dbtype = "duckdb"
  )

if (interactive()) {

# También es posible obtener un
# archivo remoto y guardarlo en una
# base de datos en memoria; use el
# combo `file_tbl()` + `db_tbl()`
all_revenue_large_duckdb <-
  db_tbl(
    table = file_tbl(
      file = from_github(
        file = "all_revenue_large.rds",
        repo = "rich-iannone/intendo",
        subdir = "data-large"
      )
    ),
    dbname = ":memory:",
    dbtype = "duckdb"
  )

# Para las bases de datos remotas,
# es muy similar; aquí hay un ejemplo
# que accede a la tabla `rna` (en la
# base de datos pública de RNA Central)
# usando `db_tbl()`
rna_db_tbl <- 
  db_tbl(
    table = "rna",
    dbname = "pfmegrnargs",
    dbtype = "postgres", 
    host = "hh-pgsql-public.ebi.ac.uk",
    port = 5432,
    user = I("reader"),
    password = I("NWDMCE5xdipIjRrp")
  )

# El uso de `I()` para el nombre de
# usuario y la contraseña significa que
# está pasando los valores reales pero,
# normalmente, querrá usar los nombres
# de las variables de entorno (envvars)
# para acceder de forma segura a los
# valores de nombre de usuario y
# contraseña apropiados cuando se
# conecte a una base de datos:
example_db_tbl <- 
  db_tbl(
    table = "<table_name>",
    dbname = "<database_name>",
    dbtype = "<database_type_shortname>", 
    host = "<connection_url>",
    port = "<connection_port>",
    user = "<DB_USER_NAME>",
    password = "<DB_PASSWORD>"
  )

# Las variables de entorno se pueden
# crear editando el archivo de usuario
# `.Renviron` y la función
# `usethis::edit_r_environ()` hace que
# esto sea bastante fácil de hacer

# El almacenamiento de fórmulas de
# preparación de tablas en una tienda
# de tablas facilita el trabajo con
# tablas de base de datos en blanco;
# aquí se explica cómo generar un
# almacén de tablas con dos entradas
# con nombre para la preparación de
# la tabla
tbls <-
  tbl_store(
    small_table_duck ~ db_tbl(
      table = pointblank::small_table,
      dbname = ":memory:",
      dbtype = "duckdb"
    ),
    small_high_duck ~ db_tbl(
      table = pointblank::small_table,
      dbname = ":memory:",
      dbtype = "duckdb"
    ) %>%
      dplyr::filter(f == "high")
  )

# Ahora es fácil acceder a cualquiera
# de estas tablas (la segunda es una
# versión mutada) a través de la
# función `tbl_get()`
tbl_get("small_table_duck", store = tbls)
tbl_get("small_high_duck", store = tbls)

# Las fórmulas de preparación de tablas
# en `tbls` también podrían usarse en
# funciones con el argumento `read_fn`;
# esto es gracias a la función
# `tbl_source()`
agent <- 
  create_agent(
    read_fn = ~ tbl_source(
      "small_table_duck",
      store = tbls
    )
  )

informant <- 
  create_informant(
    read_fn = ~ tbl_source(
      "small_high_duck",
      store = tbls
    )
  )

}