Redacte un archivo inicial pointblank validation .R / .Rmd con una tabla de datos

Genere un borrador del plan de validación en un nuevo archivo .R o .Rmd usando una tabla de datos de entrada. Con este flujo de trabajo, se escaneará la tabla de datos para conocer los datos de su columna y se escribirá un conjunto de pasos de validación de inicio (que constituyen un plan de validación). Es mejor utilizar un extracto de datos que contenga al menos 1000 filas y que esté relativamente libre de datos falsos.

Una vez en el archivo, es posible modificar los pasos de validación para que se ajusten mejor a las expectativas del dominio en particular. Si bien la inferencia de columna se utiliza para generar planes de validación razonables, es difícil inferir los valores aceptables sin experiencia en el dominio. Sin embargo, usar draft_validation() podría ayudarlo a comenzar en el piso 10 para abordar los problemas de calidad de los datos y, en cualquier caso, es mejor que comenzar con una vista de editor de código vacía.

Uso

draft_validation(
  tbl,
  tbl_name = NULL,
  file_name = tbl_name,
  path = NULL,
  lang = NULL,
  output_type = c("R", "Rmd"),
  add_comments = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)

Argumentos

tbl

La tabla de entrada. Puede ser un marco de datos, tibble, un objeto tbl_dbi o un objeto tbl_spark.

tbl_name

Un nombre opcional para asignar al objeto de la tabla de entrada. Si no se proporciona ningún valor, se generará un nombre en función de la información disponible. Este nombre de tabla se mostrará en el área de encabezado del informe del agente generado al imprimir el agent o llamar a get_agent_report().

file_name

Un nombre opcional para el archivo .R o .Rmd. Debe ser un nombre sin extensión. Por defecto, esto se toma del tbl_name pero si no se proporciona nada para eso, el nombre contendrá el texto "draft_validation_" seguido de la fecha y hora actuales.

path

Aquí se puede especificar una ruta si no se debe intentar colocar el archivo generado en el directorio de trabajo.

lang

El idioma que se utilizará al crear comentarios para los pasos de validación generados automáticamente. De forma predeterminada, NULL creará texto en inglés ("en"). Otras opciones incluyen francés ("fr"), alemán ("de"), italiano ("it"), español ("es"), portugués ("pt"), turco ("tr"), chino ("zh"), ruso ("ru"), polaco ("pl"), danés ("da"), sueco ("sv" ) y holandés ("nl").

output_type

Una opción para elegir qué tipo de salida se debe generar. De forma predeterminada, este es un script .R ("R") pero podría ser alternativamente un documento R Markdown ("Rmd").

add_comments

¿Debería haber comentarios que expliquen las características del plan de validación en el documento generado? De forma predeterminada, es TRUE.

overwrite

¿Debería sobrescribirse un archivo con el mismo nombre? De forma predeterminada, es FALSE.

quiet

¿Debería la función no informar cuando se escribe el archivo? Por defecto, esto es FALSE.

Valor

Devuelve de forma invisible TRUE si el archivo se ha escrito.

Function ID

1-11

Ver también

Ejemplos

if (interactive()) {

# Proyecto de plan de validación
# para la tabla `dplyr::storms`
draft_validation(tbl = dplyr::storms)

}