Envíe un correo electrónico en un paso de validación o al final de un interrogatorio

La función email_blast() es útil para enviar un mensaje de correo electrónico que explica el resultado de una validación de pointblank. Se alimenta de los paquetes blastula y glue. Esta función debe ser invocada como parte del argumento end_fns de create_agent(). También es posible invocar email_blast() como parte del argumento fns de la función action_levels() (es decir, para enviar múltiples mensajes de correo electrónico en la granularidad de diferentes pasos de validación que superen los umbrales de fallo).

Para manejar mejor el envío de correos electrónicos con email_blast(), se puede utilizar la función análoga email_create() con un objeto agente pointblank o una lista x obtenida mediante la función get_agent_x_list().

Uso

email_blast(
  x,
  to,
  from,
  credentials = NULL,
  msg_subject = NULL,
  msg_header = NULL,
  msg_body = stock_msg_body(),
  msg_footer = stock_msg_footer(),
  send_condition = ~TRUE %in% x$notify
)

Argumentos

x

Una referencia al objeto x-list preparado internamente por el agente. Esta versión de la x-list es la misma que la generada a través de get_agent_x_list(<agent>) excepto que esta versión se genera internamente y por lo tanto sólo está disponible en un contexto de evaluación interna.

to, from

Las direcciones de correo electrónico de los destinatarios y del remitente.

credentials

Un objeto de lista de credenciales producido por cualquiera de las funciones blastula::creds(), blastula::creds_anonymous(), blastula::creds_key() o blastula::creds_file(). Consulte la documentación de blastula para obtener información sobre cómo utilizar estas funciones.

msg_subject

La línea de asunto del mensaje de correo electrónico.

msg_header, msg_body, msg_footer

Contenido para la cabecera, el cuerpo y el pie de página del mensaje de correo electrónico HTML.

send_condition

Una expresión que debe evaluarse como un vector lógico de longitud 1. Si se evalúa como TRUE entonces se enviará el correo electrónico, si FALSE entonces no ocurrirá. La expresión puede usar variables de la lista x (por ejemplo, x$notify, x$type, etc.) y todas esas variables pueden ser exploradas usando la función get_agent_x_list(). La expresión por defecto es ~TRUE %in% x$notify, que da como resultado TRUE si hay algún valor TRUE en el vector lógico x$notify (es decir, cualquier paso de validación da como resultado una condición condición de 'notify').

YAML

Un objeto agente pointblank puede escribirse en YAML con yaml_write() y el YAML resultante puede utilizarse para regenerar un agente (con yaml_read_agent()) o interrogar la tabla de destino (mediante yaml_agent_interrogate()). He aquí un ejemplo de cómo se expresa el uso de email_blast() dentro del argumento end_fns de create_agent() en código R y en la correspondiente representación YAML.

# Código R
create_agent(
  read_fn = ~ small_table,
  tbl_name = "small_table",
  label = "Un ejemplo.",
  actions = al,
  end_fns = list(
    ~ email_blast(
      x,
      to = "joe_public@example.com",
      from = "pb_notif@example.com",
      msg_subject = "Table Validation",
      credentials = blastula::creds_key(
        id = "smtp2go"
      ),
    )
  )
) %>%
  col_vals_gt(vars(a), 1) %>%
  col_vals_lt(vars(a), 7) 

# Representación YAML
type: agent
read_fn: ~small_table
tbl_name: small_table
label: Un ejemplo.
lang: en
locale: en
actions:
  warn_count: 1.0
  notify_count: 2.0
end_fns: ~email_blast(x, to = "joe_public@example.com", 
  from = "pb_notif@example.com", msg_subject = "Table Validation",
  credentials = blastula::creds_key(id = "smtp2go"),
  )
embed_report: true
steps:
  - col_vals_gt:
    columns: vars(a)
    value: 1.0
  - col_vals_lt:
    columns: vars(a)
    value: 7.0

Function ID

4-1

Ver también

Ejemplos

# Crear una lista `action_levels()`
# con valores absolutos para los
# estados `warn` y `notify` (con
# umbrales de 1 y 2 unidades
# de "fallo")
al <- 
  action_levels(
    warn_at = 1,
    notify_at = 2
  )
  
if (interactive()) {

# Validar que los valores de la
# columna `a` de `small_tbl` sean
# siempre > 1 y que sean siempre < 7;
# en primer lugar, aplicar la
# directiva `actions_levels()` a `actions`
# y configurar un `email_blast()` como
# uno de los `end_fns` (por defecto, el
# correo electrónico se enviará si hay
# un único estado `notify` en todos
# los pasos de validación)
agent <-
  create_agent(
    read_fn = ~ small_table,
    tbl_name = "small_table",
    label = "Un ejemplo.",
    actions = al,
    end_fns = list(
      ~ email_blast(
        x,
        to = "joe_public@example.com",
        from = "pb_notif@example.com",
        msg_subject = "Table Validation",
        credentials = blastula::creds_key(
          id = "smtp2go"
        ),
      )
    )
  ) %>%
  col_vals_gt(vars(a), value = 1) %>%
  col_vals_lt(vars(a), value = 7) %>%
  interrogate()

}

# El ejemplo anterior no se ha ejecutado
# intencionadamente porque las
# credenciales de correo electrónico no
# están disponibles y las direcciones de
# correo electrónico `to` y `from`
# son inexistentes

# Para obtener un objeto de correo
# electrónico de blastula en lugar de
# enviar el mensaje de forma ansiosa,
# podemos utilizar la función función
# `email_create()`
email_object <-
  create_agent(
    read_fn = ~ small_table,
    tbl_name = "small_table",
    label = "Un ejemplo.",
    actions = al
  ) %>%
  col_vals_gt(vars(a), value = 5) %>%
  col_vals_lt(vars(a), value = 7) %>%
  interrogate() %>%
  email_create()