Cree un objeto de correo electrónico a partir de un pointblank agent o informant

La función email_create() produce un objeto de mensaje de correo electrónico que puede ser enviar utilizando el paquete blastula. La "x" que necesitamos para esto puede ser ser un agente pointblank, la lista x del agente (producida desde el agente con la función get_agent_x_list()), o un pointblank informante. En todos los casos, el mensaje de correo electrónico aparecerá en el Visor y una blastula email_message object será devuelto.

Uso

email_create(
  x,
  msg_header = NULL,
  msg_body = stock_msg_body(),
  msg_footer = stock_msg_footer()
)

Argumentos

x

A pointblank agent, an agent x-list, or a pointblank informant. The x-list object can be created with the get_agent_x_list() function. It is recommended that the option i = NULL be used with get_agent_x_list() if supplying an x-list as x. Furthermore, The option generate_report = TRUE could be used with create_agent() so that the agent report is available within the email.

msg_header, msg_body, msg_footer

Content for the header, body, and footer components of the HTML email message.

Valor

A blastulaemail_message object.

Function ID

4-2

Ver también

Ejemplos

if (interactive()) {

# Crear una lista `action_levels()`
# con valores absolutos para los
# estados `warn`, y `notify` (con
# umbrales de 1 y 2 unidades
# de `fallo`)
al <- 
  action_levels(
    warn_at = 1,
    notify_at = 2
  )

# En un flujo de trabajo que
# implique un objeto `agente`,
# podemos hacer uso del argumento
# `end_fns` y enviar el informe por
# correo electrónico de forma
# programática con la función
# `email_blast()`, sin embargo, un
# flujo de trabajo alternativo es
# producir el objeto de correo
# electrónico y elegir enviarlo
# fuera de la API de pointblank; la
# función `email_create()` nos
# permite hacer esto con un
# objeto `agente`
email_object_1 <-
  create_agent(
    read_fn = ~ small_table,
    tbl_name = "small_table",
    label = "Un ejemplo.",
    actions = al
  ) %>%
  col_vals_gt(vars(a), value = 1) %>%
  col_vals_lt(vars(a), value = 7) %>%
  interrogate() %>%
  email_create()

# Podemos ver el correo HTML
# simplemente imprimiendo
# `email_object`; debería debería
# aparecer en el Visor

# La función `email_create()` también
# puede utilizarse en una x-list de
# agentes para obtener el mismo
# objeto de mensaje de correo
# electrónico
email_object_2 <-
  create_agent(
    read_fn = ~ small_table,
    tbl_name = "small_table",
    label = "Un ejemplo.",
    actions = al
  ) %>%
  col_vals_gt(vars(a), value = 5) %>%
  col_vals_lt(vars(b), value = 5) %>%
  interrogate() %>%
  get_agent_x_list() %>%
  email_create()

# Un informe producido por el
# informante puede convertirse en
# un objeto de mensaje de correo
# electrónico; creemos un informante
# y utilicemos `email_create()`
email_object_3 <-
  create_informant(
    read_fn = ~ small_table,
    tbl_name = "small_table",
    label = "Un ejemplo."
  ) %>%
  info_tabular(
    info = "A simple table in the
    *Examples* section of the function
    called `email_create()`."
  ) %>%
  info_columns(
    columns = vars(a),
    info = "Numbers. On the high side."
  ) %>%
  info_columns(
    columns = vars(b),
    info = "Lower numbers. Zeroes, even."
  ) %>%
  incorporate() %>%
  email_create()

}