Especificar un archivo para descargar desde GitHub

La función from_github() es útil para generar una URL válida que apunte a un archivo de datos en un repositorio público de GitHub. Esta función se puede utilizar en el argumento file de la función file_tbl() o en cualquier otro lugar donde se necesiten URLs de GitHub para el contenido bruto del usuario.

Uso

from_github(file, repo, subdir = NULL, default_branch = "main")

Argumentos

file

El nombre del archivo al que se apunta en un repositorio de GitHub. Puede ser una ruta que lleve al archivo y lo incluya. Esto se combina con cualquier ruta dada en subdir.

repo

La dirección del repositorio de GitHub en el formato username/repo[/subdir][@ref|#pull|@*release].

subdir

Una cadena de ruta que representa un subdirectorio en el repositorio de GitHub de GitHub. Se combina con cualquier componente de la ruta incluido en file.

Valor

Un vector de caracteres de longitud 1 que contiene una URL.

Function ID

13-6

Ver también

Ejemplos

# Una URL válida a un archivo de datos
# en GitHub puede ser obtenida en el
# HEAD de la rama por defecto
# from_github(
#   file = "inst/data_files/small_table.csv",
#   repo = "rich-iannone/pointblank"
# )

# La ruta de acceso a la ubicación del
# archivo puede ser suministrada total
# o parcialmente a `subdir`.
# from_github(
#   file = "small_table.csv",
#   repo = "rich-iannone/pointblank",
#   subdir = "inst/data_files"
# )

# Podemos utilizar la primera llamada
# en combinación con `file_tbl()` y
# `create_agent()`; esto proporciona
# una fórmula de preparación de tablas
# que obtiene un archivo CSV del
# repositorio de GitHub para el
# paquete pointblank 
# agent <- 
#   create_agent(
#     read_fn = ~ file_tbl(
#       file = from_github(
#         file = "inst/data_files/small_table.csv",
#         repo = "rich-iannone/pointblank"
#       ),
#       col_types = "TDdcddlc"
#     )
#   ) %>%
#   col_vals_gt(vars(a), 0) %>%
#   interrogate()

# La función de ayuda `from_github()`
# es bastante potente y puede llegar
# a muchos diferentes archivos en
# un repositorio

# Un archivo de datos de GitHub puede
# obtenerse de un commit en el
# momento de la publicación
# from_github(
#   file = "inst/extdata/small_table.csv",
#   repo = "rich-iannone/pointblank@v0.2.1"
# )

# También se puede obtener un archivo
# de un repo en el momento de un commit
# específico (se puede utilizar el hash
# parcial o hash SHA-1 completo de
# la confirmación)
# from_github(
#   file = "data-raw/small_table.csv",
#   repo = "rich-iannone/pointblank@e04a71"
# )

# Un archivo también puede obtenerse
# de un *abierto* pull request
# from_github(
#   file = "data-raw/small_table.csv",
#   repo = "rich-iannone/pointblank#248"
# )