Una tabla con datos de ingresos del juego

Esta tabla es un subconjunto de la tabla sj_all_revenue del paquete de datos intendo. Son las primeras 2000 filas de esa tabla donde los registros de ingresos van desde 2015-01-01 hasta 2015-01-21.

Uso

game_revenue

Formato

Un tibble con 2.000 filas y 11 variables:

player_id

Una columna de character con identificadores únicos para cada usuario/jugador.

session_id

Una columna de character que contiene identificadores únicos para cada sesión de jugador.

session_start

Una columna de fecha-hora que indica cuándo la sesión (que contiene el evento de ingresos) comenzó.

time

Una columna de fecha-hora que indica exactamente cuándo el jugador (o evento de ingresos).

item_type

Una columna de character que proporciona la clase del artículo comprado.

item_name

Una columna de character que proporciona el nombre del artículo comprado.

item_revenue

Una columna numeric con los importes de los ingresos por artículo comprado.

session_duration

Una columna numeric que indica la duración de la sesión (en minutos) en la que se produjo la compra.

start_day

Una columna Date que proporciona la fecha del primer ingreso de el jugador que realiza la compra.

acquisition

Una columna de character que proporciona el método de adquisición para el jugador.

country

Una columna de character que proporciona el probable país de residencia del jugador.

Function ID

14-4

Ver también

Ejemplos

# Aquí hay un vistazo a los datos
# disponibles en `game_revenue`
dplyr::glimpse(game_revenue)
#> Rows: 2,000
#> Columns: 11
#> $ player_id        <chr> "ECPANOIXLZHF896", "ECPANOIXLZHF896", "ECPANOIXLZHF89…
#> $ session_id       <chr> "ECPANOIXLZHF896-eol2j8bs", "ECPANOIXLZHF896-eol2j8bs…
#> $ session_start    <dttm> 2015-01-01 01:31:03, 2015-01-01 01:31:03, 2015-01-01…
#> $ time             <dttm> 2015-01-01 01:31:27, 2015-01-01 01:36:57, 2015-01-01…
#> $ item_type        <chr> "iap", "iap", "iap", "ad", "ad", "ad", "ad", "ad", "a…
#> $ item_name        <chr> "offer2", "gems3", "gold7", "ad_20sec", "ad_5sec", "a…
#> $ item_revenue     <dbl> 8.991, 22.491, 107.991, 0.760, 0.030, 0.070, 0.080, 1…
#> $ session_duration <dbl> 16.3, 16.3, 16.3, 16.3, 35.2, 35.2, 35.2, 35.2, 35.2,…
#> $ start_day        <date> 2015-01-01, 2015-01-01, 2015-01-01, 2015-01-01, 2015…
#> $ acquisition      <chr> "google", "google", "google", "google", "google", "go…
#> $ country          <chr> "Germany", "Germany", "Germany", "Germany", "Germany"…