Obtenga la x-list del agente

La x-list del agente es un registro de la información que posee el agente en un momento dado. La x-list contendrá la información más completa después de que haya tenido lugar una interrogación (antes de eso, los datos reflejan en gran medida el plan de validación). La x-list puede limitarse a un paso de validación concreto (proporcionando el número de paso al argumento i), o bien, podemos obtener la información de todos los pasos de validación dejando i sin especificar. La x-list es un objeto R list que contiene mucha información.

Para un objeto x-list obtenido con i especificado para un paso de validación, están disponibles los siguientes componentes:

time_start: la hora a la que comenzó el interrogatorio (POSIXct [0 o 1])

time_end: la hora a la que terminó el interrogatorio (POSIXct [0 o 1])

label: la etiqueta opcional dada al agente (chr [1])

tbl_name: el nombre del objeto de la tabla, si está disponible (chr [1])

tbl_src: el tipo de tabla utilizada en la validación (chr [1])

tbl_src_details: si la tabla es una tabla de base de datos, esto proporciona más detalles sobre la tabla de la base de datos (chr [1])

tbl: el propio objeto de la tabla

col_names: los nombres de las columnas de la tabla (chr [ncol(tbl)])

col_types: los tipos de columna de la tabla (chr [ncol(tbl)])

i: el índice de pasos de validación (int [1])

type: el tipo de validación, el valor es el nombre de la función de validación (chr [1])

columns: las columnas especificadas para la función de validación (chr [variable length])

values: los valores especificados para la función de validación (mixed types [variable length])

briefs: el resumen del paso de validación en el lang especificado (chr [1])

eval_error, eval_warning: indica si la evaluación de la función de paso función de paso, durante la interrogación, dio lugar a un error o a una advertencia (lgl [1])

capture_stack: una lista de los errores o avisos capturados durante la evaluación de la función en el momento de la interrogación (list [1])

n: el número de unidades de prueba para la etapa de validación (num [1])

n_passed, n_failed: el número de unidades de prueba que pasan y fallan para el paso de validación (num [1])

f_passed: la fracción de unidades de prueba aprobadas para el paso de validación, n_passed / n (num [1])

f_failed: la fracción de unidades de prueba que fallan en el paso de validación, n_failed / n (num [1])

warn, stop, notify: un valor lógico que indica si el nivel de unidades de prueba fallidas hizo que se introdujeran las condiciones correspondientes (lgl [1])

lang: el código de idioma de dos letras que indica qué idioma debe utilizarse para todos los escritos, el informe del agente y el informe generados por la función scan_data() (chr [1])

Si i no se especifica (es decir, no se restringe a un paso de validación específico) entonces ciertos componentes de longitud uno en el objeto x-list se expandirán al número total de pasos de validación (estos son: i, type, columns, values, briefs, eval_error, eval_warning, capture_stack, n, n_passed, n_failed, f_passed, f_failed, warn, stop y notify). El objeto x-list también tendrá componentes adicionales cuando i sea NULL, que son:

report_object: un objeto de tabla gt, que también se presenta como el método de impresión por defecto para un ptblank_agent

email_object: un objeto blastula email_message con un conjunto de componentes

report_html: la fuente HTML para el report_object, proporcionada como un vector de un carácter de longitud

report_html_small: la fuente HTML para una versión más reducida y condensada de report_object, proporcionada como un vector de un carácter de longitud; El HTML tiene estilos inline, lo que lo hace más adecuado para los cuerpos de los mensajes de correo electrónico

Uso

get_agent_x_list(agent, i = NULL)

Argumentos

agent

Un objeto de agente de clase ptblank_agent.

i

El número de paso de validación, que se asigna a cada paso de validación en el orden de invocación. Si es NULL (el valor por defecto), la x-list proporcionará información para todos los pasos de validación. Si se proporciona un número de paso válido, la x-list tendrá información perteneciente sólo a ese paso.

Valor

Un objeto list.

Function ID

8-1

Ver también

Ejemplos

# Crear un simple data frame con 
# una columna de valores numéricos
tbl <- dplyr::tibble(a = c(5, 7, 8, 5))

# Crear un objeto `action_levels()`
# con valores fraccionados para los
# `warn`, `stop` y `notify`
al <-
  action_levels(
    warn_at = 0.2,
    stop_at = 0.8,
    notify_at = 0.345
  )

# Crear un agente (dándole los
# objetos `tbl` y `al`), utilizar
# dos funciones de validación,
# y luego interrogar
agent <-
  create_agent(
    tbl = tbl,
    actions = al
  ) %>%
  col_vals_gt(vars(a), value = 7) %>%
  col_is_numeric(vars(a)) %>%
  interrogate()
  
# Obtener el objeto x-list
x <- get_agent_x_list(agent)

# Imprimir el objeto x-list
x
#> ── The x-list for `tbl` ─────────────────────────────────────────── ALL STEPS ──
#> $time_start $time_end (POSIXct [1])
#> $label $tbl_name $tbl_src $tbl_src_details (chr [1])
#> $tbl (tbl_df, tbl, and data.frame)
#> $col_names $col_types (chr [1])
#> $i $type $columns $values $label $briefs (mixed [2])
#> $eval_error $eval_warning (lgl [2])
#> $capture_stack (list [2])
#> $n $n_passed $n_failed $f_passed $f_failed (num [2])
#> $warn $stop $notify (lgl [2])
#> $validation_set (tbl_df [2, 35])
#> $lang (chr [1])
#> $report_object (gt_tbl)
#> $email_object (blastula_message)
#> $report_html $report_html_small (chr [1])
#> ────────────────────────────────────────────────────────────────────────────────

# Obtener el componente `f_passed`
# del objeto x-list
x$f_passed
#> [1] 0.25 1.00