Recopile extractos de datos de un paso de validación

En un flujo de trabajo basado en agentes (es decir, iniciado con create_agent()), después de la interrogación con interrogate(), podemos extraer los datos de las filas que no pasaron los pasos de validación basados en filas con la función get_data_extracts(). Hay un extracto discreto por cada paso de validación basado en filas y la cantidad de datos disponibles en un extracto particular depende tanto de la fracción de unidades de prueba que no pasaron el paso de validación como del nivel de muestreo o recolección explícita de ese conjunto de unidades. Estos extractos pueden recogerse mediante programación a través de get_data_extracts() pero también pueden descargarse como archivos CSV desde el informe HTML generado por el método print o mediante el uso de get_agent_report().

La disponibilidad de extractos de datos para cada paso de validación basado en filas depende de si extract_failed se establece como TRUE dentro de la llamada interrogate() (lo es por defecto). La cantidad de filas failed extraídas depende de los parámetros de recogida en interrogate(), y el comportamiento por defecto es recoger hasta las primeras 5000 filas fail.

Los pasos de validación por filas se basan en aquellas funciones de validación de la forma col_vals_*() y también incluyen conjointly() y rows_distinct(). Sólo las funciones de ese conjunto de funciones de validación pueden producir datos extraídos.

Uso

get_data_extracts(agent, i = NULL)

Argumentos

agent

Un objeto de agente de clase ptblank_agent. Se debería haber llamado a interrogate() sobre él, de manera que los pasos de validación se llevaran a cabo y cualquier fila de muestra de las validaciones que no pasaran pudiera estar disponibles en el objeto.

i

El número de paso de validación, que es asignado a cada paso de validación por pointblank en el orden de definición. Si es NULL (el valor por defecto), todas las tablas de extracto de datos se proporcionarán en un objeto de lista.

Valor

Una lista de tablas si no se proporciona i, o, una tabla independiente si se proporciona i.

Function ID

8-2

Ver también

Ejemplos

# Crea una serie de dos pasos de
# validación centrados en la
# comprobación de los valores de
# las filas de parte del objeto
# `small_table`; `interrogate()`
# inmediatamente
agent <-
  create_agent(
    read_fn = ~ small_table %>%
      dplyr::select(a:f),
    label = "`get_data_extracts()`"
  ) %>%
  col_vals_gt(vars(d), value = 1000) %>%
  col_vals_between(
    vars(c),
    left = vars(a), right = vars(d),
    na_pass = TRUE
  ) %>%
  interrogate()

# El uso de `get_data_extracts()`
# con sus valores predeterminados
# devuelve una lista de tablas, donde
# cada tabla se denomina según el
# paso de validación que tiene un
# extracto disponible
agent %>% get_data_extracts()
#> $`1`
#> # A tibble: 6 × 6
#>       a b             c     d e     f    
#>   <int> <chr>     <dbl> <dbl> <lgl> <chr>
#> 1     8 3-ldm-038     7  284. TRUE  low  
#> 2     7 1-knw-093     3  843. TRUE  high 
#> 3     3 5-bce-642     9  838. FALSE high 
#> 4     3 5-bce-642     9  838. FALSE high 
#> 5     4 2-dmx-010     7  834. TRUE  low  
#> 6     2 7-dmx-010     8  108. FALSE low  
#> 
#> $`2`
#> # A tibble: 4 × 6
#>       a b             c     d e     f    
#>   <int> <chr>     <dbl> <dbl> <lgl> <chr>
#> 1     6 8-kdg-938     3 2343. TRUE  high 
#> 2     8 3-ldm-038     7  284. TRUE  low  
#> 3     7 1-knw-093     3  843. TRUE  high 
#> 4     4 5-boe-639     2 1036. FALSE low  
#> 

# Podemos obtener un extracto de
# un paso específico especificándolo
# en el argumento `I`; obtengamos las
# filas que fallan del primer paso
# de validación (`col_vals_gt`)
agent %>% get_data_extracts(i = 1)
#> # A tibble: 6 × 6
#>       a b             c     d e     f    
#>   <int> <chr>     <dbl> <dbl> <lgl> <chr>
#> 1     8 3-ldm-038     7  284. TRUE  low  
#> 2     7 1-knw-093     3  843. TRUE  high 
#> 3     3 5-bce-642     9  838. FALSE high 
#> 4     3 5-bce-642     9  838. FALSE high 
#> 5     4 2-dmx-010     7  834. TRUE  low  
#> 6     2 7-dmx-010     8  108. FALSE low