Obtener un informe de información de la tabla de un objeto informant

We can get a table information report from an informant object that's generated by the create_informant() function. The report is provided as a gt based display table. The amount of information shown depends on the extent of that added via the use of the info_*() functions or through direct editing of a pointblank YAML file (an informant can be written to pointblank YAML with yaml_write(informant = <informant>, ...)).

Uso

get_informant_report(
  informant,
  size = "standard",
  title = ":default:",
  lang = NULL,
  locale = NULL
)

Argumentos

informant

An informant object of class ptblank_informant.

size

The size of the display table, which can be either "standard" (the default, with a width of 875px) or "small" (width of 575px).

title

Options for customizing the title of the report. The default is the keyword ":default:" which produces generic title text that refers to the pointblank package in the language governed by the lang option. Another keyword option is ":tbl_name:", and that presents the name of the table as the title for the report. If no title is wanted, then the ":none:" keyword option can be used. Aside from keyword options, text can be provided for the title and glue::glue() calls can be used to construct the text string. If providing text, it will be interpreted as Markdown text and transformed internally to HTML. To circumvent such a transformation, use text in I() to explicitly state that the supplied text should not be transformed.

lang

The language to use for the information report (a summary table that provides the validation plan and the results from the interrogation. De forma predeterminada, NULL creará texto en inglés ("en"). Otras opciones incluyen francés ("fr"), alemán ("de"), italiano ("it"), español ("es"), portugués ("pt"), turco ("tr"), chino ("zh"), ruso ("ru"), polaco ("pl"), danés ("da"), sueco ("sv" ) y holandés ("nl"). This lang option will override any previously set language setting (e.g., by the create_agent() call).

locale

An optional locale ID to use for formatting values in the information report summary table according the locale's rules. Examples include "en_US" for English (United States) and "fr_FR" for French (France); more simply, this can be a language identifier without a country designation, like "es" for Spanish (Spain, same as "es_ES"). This locale option will override any previously set locale value (e.g., by the create_agent() call).

Valor

A gt table object.

Function ID

7-2

Ver también

Ejemplos

# Generate an informant object using
# the `small_table` dataset
informant <- create_informant(small_table)

# This function creates some information
# without any extra help by profiling
# the supplied table object; it adds
# the sections 'table' and columns' and
# we can print the object to see the
# table information report

# Alternatively, we can get the same report
# by using `get_informant_report()`
report <- get_informant_report(informant)
class(report)
#> [1] "ptblank_informant_report" "gt_tbl"                  
#> [3] "list"