Get a table metadata report from a metadata object

We can get a table metadata report from a metadata object that's generated by the create_metadata() function. The report is provided as a gt based display table. The amount of metadata shown depends on the extent of that added via the use of the meta_*() functions or addition of metadata through direct editing of a pointblank YAML file.

get_metadata_report(metadata)

Arguments

metadata
A metadata object of class ptblank_metadata.

Value

A gt table object.

Examples

# Generate a metadata object using
# the `small_table` dataset
metadata <- create_metadata(small_table)

# This function creates some metadata
# without any extra help by profiling
# the supplied table object; it adds
# the sections: (1) 'table' and
# (2) 'columns' and we can print the
# object to see the metadata report
# metadata

# Alternatively, we can visualize this
# metadata into the same report by using
# `get_metadata_report()`
report <- get_metadata_report(metadata)
class(report)

#> [1] "gt_tbl" "list"