Obtenga un informe resumido utilizando varios agentes

Podemos obtener una tabla resumen informativa de un colectivo de agentes utilizando la función get_multiagent_report(). La información de un multiagent puede proporcionarse de tres formas muy distintas: (1) el Long Display (informes apilados), (2) el Wide Display (un informe comparativo), (3) como un tibble con columnas empaquetadas.

Uso

get_multiagent_report(
  multiagent,
  display_table = TRUE,
  display_mode = c("long", "wide"),
  title = ":default:",
  lang = NULL,
  locale = NULL
)

Argumentos

multiagent

Un objeto multiagente de la clase ptblank_multiagent.

display_table

¿Debe generarse una tabla de visualización? Si es TRUE (por defecto) se mostrará una tabla de visualización del informe en el Visor. Si FALSE entonces se devolverá un tibble.

display_mode

Si obtenemos una tabla de visualización, ¿los datos del agente deben presentarse de forma "long" o "wide"? El valor por defecto es "long" pero cuando se comparan varias ejecuciones en las que la tabla de destino es la misma puede ser preferible elegir "wide".

title

Opciones para personalizar el título del informe cuando display_table = TRUE. El valor por defecto es la palabra clave ":default:" que produce un texto de título genérico. Si no se desea ningún título, se puede utilizar la opción de palabra clave ":none:". Otra opción de palabra clave es ":tbl_name:", que presenta el nombre de la tabla como título del informe (sólo puede utilizarse cuando display_mode = "long"). Aparte de las opciones de palabras clave, se puede proporcionar texto para el título y se pueden utilizar llamadas a glue::glue() para construir la cadena de texto. Si se proporciona texto, se interpretará como texto Markdown y se transformará internamente en HTML. Para evitar dicha transformación, utilice text en I() para indicar explícitamente que el suministrado no debe ser transformado.

lang

La lengua a utilizar para los formularios de informes largos o amplios. De forma predeterminada, NULL creará texto en inglés ("en"). Otras opciones incluyen francés ("fr"), alemán ("de"), italiano ("it"), español ("es"), portugués ("pt"), turco ("tr"), chino ("zh"), ruso ("ru"), polaco ("pl"), danés ("da"), sueco ("sv" ) y holandés ("nl").

locale

Un identificador opcional de la configuración regional que se utilizará para formatear los valores en los formularios largos o en los formularios de informes anchos (según las reglas de la configuración regional). Los ejemplos incluyen "en_US" para el inglés (Estados Unidos) y "fr_FR" para el francés (Francia); más simplemente, puede ser un identificador de idioma sin designación de país, como "es" para el español (España, igual que "es_ES"). Esta opción locale anulará cualquier valor de locale previamente establecido.

Valor

Un objeto de tabla gt si display_table = TRUE o un tibble si display_table = FALSE.

La pantalla larga

Cuando se muestra como "long" el informe multiagente apilará los informes de los agentes individuales en un único documento en el orden de los agentes en el objeto multiagente.

Se proporcionará cada plan de validación (posiblemente con información de interrogación) y la salida de cada uno es equivalente a llamar a get_agent_report() en cada de los agentes dentro del objeto multiagente.

La pantalla ancha

Cuando se muestra como "wide" el informe multiagente mostrará los datos de los agentes individuales como columnas, con filas que representan pasos de validación comunes comunes a todos los agentes.

Cada paso de validación se representa con un icono (que representa el nombre de la función de validación) y el hash SHA1 asociado. Esta es una forma muy fiable de saber qué pasos de validación son efectivamente idénticos en todas las interrogaciones. Esta forma de organizar el informe es beneficiosa porque diferentes agentes pueden haber utilizado diferentes pasos y queremos hacer un seguimiento de los resultados de la validación cuando el paso de validación no cambia pero la tabla de destino sí (es decir, se añaden nuevas filas, se actualizan las filas existentes, etc.).

La tabla única de este modo de visualización tendrá las siguientes columnas:

STEP: el hash SHA1 para el paso de validación, posiblemente compartido entre varias interrogaciones.

todas las demás columnas: cada columna más allá de STEP representa una interrogación separada de un objeto agent. La marca de tiempo para la finalización de cada interrogación se muestra como la etiqueta de la columna.

Function ID

10-3

Ver también

Ejemplos

if (interactive()) {

# Let's walk through several theoretical
# data quality analyses of an extremely
# small table; that table is called
# `small_table` and we can find it as a
# dataset in this package
small_table

# To set failure limits and signal
# conditions, we designate proportional
# failure thresholds to the `warn`, `stop`,
# and `notify` states using `action_levels()`
al <- 
  action_levels(
    warn_at = 0.05,
    stop_at = 0.10,
    notify_at = 0.20
  )

# We will create four different agents
# and have slightly different validation
# steps in each of them; in the first,
# `agent_1`, eight different validation
# steps are created and the agent will
# interrogate the `small_table`
agent_1 <-
  create_agent(
    read_fn = ~ small_table,
    tbl_name = "small_table",
    label = "`get_multiagent_report()`",
    actions = al
  ) %>%
  col_vals_gt(
    vars(date_time),
    value = vars(date),
    na_pass = TRUE
  ) %>%
  col_vals_gt(
    vars(b), 
    value = vars(g),
    na_pass = TRUE
  ) %>%
  rows_distinct() %>%
  col_vals_equal(
    vars(d), 
    value = vars(d),
    na_pass = TRUE
  ) %>%
  col_vals_between(
    vars(c), 
    left = vars(a), right = vars(d)
  ) %>%
  col_vals_not_between(
    vars(c),
    left = 10, right = 20,
    na_pass = TRUE
  ) %>%
  rows_distinct(vars(d, e, f)) %>%
  col_is_integer(vars(a)) %>%
  interrogate()

# The second agent, `agent_2`, retains
# all of the steps of `agent_1` and adds
# two more (the last of which is inactive)
agent_2 <- 
  agent_1 %>%
  col_exists(vars(date, date_time)) %>%
  col_vals_regex(
    vars(b), 
    regex = "[0-9]-[a-z]{3}-[0-9]{3}",
    active = FALSE
  ) %>%
  interrogate()

# The third agent, `agent_3`, adds a single
# validation step, removes the fifth one,
# and deactivates the first
agent_3 <- 
  agent_2 %>%
  col_vals_in_set(
    vars(f),
    set = c("low", "mid", "high")
  ) %>%
  remove_steps(i = 5) %>%
  deactivate_steps(i = 1) %>%
  interrogate()

# The fourth and final agent, `agent_4`,
# reactivates steps 1 and 10, and removes
# the sixth step
agent_4 <-
  agent_3 %>%
  activate_steps(i = 1) %>%
  activate_steps(i = 10) %>%
  remove_steps(i = 6) %>%
  interrogate()

# While all the agents are slightly
# different from each other, we can still
# get a combined report of them by
# creating a 'multiagent'
multiagent <-
  create_multiagent(
    agent_1, agent_2, agent_3, agent_4
  )

# Calling `multiagent` in the console
# prints the multiagent report; but we
# can use some non-default option with
# the `get_multiagent_report()` function

# By default, `get_multiagent_report()`
# gives you a tall report with agent
# reports being stacked
report_1 <- 
  get_multiagent_report(multiagent)
  
# We can modify the title with that's
# more suitable or use a keyword like
# `:tbl_name:` to give us the target
# table name in each section
report_2 <- 
  get_multiagent_report(
    multiagent,
    title = ":tbl_name:"
  )

# We can opt for a wide display of
# the reporting info, and this is
# great when reporting on multiple
# validations of the same target
# table
report_3 <- 
  get_multiagent_report(
    multiagent,
    display_mode = "wide"
  )
}