Get a parameter value from a summary table

The get_tt_param() function can help you to obtain a single parameter value from a summary table generated by the tt_*() functions tt_summary_stats(), tt_string_info(), tt_tbl_dims(), or tt_tbl_colnames(). The following parameters are to be used depending on the input tbl:

from tt_summary_stats(): "min", "p05", "q_1", "med", "q_3", "p95", "max", "iqr", "range"

from tt_string_info(): "length_mean", "length_min", "length_max"

from tt_tbl_dims(): "rows", "columns"

from tt_tbl_colnames(): any integer present in the .param. column

The tt_summary_stats() and tt_string_info() functions will generate summary tables with columns that mirror the numeric and character columns in their input tables, respectively. For that reason, a column name must be supplied to the column argument in get_tt_param().

Usage

get_tt_param(tbl, param, column = NULL)

Arguments

tbl

A summary table generated by either of the tt_summary_stats(), tt_string_info(), tt_tbl_dims(), or tt_tbl_colnames() functions.

param

The parameter name associated to the value that is to be gotten. These parameter names are always available in the first column (.param.) of a summary table obtained by tt_summary_stats(), tt_string_info(), tt_tbl_dims(), or tt_tbl_colnames().

column

The column in the summary table for which the data value should be obtained. This must be supplied for summary tables generated by tt_summary_stats() and tt_string_info() (the tt_tbl_dims() and tt_tbl_colnames() functions will always generate a two-column summary table).

Function ID

12-7

See also

Examples

# Get summary statistics for the
# first quarter of the `game_revenue`
# dataset that's included in the package
stat_tbl <- 
  game_revenue %>%
  tt_time_slice(slice_point = 0.25) %>%
  tt_summary_stats()

# Based on player behavior for the first
# quarter of the year, test whether the
# maximum session duration during the
# rest of the year is never higher
game_revenue %>%
  tt_time_slice(
    slice_point = 0.25,
    keep = "right"
  ) %>%
  test_col_vals_lte(
    columns = vars(session_duration), 
    value = get_tt_param(
      tbl = stat_tbl,
      param = "max",
      column = "session_duration"
    )
  )
#> [1] TRUE