Add column information from another data table

The info_columns_from_tbl() function is a wrapper around the info_columns() function and is useful if you wish to apply info text to columns where that information already exists in a data frame (or in some form that can readily be coaxed into a data frame). The form of the input tbl (the one that contains column metadata) has a few basic requirements:

the data frame must have two columns

both columns must be of class character

the first column should contain column names and the second should contain the info text

Each column that matches across tables (i.e., the tbl and the target table of the informant) will have a new entry for the "info" property. Empty or missing info text will be pruned from tbl.

info_columns_from_tbl(x, tbl, .add = TRUE)

Arguments

x
An informant object of class ptblank_informant.

tbl
The two-column data frame which contains metadata about the target table in the informant object.

.add
Should new text be added to existing text? This is TRUE by default; setting to FALSE replaces any existing text for the "info" property.

Value

A ptblank_informant object.

Function ID

3-3

See also

Examples

# Create a pointblank `informant`
# object with `create_informant()`;
# we specify a `read_fn` with the
# `~` followed by a statement that
# gets the `game_revenue` dataset
informant <- 
  create_informant(
    read_fn = ~ game_revenue,
    tbl_name = "game_revenue",
    label = "An example."
  )

# We can add *info text* to describe
# the columns in the table by using
# information in another table; the
# `info_columns_from_tbl()` takes a
# table object where the first column
# has the column names and the second
# contains the *info text* (the
# `game_revenue_info` dataset contains
# metadata for `game_revenue`)
informant <-
  informant %>%
  info_columns_from_tbl(
    tbl = game_revenue_info
  )

# We can continue to add more *info
# text* since the process is additive;
# the `info_columns_from_tbl()`
# function populates the `info`
# subsection
informant <-
  informant %>%
  info_columns(
    columns = "item_revenue",
    info = "Revenue reported in USD."
  ) %>%
  info_columns(
    columns = "acquisition",
    `top list` = "{top5_aq}"
  ) %>%
  info_snippet(
    snippet_name = "top5_aq",
    fn = snip_list(column = "acquisition")
  ) %>%
  incorporate()

#> 

#> ── Incorporation Started - there is a single snippet to process ────────────────

#> ✓ Information gathered.

#> ✓ Snippets processed.

#> ✓ Information built.

#> 

#> ── Incorporation Completed ─────────────────────────────────────────────────────