Una pequeña tabla que es útil para realizar pruebas

Esta es una pequeña tabla con algunos tipos diferentes de columnas. Probablemente sea útil cuando se prueban las funciones desde pointblank. Las filas 9 y 10 son duplicados exactos. La columna c contiene dos valores NA.

Usage

small_table

Format

Un tibble con 13 filas y 8 variables:

date_time

Una columna de fecha y hora (de la clase POSIXct) con fechas que corresponden exactamente a las de la columna date. Los valores de tiempo son algo aleatorios, pero todos los valores de 'segundos' son 00.

date

Una columna Date con fechas desde 2016-01-04 a 2016-01-30.

a

Una columna integer con valores comprendidos entre 1 y 8.

b

Una columna de character con valores que se adhieren a un patrón común.

c

Una columna integer con valores comprendidos entre 2 y 9. Contiene dos valores NA.

d

Una columna numérica con valores que van desde 108 a 10000.

e

Una columna logical.

f

Una columna character con valores "low", "mid" y "high"

Function ID

14-1

See also

Examples

# Aquí hay un vistazo a los datos
# disponibles en `small_table`
dplyr::glimpse(small_table)
#> Rows: 13
#> Columns: 8
#> $ date_time <dttm> 2016-01-04 11:00:00, 2016-01-04 00:32:00, 2016-01-05 13:32:…
#> $ date      <date> 2016-01-04, 2016-01-04, 2016-01-05, 2016-01-06, 2016-01-09,…
#> $ a         <int> 2, 3, 6, 2, 8, 4, 7, 4, 3, 3, 4, 2, 1
#> $ b         <chr> "1-bcd-345", "5-egh-163", "8-kdg-938", "5-jdo-903", "3-ldm-0…
#> $ c         <dbl> 3, 8, 3, NA, 7, 4, 3, 2, 9, 9, 7, 8, NA
#> $ d         <dbl> 3423.29, 9999.99, 2343.23, 3892.40, 283.94, 3291.03, 843.34,…
#> $ e         <lgl> TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE, FAL…
#> $ f         <chr> "high", "low", "high", "mid", "low", "mid", "high", "low", "…