Un fn para info_snippet(): obtener una lista de categorías de columnas

The snip_list() function can be used as an info_snippet() function (i.e., provided to fn) to get a catalog list from a table column. You can limit the of items in that list with the limit value.

Uso

snip_list(
  column,
  limit = 5,
  sorting = c("inorder", "infreq", "inseq"),
  reverse = FALSE,
  sep = ",",
  and_or = NULL,
  oxford = TRUE,
  as_code = TRUE,
  quot_str = NULL,
  lang = NULL
)

Argumentos

column

El nombre de la columna que contiene los valores de destino.

limit

Un límite de elementos puestos en la lista generada. El texto devuelto indicará el número restante de elementos más allá del limit. Por defecto, el límite es 5.

sorting

Una palabra clave utilizada para designar el tipo de ordenación a utilizar para la lista. Las tres opciones son "inorder" (por defecto), "infreq" e "inseq". Con "inorder", los distintos elementos se listan en el orden en que aparecen por primera vez. Con "infreq" se ordenan los elementos por la frecuencia decreciente de cada uno de ellos. La opción "inseq" aplica una ordenación alfanumérica a los distintos elementos de la lista.

reverse

Una opción para invertir el orden de los elementos de la lista. Por defecto, esto es FALSE pero si se utiliza TRUE se invertirán los elementos antes de aplicar el limit.

sep

El separador a utilizar entre los elementos de la lista. Por defecto, es una coma.

and_or

El tipo de conjunción a utilizar entre el último y el penúltimo elemento de la lista (en caso de que la longitud del elemento sea inferior al valor limit). Si se utiliza NULL (el valor por defecto), se utilizará la conjunción y. Como alternativa, se pueden utilizar las siguientes palabras clave: "y", "o", o una cadena vacía (para no usar ninguna conjunción).

oxford

Si se utiliza la coma de Oxford en determinadas condiciones. En por defecto, es TRUE.

as_code

¿Debe aparecer cada elemento de la lista en una "fuente de código" (es decir, como texto monoespaciado)? Por defecto es TRUE. El uso de FALSE mantiene todos los elementos de la lista de la lista en el mismo tipo de letra que el resto del informe.

quot_str

Una opción para saber si los elementos de la lista deben ir entre comillas dobles. Si es NULL (el valor por defecto), las comillas se asocian principalmente a los elementos de la lista derivados de valores de character o factor; los números, las fechas y los valores lógicos no tendrán comillas. Podemos utilizar explícitamente comillas (o no) con TRUE o FALSE aquí.

lang

El idioma que se utilizará para cualquier palabra de unión (de la opción and_or ) o palabras adicionales en la cadena de la lista generada. Por defecto, NULLutilizará cualquier configuración delangque esté disponible en el objeto *informante* padre (se puede establecer en el argumentolang de [create_informant()]). Si se especifica aquí como una anulación, las opciones de idioma son francés ("fr"), alemán ("de"), italiano ("it"), español ("es"), portugués ("pt"), turco ("tr"), chino ("zh"), ruso ("ru"), polaco ("pl"), danés ("da"), sueco ("sv" ) y holandés ("nl"`).

Valor

Una fórmula necesaria para el argumento fn de info_snippet().

Function ID

3-6

Ver también

Ejemplos

# Generar un objeto informante,
# añadir un snippet con
# `info_snippet()` y
# `snip_list()` (dándonos un
# método para obtener una lista
# distinta de valores de columna
# para la columna `f`); definir
# una ubicación para el resultado
# del snippet en `{ }` y luego
# `incorporate()` el snippet en
# el texto informativo
informant <- 
  create_informant(
    read_fn = ~ small_table,
    tbl_name = "small_table",
    label = "Un ejemplo."
  ) %>% 
  info_columns(
    columns = "f",
    `Items` = "This column contains {values_f}."
  ) %>%
  info_snippet(
    snippet_name = "values_f",
    fn = snip_list(column = "f")
  ) %>%
  incorporate()
#> 
#> ── Incorporation Started - there is a single snippet to process ────────────────
#> ✔ Information gathered.
#> ✔ Snippets processed.
#> ✔ Information built.
#> 
#> ── Incorporation Completed ─────────────────────────────────────────────────────

# Podemos imprimir el objeto
# `informant` para ver el informe
# de información