Definir un almacén de tablas con fórmulas de preparación de tablas: un "table store"

Puede ser útil configurar todas las fuentes de datos que necesites y sólo extraerlas cuando sea necesario. Esta configuración por adelantado con tbl_store() nos permite definir los métodos para obtener datos tabulares de fuentes mixtas (por ejemplo, tablas de bases de datos, tablas generadas a partir de archivos planos, etc.) y proporcionar nombres para estos procedimientos de preparación de datos. Entonces tenemos una forma conveniente de acceder a las tablas materializadas con tbl_get(), o, a las fórmulas de preparación de tablas con tbl_source(). Las fórmulas table-prep pueden ser tan simples como obtener una tabla de una ubicación, o, puede involucrar tanta mutación como sea necesaria (imagine obtener varias variaciones mutadas de la misma tabla fuente, generar una tabla desde múltiples fuentes, o pre-filtrar una tabla de la base de datos de acuerdo al tiempo del sistema). Otro aspecto agradable de organizar las fórmulas de preparación de tablas en un solo objeto es suministrarlo al argumento read_fn de create_agent() o create_informant() mediante la notación $ (por ejemplo, create_agent(read_fn = <tbl_store>$<nombre>)) o con tbl_source() (por ejemplo, create_agent(read_fn = ~ tbl_source("<nombre>", <tbl_store>))).

Uso

tbl_store(..., .list = list2(...))

Argumentos

...

Expresiones que contienen fórmulas de preparación de tablas y nombres de tablas para la recuperación de datos. Deben utilizarse fórmulas de dos lados (por ejemplo, <LHS> ~ <RHS>), donde el lado izquierdo es un nombre dado y el derecho es la parte que se utiliza para obtener la tabla.

.list

Permite el uso de una lista como alternativa de entrada a ....

Valor

Un objeto tbl_store que contiene fórmulas de preparación de tablas.

YAML

Un pointblank tbl_store se puede escribir en YAML con yaml_write() y el YAML resultante se puede utilizar de varias maneras. El escenario ideal es tener agentes e informantes pointblank también en forma YAML. De esta manera el agente y el informante pueden referirse a la tabla de almacenamiento YAML (a través de tbl_source()), y, el procesamiento de ambos agentes e informantes puede realizarse con yaml_agent_interrogate() y yaml_informant_incorporate(). Con el siguiente código de R, se genera un almacén de tablas con dos fórmulas de preparación de tablas y se escribe en YAML (si no se da un nombre de archivo, el YAML se escribe en "tbl_store.yml").

# Código R para generar el archivo "tbl_store.yml"
tbl_store(
  tbl_duckdb ~ db_tbl(small_table, dbname = ":memory:", dbtype = "duckdb"),
  sml_table_high ~ small_table %>% dplyr::filter(f == "high")
) %>%
  yaml_write()

# Representación YAML ("tbl_store.yml")
tbls:
  tbl_duckdb: ~ db_tbl(small_table, dbname = ":memory:", dbtype = "duckdb")
  sml_table_high: ~ small_table %>% dplyr::filter(f == "high")

Esto es útil cuando se quiere obtener tiradas frescas de datos preparados de una fuente materializada en una sesión de R (con la función tbl_get(). Por ejemplo, la tabla sml_table_high puede obtenerse utilizando tbl_get("sml_table_high", "tbl_store.yml"). Para conseguir que un agente compruebe estos datos preparados periódicamente, será útil el siguiente ejemplo con tbl_source():

# Generar un objeto agente que compruebe
# `sml_table_high`, escribirlo en YAML
create_agent(
  read_fn = ~ tbl_source("sml_table_high", "tbl_store.yml"),
  label = "An example that uses a table store.",
  actions = action_levels(warn_at = 0.10)
) %>% 
  col_exists(vars(date, date_time)) %>%
  write_yaml()
  
# Representación YAML ("agent-sml_table_high.yml")
read_fn: ~ tbl_source("sml_table_high", "tbl_store.yml")
tbl_name: sml_table_high
label: An example that uses a table store.
actions:
  warn_fraction: 0.1
locale: en
steps:
  - col_exists:
    columns: vars(date, date_time)

Ahora, siempre que la tabla sml_table_high necesite ser validada, se puede hacer con yaml_agent_interrogate() (por ejemplo, yaml_agent_interrogate("agent-sml_table_high.yml")).

Function ID

1-8

Ver también

Ejemplos

if (interactive()) {

# Definir un objeto `tbl_store` añadiendo
# fórmulas de preparación de tablas
# dentro de la `tbl_store()` llamada
tbls <- 
  tbl_store(
    small_table_duck ~ db_tbl(
      table = small_table,
      dbname = ":memory:",
      dbtype = "duckdb"
    ),
    ~ db_tbl(
      table = "rna",
      dbname = "pfmegrnargs",
      dbtype = "postgres",
      host = "hh-pgsql-public.ebi.ac.uk",
      port = 5432,
      user = I("reader"),
      password = I("NWDMCE5xdipIjRrp")
    ),
    all_revenue ~ db_tbl(
      table = file_tbl(
        file = from_github(
          file = "all_revenue_large.rds",
          repo = "rich-iannone/intendo",
          subdir = "data-large"
        )
      ),
      dbname = ":memory:",
      dbtype = "duckdb"
    ),
    sml_table ~ pointblank::small_table
  )

# Una vez que este objeto está disponible,
# se puede comprobar que la tabla de
# interés se produce a su especificación
# con la función `tbl_get()`.
tbl_get(
  tbl = "small_table_duck",
  store = tbls
)

# Otra forma más sencilla de obtener la
# misma tabla materializada es utilizando
# `$` para obtener la entrada de
# elección para `tbl_get()`.
tbls$small_table_duck %>% tbl_get()

# La creación de un agente es fácil
# cuando todas las fórmulas de
# preparación de tablas están
# encapsuladas en un objeto `tbl_store`;
# utilice la notación `$` para pasar el
# procedimiento apropiado para leer una
# tabla al argumento `read_fn`
agent_1 <-
  create_agent(
    read_fn = tbls$small_table_duck
  )
  
# Existen otras formas de utilizar el
# almacén de tablas para asignar una
# tabla de destino a un agente, como
# el uso de la función `tbl_source()`
agent_2 <-
  create_agent(
    read_fn = ~ tbl_source(
      tbl = "small_table_duck",
      store = tbls
      )
  )

# El almacén de tablas puede ser
# trasladado a YAML con `yaml_write`
# y la llamada `tbl_source()` podría
# entonces referirse a ese almacén
# de tablas en disco; hagamos esa
# conversión a YAML
yaml_write(tbls)

# Lo anterior escribe el archivo
# `tbl_store.yml` (al no
# proporcionar un `filename` se
# elige este nombre de archivo por
# defecto); a continuación, modifica
# el `tbl_source()` para que `store`
# haga referencia al archivo YAML 
agent_3 <-
  create_agent(
    read_fn = ~ tbl_source(
      tbl = "small_table_duck",
      store = "tbl_store.yml"
    )
  )

}