Table Transformer: get a table's column names

With any table object, you can produce a summary table that contains table's column names. The output summary table will have two columns and as many rows as there are columns in the input table. The first column is the ".param." column, which is an integer-based column containing the indices of the columns from the input table. The second column, "value", contains the column names from the input table.

tt_tbl_colnames(tbl)

Arguments

tbl
A table object to be used as input for the transformation. This can be a data frame, a tibble, a tbl_dbi object, or a tbl_spark object.

Value

A tibble object.

Function ID

12-4

See also

Examples

# Get the column names of the
# `game_revenue` dataset that's
# included in the package
tt_tbl_colnames(game_revenue)

#> # A tibble: 11 x 2
#>    .param. value           
#>      <int> <chr>           
#>  1       1 player_id       
#>  2       2 session_id      
#>  3       3 session_start   
#>  4       4 time            
#>  5       5 item_type       
#>  6       6 item_name       
#>  7       7 item_revenue    
#>  8       8 session_duration
#>  9       9 start_day       
#> 10      10 acquisition     
#> 11      11 country         

# This output table is useful when
# you want to validate the
# column names of the table; here,
# we check that `game_revenue` has
# certain column names present
tt_tbl_colnames(game_revenue) %>%
  test_col_vals_make_subset(
    columns = vars(value),
    set = c("acquisition", "country")
  )

#> [1] TRUE

# We can check to see whether the
# column names in the `specifications`
# table are all less than 15
# characters in length
specifications %>%
  tt_tbl_colnames() %>%
  tt_string_info() %>%
  test_col_vals_lt(
    columns = vars(value),
    value = 15
  )

#> [1] FALSE